refactor(store): split pg.go into per-entity files
pg.go went from 5057 lines to ~280: it now holds only the PG core
(pool, migrations) and cross-entity helpers (cursor codec, scan/marshal
utilities, isUniqueViolation, escapeLike). Entity CRUD moves unchanged
into pg_clusters.go, pg_nodes.go, pg_namespaces.go, pg_pods.go,
pg_workloads.go, pg_services.go, pg_ingresses.go,
pg_persistent_volumes.go, pg_persistent_volume_claims.go and
pg_settings.go — completing the per-domain layout the satellite files
(pg_auth.go, pg_applications.go, …) already follow.

Pure move, no behaviour change; store integration tests pass against
PostgreSQL 16.
