+++ internal/store/pg.go
@@ -1,4 +1,13 @@
 // Package store provides the PostgreSQL implementation of api.Store.
+//
+// This file holds the PG core (pool, migrations) plus helpers shared
+// across entities (cursor codec, scan/marshal utilities, error
+// classification). Entity CRUD lives in per-entity pg_*.go files
+// (pg_clusters.go, pg_nodes.go, pg_namespaces.go, pg_pods.go,
+// pg_workloads.go, pg_services.go, pg_ingresses.go,
+// pg_persistent_volumes.go, pg_persistent_volume_claims.go,
+// pg_settings.go) alongside the pre-existing domain files
+// (pg_auth.go, pg_applications.go, pg_cloud_accounts.go, …).
 package store
 
 import (
@@ -8,7 +17,6 @@ import (
 	"encoding/json"
 	"errors"
 	"fmt"
-	"reflect"
 	"strings"
 	"time"
 
@@ -21,7 +29,6 @@ import (
 
 	"github.com/sthalbert/longue-vue/internal/api"
 	"github.com/sthalbert/longue-vue/internal/secrets"
-	"github.com/sthalbert/longue-vue/internal/timetravel"
 	"github.com/sthalbert/longue-vue/migrations"
 )
 
@@ -138,530 +145,6 @@ func (r scanRowWith) Scan(dest ...any) error {
 	return nil
 }
 
-// EnsureCluster is idempotent on the cluster name. It has three branches:
-//
-//   - RESTORE — a terminated row already exists. Clear terminated_at,
-//     write a `restore` history row, return the existing row. Mirrors
-//     the auto-restore behaviour of UpsertNamespace/Node/Workload so a
-//     collector that keeps pushing after a soft-delete resurrects the
-//     cluster instead of being silently ignored (Fix 2 in the spec at
-//     docs/superpowers/specs/2026-05-18-cluster-cascade-delete-design.md).
-//   - NO-OP — a live row already exists. Return it unchanged. No history.
-//   - CREATE — no row exists. INSERT and capture create history.
-//
-// The returned bool is true only on CREATE. Concurrent inserts of the
-// same name are resolved by INSERT … ON CONFLICT (name) DO NOTHING:
-// the loser falls through to a re-fetch (still false).
-//
-//nolint:gocyclo,gocritic // branch-heavy idempotent insert; hugeParam: Store interface
-func (p *PG) EnsureCluster(ctx context.Context, in api.ClusterCreate) (api.Cluster, bool, error) {
-	id := uuid.New()
-	now := time.Now().UTC()
-
-	labelsJSON, err := marshalLabels(in.Labels)
-	if err != nil {
-		return api.Cluster{}, false, err
-	}
-	annotationsJSON, err := marshalLabels(in.Annotations)
-	if err != nil {
-		return api.Cluster{}, false, fmt.Errorf("marshal cluster annotations: %w", err)
-	}
-
-	tx, err := p.pool.Begin(ctx)
-	if err != nil {
-		return api.Cluster{}, false, fmt.Errorf("begin ensure cluster: %w", err)
-	}
-	defer func() { _ = tx.Rollback(ctx) }()
-
-	actor := timetravel.ActorFromContext(ctx)
-
-	// Snapshot any existing row by name so we can pick a branch.
-	var prevID *uuid.UUID
-	var prevTerminatedAt *time.Time
-	_ = tx.QueryRow(ctx,
-		`SELECT id, terminated_at FROM clusters WHERE name=$1`,
-		in.Name,
-	).Scan(&prevID, &prevTerminatedAt)
-
-	// Branch RESTORE: terminated row exists → clear terminated_at + history.
-	if prevID != nil && prevTerminatedAt != nil {
-		if _, err := tx.Exec(ctx,
-			`UPDATE clusters SET terminated_at = NULL, updated_at = $1 WHERE id = $2`,
-			now, *prevID); err != nil {
-			return api.Cluster{}, false, fmt.Errorf("restore cluster: %w", err)
-		}
-		if snap, err := clusterRowMapNoLock(ctx, tx, *prevID); err == nil {
-			_ = timetravel.Capture(ctx, tx, timetravel.KindCluster, *prevID, nil, snap, changeTypeRestore, actor)
-		}
-		if err := tx.Commit(ctx); err != nil {
-			return api.Cluster{}, false, fmt.Errorf("commit ensure cluster (restore): %w", err)
-		}
-		restored, getErr := p.GetClusterByName(ctx, in.Name)
-		if getErr != nil {
-			return api.Cluster{}, false, fmt.Errorf("ensure cluster %q: fetch restored: %w", in.Name, getErr)
-		}
-		return restored, false, nil
-	}
-
-	// Branch NO-OP: live row exists → return as-is.
-	if prevID != nil {
-		if err := tx.Commit(ctx); err != nil {
-			return api.Cluster{}, false, fmt.Errorf("commit ensure cluster (existing): %w", err)
-		}
-		existing, getErr := p.GetClusterByName(ctx, in.Name)
-		if getErr != nil {
-			return api.Cluster{}, false, fmt.Errorf("ensure cluster %q: fetch existing: %w", in.Name, getErr)
-		}
-		return existing, false, nil
-	}
-
-	// Branch CREATE: insert. Use ON CONFLICT DO NOTHING to absorb a race
-	// where two transactions reach this point with the same name.
-	const q = `
-		INSERT INTO clusters (
-			id, name, display_name, environment, provider, region,
-			kubernetes_version, api_endpoint, labels,
-			owner, criticality, notes, runbook_url, annotations,
-			created_at, updated_at
-		) VALUES ($1, $2, $3, $4, $5, $6, $7, $8, $9, $10, $11, $12, $13, $14, $15, $15)
-		ON CONFLICT (name) DO NOTHING
-		RETURNING id
-	`
-	var insertedID uuid.UUID
-	scanErr := tx.QueryRow(ctx, q,
-		id, in.Name, in.DisplayName, in.Environment, in.Provider, in.Region,
-		in.KubernetesVersion, in.ApiEndpoint, labelsJSON,
-		in.Owner, in.Criticality, in.Notes, in.RunbookUrl, annotationsJSON,
-		now,
-	).Scan(&insertedID)
-	switch {
-	case scanErr == nil:
-		if snap, err := clusterRowMapNoLock(ctx, tx, insertedID); err == nil {
-			_ = timetravel.Capture(ctx, tx, timetravel.KindCluster, insertedID, nil, snap, changeTypeCreate, actor)
-		}
-		if err := tx.Commit(ctx); err != nil {
-			return api.Cluster{}, false, fmt.Errorf("commit ensure cluster: %w", err)
-		}
-		return api.Cluster{
-			Id:                &insertedID,
-			Name:              in.Name,
-			DisplayName:       in.DisplayName,
-			Environment:       in.Environment,
-			Provider:          in.Provider,
-			Region:            in.Region,
-			KubernetesVersion: in.KubernetesVersion,
-			ApiEndpoint:       in.ApiEndpoint,
-			Labels:            in.Labels,
-			Owner:             in.Owner,
-			Criticality:       in.Criticality,
-			Notes:             in.Notes,
-			RunbookUrl:        in.RunbookUrl,
-			Annotations:       in.Annotations,
-			CreatedAt:         &now,
-			UpdatedAt:         &now,
-		}, true, nil
-	case errors.Is(scanErr, pgx.ErrNoRows):
-		// Race: a concurrent tx inserted between our pre-snapshot and INSERT.
-		// Commit (the tx has no effect) and re-fetch.
-		if err := tx.Commit(ctx); err != nil {
-			return api.Cluster{}, false, fmt.Errorf("commit ensure cluster (race): %w", err)
-		}
-		existing, getErr := p.GetClusterByName(ctx, in.Name)
-		if getErr != nil {
-			return api.Cluster{}, false, fmt.Errorf("ensure cluster %q: fetch existing after race: %w", in.Name, getErr)
-		}
-		return existing, false, nil
-	default:
-		return api.Cluster{}, false, fmt.Errorf("insert cluster: %w", scanErr)
-	}
-}
-
-// GetCluster fetches a cluster by id.
-func (p *PG) GetCluster(ctx context.Context, id uuid.UUID) (api.Cluster, error) {
-	const q = `
-		SELECT id, name, display_name, environment, provider, region,
-		       kubernetes_version, api_endpoint, labels,
-		       owner, criticality, notes, runbook_url, annotations,
-		       created_at, updated_at
-		FROM clusters
-		WHERE id = $1
-	`
-	row := p.pool.QueryRow(ctx, q, id)
-	c, err := scanCluster(row)
-	if err != nil {
-		if errors.Is(err, pgx.ErrNoRows) {
-			return api.Cluster{}, api.ErrNotFound
-		}
-		return api.Cluster{}, fmt.Errorf("select cluster: %w", err)
-	}
-	return c, nil
-}
-
-// GetClusterByName fetches a cluster by its unique name column.
-func (p *PG) GetClusterByName(ctx context.Context, name string) (api.Cluster, error) {
-	const q = `
-		SELECT id, name, display_name, environment, provider, region,
-		       kubernetes_version, api_endpoint, labels,
-		       owner, criticality, notes, runbook_url, annotations,
-		       created_at, updated_at
-		FROM clusters
-		WHERE name = $1
-	`
-	row := p.pool.QueryRow(ctx, q, name)
-	c, err := scanCluster(row)
-	if err != nil {
-		if errors.Is(err, pgx.ErrNoRows) {
-			return api.Cluster{}, api.ErrNotFound
-		}
-		return api.Cluster{}, fmt.Errorf("select cluster by name: %w", err)
-	}
-	return c, nil
-}
-
-// ListClusters returns up to limit clusters in (created_at DESC, id DESC) order,
-// starting after the opaque cursor if non-empty.
-//
-//nolint:gocyclo // cursor-paginated query builder with optional filters
-func (p *PG) ListClusters(ctx context.Context, limit int, cursor string, includeTerminated bool) ([]api.Cluster, string, error) {
-	if limit <= 0 {
-		limit = 50
-	}
-	if limit > 200 {
-		limit = 200
-	}
-
-	var (
-		rows pgx.Rows
-		err  error
-	)
-	if cursor == "" {
-		const q = `
-			SELECT id, name, display_name, environment, provider, region,
-			       kubernetes_version, api_endpoint, labels,
-			       owner, criticality, notes, runbook_url, annotations,
-			       created_at, updated_at
-			FROM clusters
-			WHERE ($1 OR terminated_at IS NULL)
-			ORDER BY created_at DESC, id DESC
-			LIMIT $2
-		`
-		rows, err = p.pool.Query(ctx, q, includeTerminated, limit+1)
-	} else {
-		ts, id, cerr := decodeCursor(cursor)
-		if cerr != nil {
-			return nil, "", cerr
-		}
-		const q = `
-			SELECT id, name, display_name, environment, provider, region,
-			       kubernetes_version, api_endpoint, labels,
-			       owner, criticality, notes, runbook_url, annotations,
-			       created_at, updated_at
-			FROM clusters
-			WHERE (created_at, id) < ($1, $2)
-			  AND ($3 OR terminated_at IS NULL)
-			ORDER BY created_at DESC, id DESC
-			LIMIT $4
-		`
-		rows, err = p.pool.Query(ctx, q, ts, id, includeTerminated, limit+1)
-	}
-	if err != nil {
-		return nil, "", fmt.Errorf("query clusters: %w", err)
-	}
-	defer rows.Close()
-
-	items := make([]api.Cluster, 0, limit)
-	for rows.Next() {
-		c, err := scanCluster(rows)
-		if err != nil {
-			return nil, "", fmt.Errorf("scan cluster: %w", err)
-		}
-		items = append(items, c)
-	}
-	if err := rows.Err(); err != nil {
-		return nil, "", fmt.Errorf("iterate clusters: %w", err)
-	}
-
-	var next string
-	if len(items) > limit {
-		last := items[limit-1]
-		if last.CreatedAt != nil && last.Id != nil {
-			next = encodeCursor(*last.CreatedAt, *last.Id)
-		}
-		items = items[:limit]
-	}
-	return items, next, nil
-}
-
-// UpdateCluster applies merge-patch semantics: only fields that are non-nil
-// on ClusterUpdate are written. updated_at is always refreshed.
-//
-//nolint:gocyclo,gocritic // merge-patch nil checks are inherently repetitive; hugeParam: Store interface requires value param
-func (p *PG) UpdateCluster(ctx context.Context, id uuid.UUID, in api.ClusterUpdate) (api.Cluster, error) {
-	sets := make([]string, 0, 8)
-	args := make([]any, 0, 10)
-	idx := 1
-
-	appendSet := func(column string, value any) {
-		sets = append(sets, fmt.Sprintf("%s=$%d", column, idx))
-		args = append(args, value)
-		idx++
-	}
-
-	if in.DisplayName != nil {
-		appendSet("display_name", *in.DisplayName)
-	}
-	if in.Environment != nil {
-		appendSet("environment", *in.Environment)
-	}
-	if in.Provider != nil {
-		appendSet("provider", *in.Provider)
-	}
-	if in.Region != nil {
-		appendSet("region", *in.Region)
-	}
-	if in.KubernetesVersion != nil {
-		appendSet("kubernetes_version", *in.KubernetesVersion)
-	}
-	if in.ApiEndpoint != nil {
-		appendSet("api_endpoint", *in.ApiEndpoint)
-	}
-	if in.Labels != nil {
-		b, err := marshalLabels(in.Labels)
-		if err != nil {
-			return api.Cluster{}, err
-		}
-		appendSet("labels", b)
-	}
-	// Curated metadata — never written by the collector (it only ever
-	// patches KubernetesVersion), so omission here is already safe.
-	if in.Owner != nil {
-		appendSet("owner", *in.Owner)
-	}
-	if in.Criticality != nil {
-		appendSet("criticality", *in.Criticality)
-	}
-	if in.Notes != nil {
-		appendSet("notes", *in.Notes)
-	}
-	if in.RunbookUrl != nil {
-		appendSet("runbook_url", *in.RunbookUrl)
-	}
-	if in.Annotations != nil {
-		b, err := marshalLabels(in.Annotations)
-		if err != nil {
-			return api.Cluster{}, fmt.Errorf("marshal cluster annotations: %w", err)
-		}
-		appendSet("annotations", b)
-	}
-
-	appendSet("updated_at", time.Now().UTC())
-
-	args = append(args, id)
-
-	// Wrap in a transaction so we can read prev, run the UPDATE, read next,
-	// and call Capture atomically.
-	tx, txErr := p.pool.Begin(ctx)
-	if txErr != nil {
-		return api.Cluster{}, fmt.Errorf("begin update cluster: %w", txErr)
-	}
-	defer func() { _ = tx.Rollback(ctx) }()
-
-	prev, prevErr := clusterRowMap(ctx, tx, id) // FOR UPDATE lock
-	if prevErr != nil {
-		return api.Cluster{}, api.ErrNotFound
-	}
-
-	q := fmt.Sprintf("UPDATE clusters SET %s WHERE id=$%d", strings.Join(sets, ", "), idx)
-	tag, err := tx.Exec(ctx, q, args...)
-	if err != nil {
-		return api.Cluster{}, fmt.Errorf("update cluster: %w", err)
-	}
-	if tag.RowsAffected() == 0 {
-		return api.Cluster{}, api.ErrNotFound
-	}
-
-	next, nextErr := clusterRowMapNoLock(ctx, tx, id)
-	if nextErr == nil {
-		actor := timetravel.ActorFromContext(ctx)
-		_ = timetravel.Capture(ctx, tx, timetravel.KindCluster, id, prev, next, changeTypeUpdate, actor)
-	}
-
-	if err := tx.Commit(ctx); err != nil {
-		return api.Cluster{}, fmt.Errorf("commit update cluster: %w", err)
-	}
-	return p.GetCluster(ctx, id)
-}
-
-// DeleteCluster removes a cluster by id.
-func (p *PG) DeleteCluster(ctx context.Context, id uuid.UUID) error {
-	tag, err := p.pool.Exec(ctx, "DELETE FROM clusters WHERE id=$1", id)
-	if err != nil {
-		return fmt.Errorf("delete cluster: %w", err)
-	}
-	if tag.RowsAffected() == 0 {
-		return api.ErrNotFound
-	}
-	return nil
-}
-
-// SoftDeleteCluster marks the cluster and all its live children
-// (namespaces, nodes, workloads) as terminated in a single transaction.
-// Mirrors ADR-0021 §IMP-007. Children that are already terminated are
-// skipped via the AND terminated_at IS NULL guard.
-//
-// Pods, services, ingresses, PVCs (per namespace) and PVs (per cluster)
-// are out-of-scope for time-travel history (ADR-0021 §IMP-007) and are
-// HARD-DELETED here. Without this step they would remain in the DB
-// attached to a terminated parent and still be queryable.
-//
-//nolint:gocyclo // cascade + history capture per entity-kind add branches; acceptable here
-func (p *PG) SoftDeleteCluster(ctx context.Context, id uuid.UUID) error {
-	tx, err := p.pool.Begin(ctx)
-	if err != nil {
-		return fmt.Errorf("begin: %w", err)
-	}
-	defer func() { _ = tx.Rollback(ctx) }()
-
-	actor := timetravel.ActorFromContext(ctx)
-
-	// Collect live workload IDs before soft-deleting them so we can write history.
-	wlIDs, err := liveWorkloadIDsForCluster(ctx, tx, id)
-	if err != nil {
-		return fmt.Errorf("list workloads for soft-delete cluster: %w", err)
-	}
-	nsIDs, err := liveNamespaceIDsForCluster(ctx, tx, id)
-	if err != nil {
-		return fmt.Errorf("list namespaces for soft-delete cluster: %w", err)
-	}
-	nodeIDs, err := liveNodeIDsForCluster(ctx, tx, id)
-	if err != nil {
-		return fmt.Errorf("list nodes for soft-delete cluster: %w", err)
-	}
-
-	// Hard-delete the unhistoried children first. These tables have ON
-	// DELETE CASCADE from cluster/namespace, but FK CASCADE only fires
-	// on hard-delete; since we soft-delete the parent it never fires.
-	if _, err := tx.Exec(ctx,
-		`DELETE FROM pods
-		   WHERE namespace_id IN (SELECT id FROM namespaces WHERE cluster_id = $1)`, id); err != nil {
-		return fmt.Errorf("cascade-delete pods: %w", err)
-	}
-	if _, err := tx.Exec(ctx,
-		`DELETE FROM services
-		   WHERE namespace_id IN (SELECT id FROM namespaces WHERE cluster_id = $1)`, id); err != nil {
-		return fmt.Errorf("cascade-delete services: %w", err)
-	}
-	if _, err := tx.Exec(ctx,
-		`DELETE FROM ingresses
-		   WHERE namespace_id IN (SELECT id FROM namespaces WHERE cluster_id = $1)`, id); err != nil {
-		return fmt.Errorf("cascade-delete ingresses: %w", err)
-	}
-	if _, err := tx.Exec(ctx,
-		`DELETE FROM persistent_volume_claims
-		   WHERE namespace_id IN (SELECT id FROM namespaces WHERE cluster_id = $1)`, id); err != nil {
-		return fmt.Errorf("cascade-delete persistent_volume_claims: %w", err)
-	}
-	if _, err := tx.Exec(ctx,
-		`DELETE FROM persistent_volumes WHERE cluster_id = $1`, id); err != nil {
-		return fmt.Errorf("cascade-delete persistent_volumes: %w", err)
-	}
-
-	if _, err := tx.Exec(ctx,
-		`UPDATE workloads SET terminated_at = NOW(), updated_at = NOW()
-		   WHERE namespace_id IN (SELECT id FROM namespaces WHERE cluster_id = $1)
-		     AND terminated_at IS NULL`, id); err != nil {
-		return fmt.Errorf("soft-delete workloads: %w", err)
-	}
-	if _, err := tx.Exec(ctx,
-		`UPDATE namespaces SET terminated_at = NOW(), updated_at = NOW()
-		   WHERE cluster_id = $1 AND terminated_at IS NULL`, id); err != nil {
-		return fmt.Errorf("soft-delete namespaces: %w", err)
-	}
-	if _, err := tx.Exec(ctx,
-		`UPDATE nodes SET terminated_at = NOW(), updated_at = NOW()
-		   WHERE cluster_id = $1 AND terminated_at IS NULL`, id); err != nil {
-		return fmt.Errorf("soft-delete nodes: %w", err)
-	}
-	tag, err := tx.Exec(ctx,
-		`UPDATE clusters SET terminated_at = NOW(), updated_at = NOW()
-		   WHERE id = $1 AND terminated_at IS NULL`, id)
-	if err != nil {
-		return fmt.Errorf("soft-delete cluster: %w", err)
-	}
-	if tag.RowsAffected() == 0 {
-		var exists bool
-		if err := tx.QueryRow(ctx,
-			`SELECT EXISTS(SELECT 1 FROM clusters WHERE id = $1)`, id).Scan(&exists); err != nil {
-			return fmt.Errorf("check cluster exists: %w", err)
-		}
-		if !exists {
-			return api.ErrNotFound
-		}
-		// Already terminated → idempotent success.
-	}
-
-	// Capture history for cascade-affected entities (after UPDATEs so
-	// the rows reflect terminated_at).
-	for _, wlID := range wlIDs {
-		if snap, err := workloadRowMapNoLock(ctx, tx, wlID); err == nil {
-			_ = timetravel.Capture(ctx, tx, timetravel.KindWorkload, wlID, nil, snap, changeTypeSoftDelete, actor)
-		}
-	}
-	for _, nsID := range nsIDs {
-		if snap, err := namespaceRowMapNoLock(ctx, tx, nsID); err == nil {
-			_ = timetravel.Capture(ctx, tx, timetravel.KindNamespace, nsID, nil, snap, changeTypeSoftDelete, actor)
-		}
-	}
-	for _, nodeID := range nodeIDs {
-		if snap, err := nodeRowMapNoLock(ctx, tx, nodeID); err == nil {
-			_ = timetravel.Capture(ctx, tx, timetravel.KindNode, nodeID, nil, snap, changeTypeSoftDelete, actor)
-		}
-	}
-	if snap, err := clusterRowMapNoLock(ctx, tx, id); err == nil {
-		_ = timetravel.Capture(ctx, tx, timetravel.KindCluster, id, nil, snap, changeTypeSoftDelete, actor)
-	}
-
-	if err := tx.Commit(ctx); err != nil {
-		return fmt.Errorf("commit soft-delete cluster: %w", err)
-	}
-	return nil
-}
-
-// liveWorkloadIDsForCluster returns ids of non-terminated workloads in namespaces of cluster.
-func liveWorkloadIDsForCluster(ctx context.Context, tx pgx.Tx, clusterID uuid.UUID) ([]uuid.UUID, error) {
-	rows, err := tx.Query(ctx,
-		`SELECT id FROM workloads
-		  WHERE namespace_id IN (SELECT id FROM namespaces WHERE cluster_id = $1)
-		    AND terminated_at IS NULL`,
-		clusterID)
-	if err != nil {
-		return nil, fmt.Errorf("query live workloads for cluster: %w", err)
-	}
-	return scanUUIDs(rows)
-}
-
-// liveNamespaceIDsForCluster returns ids of non-terminated namespaces of cluster.
-func liveNamespaceIDsForCluster(ctx context.Context, tx pgx.Tx, clusterID uuid.UUID) ([]uuid.UUID, error) {
-	rows, err := tx.Query(ctx,
-		`SELECT id FROM namespaces WHERE cluster_id = $1 AND terminated_at IS NULL`,
-		clusterID)
-	if err != nil {
-		return nil, fmt.Errorf("query live namespaces for cluster: %w", err)
-	}
-	return scanUUIDs(rows)
-}
-
-// liveNodeIDsForCluster returns ids of non-terminated nodes of cluster.
-func liveNodeIDsForCluster(ctx context.Context, tx pgx.Tx, clusterID uuid.UUID) ([]uuid.UUID, error) {
-	rows, err := tx.Query(ctx,
-		`SELECT id FROM nodes WHERE cluster_id = $1 AND terminated_at IS NULL`,
-		clusterID)
-	if err != nil {
-		return nil, fmt.Errorf("query live nodes for cluster: %w", err)
-	}
-	return scanUUIDs(rows)
-}
-
 func scanUUIDs(rows pgx.Rows) ([]uuid.UUID, error) {
 	defer rows.Close()
 	var ids []uuid.UUID
@@ -678,2406 +161,6 @@ func scanUUIDs(rows pgx.Rows) ([]uuid.UUID, error) {
 	return ids, nil
 }
 
-// CountClusterChildren counts every child resource attached to the cluster
-// at the time of the call. Used by the audit log to snapshot the pre-delete
-// inventory before SoftDeleteCluster removes pods / services / ingresses /
-// PVCs / PVs (Fix 1, see SoftDeleteCluster) and soft-deletes the cluster /
-// namespaces / nodes / workloads. A single round-trip multi-CTE query keeps
-// the cost bounded regardless of how many resource types exist (ADR-0010).
-func (p *PG) CountClusterChildren(ctx context.Context, clusterID uuid.UUID) (api.CascadeCounts, error) {
-	const q = `
-		WITH ns_ids AS (
-			SELECT id FROM namespaces WHERE cluster_id = $1
-		),
-		ns_count   AS (SELECT COUNT(*) AS c FROM ns_ids),
-		node_count AS (SELECT COUNT(*) AS c FROM nodes WHERE cluster_id = $1),
-		pv_count   AS (SELECT COUNT(*) AS c FROM persistent_volumes WHERE cluster_id = $1),
-		pod_count  AS (SELECT COUNT(*) AS c FROM pods WHERE namespace_id IN (SELECT id FROM ns_ids)),
-		wl_count   AS (SELECT COUNT(*) AS c FROM workloads WHERE namespace_id IN (SELECT id FROM ns_ids)),
-		svc_count  AS (SELECT COUNT(*) AS c FROM services WHERE namespace_id IN (SELECT id FROM ns_ids)),
-		ing_count  AS (SELECT COUNT(*) AS c FROM ingresses WHERE namespace_id IN (SELECT id FROM ns_ids)),
-		pvc_count  AS (SELECT COUNT(*) AS c FROM persistent_volume_claims WHERE namespace_id IN (SELECT id FROM ns_ids))
-		SELECT
-			(SELECT c FROM ns_count),
-			(SELECT c FROM node_count),
-			(SELECT c FROM pod_count),
-			(SELECT c FROM wl_count),
-			(SELECT c FROM svc_count),
-			(SELECT c FROM ing_count),
-			(SELECT c FROM pv_count),
-			(SELECT c FROM pvc_count)
-	`
-	// Verify the cluster exists before running counts; a non-existent
-	// cluster would just return all zeroes, which is misleading.
-	var exists bool
-	if err := p.pool.QueryRow(ctx, "SELECT EXISTS(SELECT 1 FROM clusters WHERE id=$1)", clusterID).Scan(&exists); err != nil {
-		return api.CascadeCounts{}, fmt.Errorf("count cluster children: existence check: %w", err)
-	}
-	if !exists {
-		return api.CascadeCounts{}, api.ErrNotFound
-	}
-
-	var cc api.CascadeCounts
-	err := p.pool.QueryRow(ctx, q, clusterID).Scan(
-		&cc.Namespaces,
-		&cc.Nodes,
-		&cc.Pods,
-		&cc.Workloads,
-		&cc.Services,
-		&cc.Ingresses,
-		&cc.PersistentVolumes,
-		&cc.PersistentVolumeClaims,
-	)
-	if err != nil {
-		return api.CascadeCounts{}, fmt.Errorf("count cluster children: %w", err)
-	}
-	return cc, nil
-}
-
-// CreateNode inserts a new node. Returns api.ErrNotFound when the parent
-// cluster does not exist (FK violation), api.ErrConflict on duplicate
-// (cluster_id, name).
-// nodeColumns is the INSERT/SELECT column order used by every Node SQL
-// path — CreateNode, UpsertNode, scanNode, ListNodes. Kept as a single
-// const so adding a field is a three-line change (const + values + scan).
-const nodeColumns = `id, cluster_id, name, display_name, role,
-	kubelet_version, kube_proxy_version, container_runtime_version,
-	os_image, operating_system, kernel_version, architecture,
-	internal_ip, external_ip, pod_cidr,
-	provider_id, instance_type, zone,
-	capacity_cpu, capacity_memory, capacity_pods, capacity_ephemeral_storage,
-	allocatable_cpu, allocatable_memory, allocatable_pods, allocatable_ephemeral_storage,
-	conditions, taints, unschedulable, ready,
-	labels,
-	owner, criticality, notes, runbook_url, annotations, hardware_model,
-	created_at, updated_at`
-
-// nodeColumnsUAliased is nodeColumns with every column qualified by the `u`
-// alias plus the denormalized cluster_name from the LEFT JOIN — used by
-// UpsertNode's final SELECT, which joins `upserted u` against the snapshot
-// CTE `old o` and clusters `c`, and must disambiguate columns present in
-// more than one source.
-const nodeColumnsUAliased = `u.id, u.cluster_id, u.name, u.display_name, u.role,
-	u.kubelet_version, u.kube_proxy_version, u.container_runtime_version,
-	u.os_image, u.operating_system, u.kernel_version, u.architecture,
-	u.internal_ip, u.external_ip, u.pod_cidr,
-	u.provider_id, u.instance_type, u.zone,
-	u.capacity_cpu, u.capacity_memory, u.capacity_pods, u.capacity_ephemeral_storage,
-	u.allocatable_cpu, u.allocatable_memory, u.allocatable_pods, u.allocatable_ephemeral_storage,
-	u.conditions, u.taints, u.unschedulable, u.ready,
-	u.labels,
-	u.owner, u.criticality, u.notes, u.runbook_url, u.annotations, u.hardware_model,
-	u.created_at, u.updated_at,
-	c.name AS cluster_name`
-
-// nodeSelectColumns is the read-side projection: every column from the
-// nodes row (aliased to `n.`) plus the denormalized cluster name joined
-// in via LEFT JOIN (ADR-0027). LEFT JOIN keeps orphan rows visible with
-// a NULL cluster_name — the UI renders that as an explicit badge.
-const nodeSelectColumns = `n.id, n.cluster_id, n.name, n.display_name, n.role,
-	n.kubelet_version, n.kube_proxy_version, n.container_runtime_version,
-	n.os_image, n.operating_system, n.kernel_version, n.architecture,
-	n.internal_ip, n.external_ip, n.pod_cidr,
-	n.provider_id, n.instance_type, n.zone,
-	n.capacity_cpu, n.capacity_memory, n.capacity_pods, n.capacity_ephemeral_storage,
-	n.allocatable_cpu, n.allocatable_memory, n.allocatable_pods, n.allocatable_ephemeral_storage,
-	n.conditions, n.taints, n.unschedulable, n.ready,
-	n.labels,
-	n.owner, n.criticality, n.notes, n.runbook_url, n.annotations, n.hardware_model,
-	n.created_at, n.updated_at,
-	c.name AS cluster_name`
-
-const nodeFromJoined = `FROM nodes n
-	LEFT JOIN clusters c ON c.id = n.cluster_id`
-
-func nodeInsertValues(in *api.NodeCreate, id uuid.UUID, now time.Time) ([]any, error) {
-	labelsJSON, err := marshalLabels(in.Labels)
-	if err != nil {
-		return nil, err
-	}
-	annotationsJSON, err := marshalLabels(in.Annotations)
-	if err != nil {
-		return nil, fmt.Errorf("marshal node annotations: %w", err)
-	}
-	conditionsJSON, err := marshalPorts(in.Conditions)
-	if err != nil {
-		return nil, err
-	}
-	taintsJSON, err := marshalPorts(in.Taints)
-	if err != nil {
-		return nil, err
-	}
-	return []any{
-		id, in.ClusterId, in.Name, in.DisplayName, in.Role,
-		in.KubeletVersion, in.KubeProxyVersion, in.ContainerRuntimeVersion,
-		in.OsImage, in.OperatingSystem, in.KernelVersion, in.Architecture,
-		in.InternalIp, in.ExternalIp, in.PodCidr,
-		in.ProviderId, in.InstanceType, in.Zone,
-		in.CapacityCpu, in.CapacityMemory, in.CapacityPods, in.CapacityEphemeralStorage,
-		in.AllocatableCpu, in.AllocatableMemory, in.AllocatablePods, in.AllocatableEphemeralStorage,
-		conditionsJSON, taintsJSON, boolOrFalse(in.Unschedulable), boolOrFalse(in.Ready),
-		labelsJSON,
-		in.Owner, in.Criticality, in.Notes, in.RunbookUrl, annotationsJSON, in.HardwareModel,
-		now,
-	}, nil
-}
-
-func boolOrFalse(b *bool) bool {
-	if b == nil {
-		return false
-	}
-	return *b
-}
-
-// CreateNode inserts a new node into the given cluster.
-//
-//nolint:gocritic // hugeParam: Store interface requires value param
-func (p *PG) CreateNode(ctx context.Context, in api.NodeCreate) (api.Node, error) {
-	id := uuid.New()
-	now := time.Now().UTC()
-
-	values, err := nodeInsertValues(&in, id, now)
-	if err != nil {
-		return api.Node{}, err
-	}
-
-	// 38 placeholders: 36 "value" slots + created_at + updated_at (both = $38).
-	const q = `
-		INSERT INTO nodes (` + nodeColumns + `)
-		VALUES ($1,$2,$3,$4,$5,$6,$7,$8,$9,$10,$11,$12,$13,$14,$15,$16,$17,$18,$19,$20,$21,$22,$23,$24,$25,$26,$27,$28,$29,$30,$31,$32,$33,$34,$35,$36,$37,$38,$38)
-	`
-	if _, err := p.pool.Exec(ctx, q, values...); err != nil {
-		var pgErr *pgconn.PgError
-		if errors.As(err, &pgErr) {
-			switch pgErr.Code {
-			case "23505":
-				return api.Node{}, fmt.Errorf("node %q in cluster %s already exists: %w", in.Name, in.ClusterId, api.ErrConflict)
-			case "23503":
-				return api.Node{}, fmt.Errorf("cluster %s does not exist: %w", in.ClusterId, api.ErrNotFound)
-			}
-		}
-		return api.Node{}, fmt.Errorf("insert node: %w", err)
-	}
-
-	return p.GetNode(ctx, id)
-}
-
-// GetNode fetches a node by id, including the denormalized cluster_name
-// from the node's cluster (ADR-0027).
-func (p *PG) GetNode(ctx context.Context, id uuid.UUID) (api.Node, error) {
-	q := `SELECT ` + nodeSelectColumns + ` ` + nodeFromJoined + ` WHERE n.id = $1`
-	row := p.pool.QueryRow(ctx, q, id)
-	n, err := scanNode(row)
-	if err != nil {
-		if errors.Is(err, pgx.ErrNoRows) {
-			return api.Node{}, api.ErrNotFound
-		}
-		return api.Node{}, fmt.Errorf("select node: %w", err)
-	}
-	return n, nil
-}
-
-// ListNodes returns up to limit nodes sorted (created_at DESC, id DESC),
-// optionally filtered by cluster id.
-//
-//nolint:gocyclo // cursor-paginated query builder with optional filters
-func (p *PG) ListNodes(ctx context.Context, clusterID *uuid.UUID, limit int, cursor string, includeTerminated bool) ([]api.Node, string, error) {
-	if limit <= 0 {
-		limit = 50
-	}
-	if limit > 200 {
-		limit = 200
-	}
-
-	sb := strings.Builder{}
-	sb.WriteString(`SELECT `)
-	sb.WriteString(nodeSelectColumns)
-	sb.WriteString(` `)
-	sb.WriteString(nodeFromJoined)
-	args := make([]any, 0, 4)
-	conds := make([]string, 0, 3)
-
-	if !includeTerminated {
-		conds = append(conds, "n.terminated_at IS NULL")
-	}
-	if clusterID != nil {
-		args = append(args, *clusterID)
-		conds = append(conds, fmt.Sprintf("n.cluster_id = $%d", len(args)))
-	}
-	if cursor != "" {
-		ts, cid, err := decodeCursor(cursor)
-		if err != nil {
-			return nil, "", err
-		}
-		args = append(args, ts)
-		tsIdx := len(args)
-		args = append(args, cid)
-		idIdx := len(args)
-		conds = append(conds, fmt.Sprintf("(n.created_at, n.id) < ($%d, $%d)", tsIdx, idIdx))
-	}
-	if len(conds) > 0 {
-		sb.WriteString(" WHERE ")
-		sb.WriteString(strings.Join(conds, " AND "))
-	}
-	args = append(args, limit+1)
-	fmt.Fprintf(&sb, " ORDER BY n.created_at DESC, n.id DESC LIMIT $%d", len(args))
-
-	rows, err := p.pool.Query(ctx, sb.String(), args...)
-	if err != nil {
-		return nil, "", fmt.Errorf("query nodes: %w", err)
-	}
-	defer rows.Close()
-
-	items := make([]api.Node, 0, limit)
-	for rows.Next() {
-		n, err := scanNode(rows)
-		if err != nil {
-			return nil, "", fmt.Errorf("scan node: %w", err)
-		}
-		items = append(items, n)
-	}
-	if err := rows.Err(); err != nil {
-		return nil, "", fmt.Errorf("iterate nodes: %w", err)
-	}
-
-	var next string
-	if len(items) > limit {
-		last := items[limit-1]
-		if last.CreatedAt != nil && last.Id != nil {
-			next = encodeCursor(*last.CreatedAt, *last.Id)
-		}
-		items = items[:limit]
-	}
-	return items, next, nil
-}
-
-// buildNodeUpdateSets converts a NodeUpdate merge-patch into the SET clause
-// fragments and positional arg slice consumed by UpdateNode. A trailing
-// updated_at set is always appended; the caller owns the id arg for the
-// WHERE clause and the surrounding transaction.
-func buildNodeUpdateSets(in *api.NodeUpdate) (sets []string, args []any, err error) {
-	sets = make([]string, 0, 24)
-	args = make([]any, 0, 26)
-	add := func(column string, value any) {
-		sets = append(sets, fmt.Sprintf("%s=$%d", column, len(sets)+1))
-		args = append(args, value)
-	}
-	addPtr := func(column string, ptr any) {
-		v := reflect.ValueOf(ptr)
-		if v.Kind() == reflect.Ptr && !v.IsNil() {
-			add(column, v.Elem().Interface())
-		}
-	}
-
-	addPtr("display_name", in.DisplayName)
-	addPtr("role", in.Role)
-	addPtr("kubelet_version", in.KubeletVersion)
-	addPtr("kube_proxy_version", in.KubeProxyVersion)
-	addPtr("container_runtime_version", in.ContainerRuntimeVersion)
-	addPtr("os_image", in.OsImage)
-	addPtr("operating_system", in.OperatingSystem)
-	addPtr("kernel_version", in.KernelVersion)
-	addPtr("architecture", in.Architecture)
-	addPtr("internal_ip", in.InternalIp)
-	addPtr("external_ip", in.ExternalIp)
-	addPtr("pod_cidr", in.PodCidr)
-	addPtr("provider_id", in.ProviderId)
-	addPtr("instance_type", in.InstanceType)
-	addPtr("zone", in.Zone)
-	addPtr("capacity_cpu", in.CapacityCpu)
-	addPtr("capacity_memory", in.CapacityMemory)
-	addPtr("capacity_pods", in.CapacityPods)
-	addPtr("capacity_ephemeral_storage", in.CapacityEphemeralStorage)
-	addPtr("allocatable_cpu", in.AllocatableCpu)
-	addPtr("allocatable_memory", in.AllocatableMemory)
-	addPtr("allocatable_pods", in.AllocatablePods)
-	addPtr("allocatable_ephemeral_storage", in.AllocatableEphemeralStorage)
-	var jsonErr error
-	addJSON := func(column string, ptr any, marshal func() ([]byte, error)) {
-		if jsonErr != nil {
-			return
-		}
-		v := reflect.ValueOf(ptr)
-		if v.Kind() != reflect.Ptr || v.IsNil() {
-			return
-		}
-		b, err := marshal()
-		if err != nil {
-			jsonErr = err
-			return
-		}
-		add(column, b)
-	}
-	addJSON("conditions", in.Conditions, func() ([]byte, error) { return marshalPorts(in.Conditions) })
-	addJSON("taints", in.Taints, func() ([]byte, error) { return marshalPorts(in.Taints) })
-	addPtr("unschedulable", in.Unschedulable)
-	addPtr("ready", in.Ready)
-	addJSON("labels", in.Labels, func() ([]byte, error) { return marshalLabels(in.Labels) })
-	// Curated metadata — collector never writes these, so merge-patch
-	// omission is enough to keep operator edits safe across polls.
-	addPtr("owner", in.Owner)
-	addPtr("criticality", in.Criticality)
-	addPtr("notes", in.Notes)
-	addPtr("runbook_url", in.RunbookUrl)
-	addJSON("annotations", in.Annotations, func() ([]byte, error) {
-		b, err := marshalLabels(in.Annotations)
-		if err != nil {
-			return nil, fmt.Errorf("marshal node annotations: %w", err)
-		}
-		return b, nil
-	})
-	if jsonErr != nil {
-		return nil, nil, jsonErr
-	}
-	addPtr("hardware_model", in.HardwareModel)
-	add("updated_at", time.Now().UTC())
-	return sets, args, nil
-}
-
-// UpdateNode applies merge-patch semantics on mutable fields only. Each
-// non-nil pointer on NodeUpdate translates to a single column set; omitted
-// fields keep their existing value.
-//
-//nolint:gocritic // hugeParam: Store interface requires value param
-func (p *PG) UpdateNode(ctx context.Context, id uuid.UUID, in api.NodeUpdate) (api.Node, error) {
-	sets, args, err := buildNodeUpdateSets(&in)
-	if err != nil {
-		return api.Node{}, err
-	}
-	idx := len(sets) + 1
-	args = append(args, id)
-
-	tx, txErr := p.pool.Begin(ctx)
-	if txErr != nil {
-		return api.Node{}, fmt.Errorf("begin update node: %w", txErr)
-	}
-	defer func() { _ = tx.Rollback(ctx) }()
-
-	prev, _ := nodeRowMap(ctx, tx, id) // FOR UPDATE; ignore error for capture
-	if prev == nil {
-		return api.Node{}, api.ErrNotFound
-	}
-
-	q := fmt.Sprintf("UPDATE nodes SET %s WHERE id=$%d", strings.Join(sets, ", "), idx)
-	tag, err := tx.Exec(ctx, q, args...)
-	if err != nil {
-		return api.Node{}, fmt.Errorf("update node: %w", err)
-	}
-	if tag.RowsAffected() == 0 {
-		return api.Node{}, api.ErrNotFound
-	}
-
-	if next, err := nodeRowMapNoLock(ctx, tx, id); err == nil {
-		actor := timetravel.ActorFromContext(ctx)
-		_ = timetravel.Capture(ctx, tx, timetravel.KindNode, id, prev, next, changeTypeUpdate, actor)
-	}
-
-	if err := tx.Commit(ctx); err != nil {
-		return api.Node{}, fmt.Errorf("commit update node: %w", err)
-	}
-	return p.GetNode(ctx, id)
-}
-
-// DeleteNodesNotIn soft-deletes every node of the given cluster whose name
-// is not in keepNames AND that is not already terminated. Returns the number
-// of rows newly soft-deleted. Despite the name, this is a soft-delete: per
-// ADR-0021 §5 the row stays in the table with terminated_at = NOW() so list
-// queries can opt back in via include_terminated, history (Phase 2) can
-// reconstruct the lifecycle, and a re-appearing node resurrects via the
-// upsert path. keepNames may be nil or empty (soft-deletes every live node
-// for the cluster).
-//
-// COALESCE guards against pgx encoding a nil []string as SQL NULL: without
-// it, 'name <> ALL(NULL)' evaluates to NULL and the UPDATE matches nothing
-// instead of clearing the cluster's nodes.
-func (p *PG) DeleteNodesNotIn(ctx context.Context, clusterID uuid.UUID, keepNames []string) (int64, error) {
-	tx, err := p.pool.Begin(ctx)
-	if err != nil {
-		return 0, fmt.Errorf("begin delete nodes not in: %w", err)
-	}
-	defer func() { _ = tx.Rollback(ctx) }()
-
-	// Collect IDs that will be soft-deleted so we can write history after the UPDATE.
-	toDeleteRows, err := tx.Query(ctx,
-		`SELECT id FROM nodes
-		  WHERE cluster_id = $1
-		    AND name <> ALL(COALESCE($2::text[], ARRAY[]::text[]))
-		    AND terminated_at IS NULL`,
-		clusterID, keepNames)
-	if err != nil {
-		return 0, fmt.Errorf("list nodes to soft-delete: %w", err)
-	}
-	toDelete, err := scanUUIDs(toDeleteRows)
-	if err != nil {
-		return 0, fmt.Errorf("scan nodes to soft-delete: %w", err)
-	}
-
-	tag, err := tx.Exec(ctx,
-		`UPDATE nodes
-		    SET terminated_at = NOW(), updated_at = NOW()
-		  WHERE cluster_id = $1
-		    AND name <> ALL(COALESCE($2::text[], ARRAY[]::text[]))
-		    AND terminated_at IS NULL`,
-		clusterID, keepNames,
-	)
-	if err != nil {
-		return 0, fmt.Errorf("soft-delete nodes not in: %w", err)
-	}
-
-	actor := timetravel.ActorFromContext(ctx)
-	for _, nodeID := range toDelete {
-		if snap, err := nodeRowMapNoLock(ctx, tx, nodeID); err == nil {
-			_ = timetravel.Capture(ctx, tx, timetravel.KindNode, nodeID, nil, snap, changeTypeSoftDelete, actor)
-		}
-	}
-
-	if err := tx.Commit(ctx); err != nil {
-		return 0, fmt.Errorf("commit delete nodes not in: %w", err)
-	}
-	return tag.RowsAffected(), nil
-}
-
-// DeleteNamespacesNotIn mirrors DeleteNodesNotIn: soft-deletes namespaces of
-// the given cluster not in keepNames and not already terminated. Per
-// ADR-0021 §5. For each soft-deleted namespace, this mirrors
-// SoftDeleteNamespace by cascade-soft-deleting its live workloads and
-// hard-deleting its unhistoried children (pods, services, ingresses, PVCs)
-// — otherwise workloads in a vanished namespace would linger forever in
-// list/EOL outputs because reconcileWorkloads only walks namespaces that
-// were present in the current tick.
-//
-//nolint:gocyclo // cascade + history capture adds branches; mirrors SoftDeleteNamespace
-func (p *PG) DeleteNamespacesNotIn(ctx context.Context, clusterID uuid.UUID, keepNames []string) (int64, error) {
-	tx, err := p.pool.Begin(ctx)
-	if err != nil {
-		return 0, fmt.Errorf("begin delete namespaces not in: %w", err)
-	}
-	defer func() { _ = tx.Rollback(ctx) }()
-
-	toDeleteRows, err := tx.Query(ctx,
-		`SELECT id FROM namespaces
-		  WHERE cluster_id = $1
-		    AND name <> ALL(COALESCE($2::text[], ARRAY[]::text[]))
-		    AND terminated_at IS NULL`,
-		clusterID, keepNames)
-	if err != nil {
-		return 0, fmt.Errorf("list namespaces to soft-delete: %w", err)
-	}
-	toDelete, err := scanUUIDs(toDeleteRows)
-	if err != nil {
-		return 0, fmt.Errorf("scan namespaces to soft-delete: %w", err)
-	}
-
-	actor := timetravel.ActorFromContext(ctx)
-
-	// Per-namespace cascade: capture each namespace's live workload IDs
-	// before mutating, hard-delete unhistoried children, soft-delete
-	// workloads, then capture workload snapshots so history reflects the
-	// final terminated state.
-	wlIDsByNS := make(map[uuid.UUID][]uuid.UUID, len(toDelete))
-	for _, nsID := range toDelete {
-		wlIDs, err := liveWorkloadIDsForNamespace(ctx, tx, nsID)
-		if err != nil {
-			return 0, fmt.Errorf("list workloads for soft-delete namespace %s: %w", nsID, err)
-		}
-		wlIDsByNS[nsID] = wlIDs
-
-		if _, err := tx.Exec(ctx,
-			`DELETE FROM pods WHERE namespace_id = $1`, nsID); err != nil {
-			return 0, fmt.Errorf("cascade-delete pods for %s: %w", nsID, err)
-		}
-		if _, err := tx.Exec(ctx,
-			`DELETE FROM services WHERE namespace_id = $1`, nsID); err != nil {
-			return 0, fmt.Errorf("cascade-delete services for %s: %w", nsID, err)
-		}
-		if _, err := tx.Exec(ctx,
-			`DELETE FROM ingresses WHERE namespace_id = $1`, nsID); err != nil {
-			return 0, fmt.Errorf("cascade-delete ingresses for %s: %w", nsID, err)
-		}
-		if _, err := tx.Exec(ctx,
-			`DELETE FROM persistent_volume_claims WHERE namespace_id = $1`, nsID); err != nil {
-			return 0, fmt.Errorf("cascade-delete persistent_volume_claims for %s: %w", nsID, err)
-		}
-		if _, err := tx.Exec(ctx,
-			`UPDATE workloads SET terminated_at = NOW(), updated_at = NOW()
-			   WHERE namespace_id = $1 AND terminated_at IS NULL`, nsID); err != nil {
-			return 0, fmt.Errorf("soft-delete workloads for %s: %w", nsID, err)
-		}
-	}
-
-	tag, err := tx.Exec(ctx,
-		`UPDATE namespaces
-		    SET terminated_at = NOW(), updated_at = NOW()
-		  WHERE cluster_id = $1
-		    AND name <> ALL(COALESCE($2::text[], ARRAY[]::text[]))
-		    AND terminated_at IS NULL`,
-		clusterID, keepNames,
-	)
-	if err != nil {
-		return 0, fmt.Errorf("soft-delete namespaces not in: %w", err)
-	}
-
-	for _, nsID := range toDelete {
-		for _, wlID := range wlIDsByNS[nsID] {
-			if snap, err := workloadRowMapNoLock(ctx, tx, wlID); err == nil {
-				_ = timetravel.Capture(ctx, tx, timetravel.KindWorkload, wlID, nil, snap, changeTypeSoftDelete, actor)
-			}
-		}
-		if snap, err := namespaceRowMapNoLock(ctx, tx, nsID); err == nil {
-			_ = timetravel.Capture(ctx, tx, timetravel.KindNamespace, nsID, nil, snap, changeTypeSoftDelete, actor)
-		}
-	}
-
-	if err := tx.Commit(ctx); err != nil {
-		return 0, fmt.Errorf("commit delete namespaces not in: %w", err)
-	}
-	return tag.RowsAffected(), nil
-}
-
-// DeleteNode removes a node by id.
-func (p *PG) DeleteNode(ctx context.Context, id uuid.UUID) error {
-	tag, err := p.pool.Exec(ctx, "DELETE FROM nodes WHERE id=$1", id)
-	if err != nil {
-		return fmt.Errorf("delete node: %w", err)
-	}
-	if tag.RowsAffected() == 0 {
-		return api.ErrNotFound
-	}
-	return nil
-}
-
-// CreateNamespace inserts a new namespace.
-//
-//nolint:gocritic // hugeParam: Store interface requires value param
-func (p *PG) CreateNamespace(ctx context.Context, in api.NamespaceCreate) (api.Namespace, error) {
-	id := uuid.New()
-	now := time.Now().UTC()
-
-	labelsJSON, err := marshalLabels(in.Labels)
-	if err != nil {
-		return api.Namespace{}, err
-	}
-	annotationsJSON, err := marshalLabels(in.Annotations)
-	if err != nil {
-		return api.Namespace{}, fmt.Errorf("marshal namespace annotations: %w", err)
-	}
-
-	const q = `
-		INSERT INTO namespaces (
-			id, cluster_id, name, display_name, phase, labels,
-			owner, criticality, notes, runbook_url, annotations,
-			created_at, updated_at
-		) VALUES ($1, $2, $3, $4, $5, $6, $7, $8, $9, $10, $11, $12, $12)
-	`
-	_, err = p.pool.Exec(ctx, q,
-		id, in.ClusterId, in.Name, in.DisplayName, in.Phase, labelsJSON,
-		in.Owner, in.Criticality, in.Notes, in.RunbookUrl, annotationsJSON,
-		now,
-	)
-	if err != nil {
-		var pgErr *pgconn.PgError
-		if errors.As(err, &pgErr) {
-			switch pgErr.Code {
-			case "23505":
-				return api.Namespace{}, fmt.Errorf("namespace %q in cluster %s already exists: %w", in.Name, in.ClusterId, api.ErrConflict)
-			case "23503":
-				return api.Namespace{}, fmt.Errorf("cluster %s does not exist: %w", in.ClusterId, api.ErrNotFound)
-			}
-		}
-		return api.Namespace{}, fmt.Errorf("insert namespace: %w", err)
-	}
-
-	return api.Namespace{
-		Id:          &id,
-		ClusterId:   in.ClusterId,
-		Name:        in.Name,
-		DisplayName: in.DisplayName,
-		Phase:       in.Phase,
-		Labels:      in.Labels,
-		Owner:       in.Owner,
-		Criticality: in.Criticality,
-		Notes:       in.Notes,
-		RunbookUrl:  in.RunbookUrl,
-		Annotations: in.Annotations,
-		CreatedAt:   &now,
-		UpdatedAt:   &now,
-	}, nil
-}
-
-// GetNamespace fetches a namespace by id.
-// Namespace read-side projection with cluster_name JOINed in (ADR-0027).
-const namespaceSelectColumns = `n.id, n.cluster_id, n.name, n.display_name, n.phase, n.labels,
-	n.owner, n.criticality, n.notes, n.runbook_url, n.annotations,
-	n.created_at, n.updated_at,
-	c.name AS cluster_name`
-
-const namespaceFromJoined = `FROM namespaces n
-	LEFT JOIN clusters c ON c.id = n.cluster_id`
-
-// GetNamespace fetches a namespace by id, including the denormalized
-// cluster_name from the namespace's cluster (ADR-0027).
-func (p *PG) GetNamespace(ctx context.Context, id uuid.UUID) (api.Namespace, error) {
-	q := `SELECT ` + namespaceSelectColumns + ` ` + namespaceFromJoined + ` WHERE n.id = $1`
-	row := p.pool.QueryRow(ctx, q, id)
-	n, err := scanNamespace(row)
-	if err != nil {
-		if errors.Is(err, pgx.ErrNoRows) {
-			return api.Namespace{}, api.ErrNotFound
-		}
-		return api.Namespace{}, fmt.Errorf("select namespace: %w", err)
-	}
-	return n, nil
-}
-
-// ListNamespaces returns up to limit namespaces sorted (created_at DESC, id DESC).
-//
-//nolint:gocyclo // cursor-paginated query builder with optional filters
-func (p *PG) ListNamespaces(
-	ctx context.Context,
-	clusterID *uuid.UUID,
-	limit int,
-	cursor string,
-	includeTerminated bool,
-) ([]api.Namespace, string, error) {
-	if limit <= 0 {
-		limit = 50
-	}
-	if limit > 200 {
-		limit = 200
-	}
-
-	sb := strings.Builder{}
-	sb.WriteString(`SELECT `)
-	sb.WriteString(namespaceSelectColumns)
-	sb.WriteString(` `)
-	sb.WriteString(namespaceFromJoined)
-	args := make([]any, 0, 4)
-	conds := make([]string, 0, 3)
-
-	if !includeTerminated {
-		conds = append(conds, "n.terminated_at IS NULL")
-	}
-	if clusterID != nil {
-		args = append(args, *clusterID)
-		conds = append(conds, fmt.Sprintf("n.cluster_id = $%d", len(args)))
-	}
-	if cursor != "" {
-		ts, cid, err := decodeCursor(cursor)
-		if err != nil {
-			return nil, "", err
-		}
-		args = append(args, ts)
-		tsIdx := len(args)
-		args = append(args, cid)
-		idIdx := len(args)
-		conds = append(conds, fmt.Sprintf("(n.created_at, n.id) < ($%d, $%d)", tsIdx, idIdx))
-	}
-	if len(conds) > 0 {
-		sb.WriteString(" WHERE ")
-		sb.WriteString(strings.Join(conds, " AND "))
-	}
-	args = append(args, limit+1)
-	fmt.Fprintf(&sb, " ORDER BY n.created_at DESC, n.id DESC LIMIT $%d", len(args))
-
-	rows, err := p.pool.Query(ctx, sb.String(), args...)
-	if err != nil {
-		return nil, "", fmt.Errorf("query namespaces: %w", err)
-	}
-	defer rows.Close()
-
-	items := make([]api.Namespace, 0, limit)
-	for rows.Next() {
-		n, err := scanNamespace(rows)
-		if err != nil {
-			return nil, "", fmt.Errorf("scan namespace: %w", err)
-		}
-		items = append(items, n)
-	}
-	if err := rows.Err(); err != nil {
-		return nil, "", fmt.Errorf("iterate namespaces: %w", err)
-	}
-
-	var next string
-	if len(items) > limit {
-		last := items[limit-1]
-		if last.CreatedAt != nil && last.Id != nil {
-			next = encodeCursor(*last.CreatedAt, *last.Id)
-		}
-		items = items[:limit]
-	}
-	return items, next, nil
-}
-
-// UpdateNamespace applies merge-patch semantics on mutable fields.
-//
-//nolint:gocyclo // merge-patch nil checks are inherently repetitive
-func (p *PG) UpdateNamespace(ctx context.Context, id uuid.UUID, in api.NamespaceUpdate) (api.Namespace, error) {
-	sets := make([]string, 0, 4)
-	args := make([]any, 0, 6)
-	idx := 1
-	appendSet := func(column string, value any) {
-		sets = append(sets, fmt.Sprintf("%s=$%d", column, idx))
-		args = append(args, value)
-		idx++
-	}
-
-	if in.DisplayName != nil {
-		appendSet("display_name", *in.DisplayName)
-	}
-	if in.Phase != nil {
-		appendSet("phase", *in.Phase)
-	}
-	if in.Labels != nil {
-		b, err := marshalLabels(in.Labels)
-		if err != nil {
-			return api.Namespace{}, err
-		}
-		appendSet("labels", b)
-	}
-	// Curated metadata — collector never writes these, so merge-patch
-	// omission is enough to keep operator edits safe across polls.
-	if in.Owner != nil {
-		appendSet("owner", *in.Owner)
-	}
-	if in.Criticality != nil {
-		appendSet("criticality", *in.Criticality)
-	}
-	if in.Notes != nil {
-		appendSet("notes", *in.Notes)
-	}
-	if in.RunbookUrl != nil {
-		appendSet("runbook_url", *in.RunbookUrl)
-	}
-	if in.Annotations != nil {
-		b, err := marshalLabels(in.Annotations)
-		if err != nil {
-			return api.Namespace{}, fmt.Errorf("marshal namespace annotations: %w", err)
-		}
-		appendSet("annotations", b)
-	}
-	appendSet("updated_at", time.Now().UTC())
-	args = append(args, id)
-
-	tx, txErr := p.pool.Begin(ctx)
-	if txErr != nil {
-		return api.Namespace{}, fmt.Errorf("begin update namespace: %w", txErr)
-	}
-	defer func() { _ = tx.Rollback(ctx) }()
-
-	prev, _ := namespaceRowMap(ctx, tx, id) // FOR UPDATE
-
-	q := fmt.Sprintf("UPDATE namespaces SET %s WHERE id=$%d", strings.Join(sets, ", "), idx)
-	tag, err := tx.Exec(ctx, q, args...)
-	if err != nil {
-		return api.Namespace{}, fmt.Errorf("update namespace: %w", err)
-	}
-	if tag.RowsAffected() == 0 {
-		return api.Namespace{}, api.ErrNotFound
-	}
-
-	if prev != nil {
-		if next, err := namespaceRowMapNoLock(ctx, tx, id); err == nil {
-			actor := timetravel.ActorFromContext(ctx)
-			_ = timetravel.Capture(ctx, tx, timetravel.KindNamespace, id, prev, next, changeTypeUpdate, actor)
-		}
-	}
-
-	if err := tx.Commit(ctx); err != nil {
-		return api.Namespace{}, fmt.Errorf("commit update namespace: %w", err)
-	}
-	return p.GetNamespace(ctx, id)
-}
-
-// DeleteNamespace removes a namespace by id.
-func (p *PG) DeleteNamespace(ctx context.Context, id uuid.UUID) error {
-	tag, err := p.pool.Exec(ctx, "DELETE FROM namespaces WHERE id=$1", id)
-	if err != nil {
-		return fmt.Errorf("delete namespace: %w", err)
-	}
-	if tag.RowsAffected() == 0 {
-		return api.ErrNotFound
-	}
-	return nil
-}
-
-// SoftDeleteNamespace soft-deletes the namespace and its workloads, and
-// hard-deletes its unhistoried children (pods, services, ingresses, PVCs).
-// PVs are cluster-scoped, not namespace-scoped, so they are unaffected.
-//
-//nolint:gocyclo // history capture adds branches; acceptable here
-func (p *PG) SoftDeleteNamespace(ctx context.Context, id uuid.UUID) error {
-	tx, err := p.pool.Begin(ctx)
-	if err != nil {
-		return fmt.Errorf("begin: %w", err)
-	}
-	defer func() { _ = tx.Rollback(ctx) }()
-
-	actor := timetravel.ActorFromContext(ctx)
-
-	wlIDs, err := liveWorkloadIDsForNamespace(ctx, tx, id)
-	if err != nil {
-		return fmt.Errorf("list workloads for soft-delete namespace: %w", err)
-	}
-
-	// Hard-delete the unhistoried namespace-scoped children. FK ON DELETE
-	// CASCADE does not fire under soft-delete; do it manually.
-	if _, err := tx.Exec(ctx,
-		`DELETE FROM pods WHERE namespace_id = $1`, id); err != nil {
-		return fmt.Errorf("cascade-delete pods: %w", err)
-	}
-	if _, err := tx.Exec(ctx,
-		`DELETE FROM services WHERE namespace_id = $1`, id); err != nil {
-		return fmt.Errorf("cascade-delete services: %w", err)
-	}
-	if _, err := tx.Exec(ctx,
-		`DELETE FROM ingresses WHERE namespace_id = $1`, id); err != nil {
-		return fmt.Errorf("cascade-delete ingresses: %w", err)
-	}
-	if _, err := tx.Exec(ctx,
-		`DELETE FROM persistent_volume_claims WHERE namespace_id = $1`, id); err != nil {
-		return fmt.Errorf("cascade-delete persistent_volume_claims: %w", err)
-	}
-
-	if _, err := tx.Exec(ctx,
-		`UPDATE workloads SET terminated_at = NOW(), updated_at = NOW()
-		   WHERE namespace_id = $1 AND terminated_at IS NULL`, id); err != nil {
-		return fmt.Errorf("soft-delete workloads: %w", err)
-	}
-	tag, err := tx.Exec(ctx,
-		`UPDATE namespaces SET terminated_at = NOW(), updated_at = NOW()
-		   WHERE id = $1 AND terminated_at IS NULL`, id)
-	if err != nil {
-		return fmt.Errorf("soft-delete namespace: %w", err)
-	}
-	if tag.RowsAffected() == 0 {
-		var exists bool
-		if err := tx.QueryRow(ctx,
-			`SELECT EXISTS(SELECT 1 FROM namespaces WHERE id = $1)`, id).Scan(&exists); err != nil {
-			return fmt.Errorf("check namespace exists: %w", err)
-		}
-		if !exists {
-			return api.ErrNotFound
-		}
-	}
-
-	for _, wlID := range wlIDs {
-		if snap, err := workloadRowMapNoLock(ctx, tx, wlID); err == nil {
-			_ = timetravel.Capture(ctx, tx, timetravel.KindWorkload, wlID, nil, snap, changeTypeSoftDelete, actor)
-		}
-	}
-	if snap, err := namespaceRowMapNoLock(ctx, tx, id); err == nil {
-		_ = timetravel.Capture(ctx, tx, timetravel.KindNamespace, id, nil, snap, changeTypeSoftDelete, actor)
-	}
-
-	if err := tx.Commit(ctx); err != nil {
-		return fmt.Errorf("commit soft-delete namespace: %w", err)
-	}
-	return nil
-}
-
-func liveWorkloadIDsForNamespace(ctx context.Context, tx pgx.Tx, namespaceID uuid.UUID) ([]uuid.UUID, error) {
-	rows, err := tx.Query(ctx,
-		`SELECT id FROM workloads WHERE namespace_id = $1 AND terminated_at IS NULL`,
-		namespaceID)
-	if err != nil {
-		return nil, fmt.Errorf("query live workloads for namespace: %w", err)
-	}
-	return scanUUIDs(rows)
-}
-
-// UpsertNamespace inserts-or-updates a namespace keyed by (cluster_id, name).
-//
-//nolint:gocritic,gocyclo // hugeParam: Store interface requires value param; history capture adds branches
-func (p *PG) UpsertNamespace(ctx context.Context, in api.NamespaceCreate) (api.Namespace, api.UpsertOutcome, error) {
-	id := uuid.New()
-	now := time.Now().UTC()
-
-	labelsJSON, err := marshalLabels(in.Labels)
-	if err != nil {
-		return api.Namespace{}, api.OutcomeNoChange, err
-	}
-
-	tx, err := p.pool.Begin(ctx)
-	if err != nil {
-		return api.Namespace{}, api.OutcomeNoChange, fmt.Errorf("begin upsert namespace: %w", err)
-	}
-	defer func() { _ = tx.Rollback(ctx) }()
-
-	// Snapshot the row before the upsert to detect create vs update vs restore.
-	var prevTerminatedAt *time.Time
-	var prevID *uuid.UUID
-	_ = tx.QueryRow(ctx,
-		`SELECT id, terminated_at FROM namespaces WHERE cluster_id=$1 AND name=$2`,
-		in.ClusterId, in.Name,
-	).Scan(&prevID, &prevTerminatedAt)
-	isCreate := prevID == nil
-	isRestore := prevID != nil && prevTerminatedAt != nil
-
-	// AUDIT_BUSINESS_FIELDS: display_name, phase, labels, terminated_at
-	// (restore flips it). updated_at is a clock field — excluded.
-	// Curator fields (owner/criticality/notes/runbook_url/annotations) are
-	// not touched by the collector upsert and excluded from the OR-chain.
-	const q = `
-		WITH old AS (
-		  SELECT display_name, phase, labels, terminated_at
-		    FROM namespaces WHERE cluster_id=$2 AND name=$3
-		),
-		upserted AS (
-		  INSERT INTO namespaces (
-		    id, cluster_id, name, display_name, phase,
-		    labels, created_at, updated_at
-		  ) VALUES ($1, $2, $3, $4, $5, $6, $7, $7)
-		  ON CONFLICT (cluster_id, name) DO UPDATE SET
-		      display_name  = EXCLUDED.display_name,
-		      phase         = EXCLUDED.phase,
-		      labels        = EXCLUDED.labels,
-		      terminated_at = NULL,
-		      updated_at    = EXCLUDED.updated_at
-		  RETURNING id, cluster_id, name, display_name, phase, labels,
-		            owner, criticality, notes, runbook_url, annotations,
-		            created_at, updated_at, terminated_at, xmax
-		)
-		SELECT u.id, u.cluster_id, u.name, u.display_name, u.phase, u.labels,
-		       u.owner, u.criticality, u.notes, u.runbook_url, u.annotations,
-		       u.created_at, u.updated_at,
-		       c.name AS cluster_name,
-		       (u.xmax = 0) AS inserted,
-		       (u.xmax <> 0 AND (
-		           o.display_name  IS DISTINCT FROM u.display_name  OR
-		           o.phase         IS DISTINCT FROM u.phase         OR
-		           o.labels        IS DISTINCT FROM u.labels        OR
-		           o.terminated_at IS DISTINCT FROM u.terminated_at
-		       )) AS business_changed
-		  FROM upserted u
-		  LEFT JOIN old o      ON true
-		  LEFT JOIN clusters c ON c.id = u.cluster_id
-	`
-	row := tx.QueryRow(ctx, q,
-		id, in.ClusterId, in.Name, in.DisplayName, in.Phase,
-		labelsJSON, now,
-	)
-	var inserted, businessChanged bool
-	n, err := scanNamespace(scanRowWith{row: row, extra: []any{&inserted, &businessChanged}})
-	if err != nil {
-		var pgErr *pgconn.PgError
-		if errors.As(err, &pgErr) && pgErr.Code == "23503" {
-			return api.Namespace{}, api.OutcomeNoChange, fmt.Errorf("cluster %s does not exist: %w", in.ClusterId, api.ErrNotFound)
-		}
-		return api.Namespace{}, api.OutcomeNoChange, fmt.Errorf("upsert namespace: %w", err)
-	}
-
-	actualID := *n.Id
-	if snap, err := namespaceRowMapNoLock(ctx, tx, actualID); err == nil {
-		actor := timetravel.ActorFromContext(ctx)
-		changeType := changeTypeUpdate
-		if isCreate {
-			changeType = changeTypeCreate
-		} else if isRestore {
-			changeType = changeTypeRestore
-		}
-		_ = timetravel.Capture(ctx, tx, timetravel.KindNamespace, actualID, nil, snap, changeType, actor)
-	}
-
-	if err := tx.Commit(ctx); err != nil {
-		return api.Namespace{}, api.OutcomeNoChange, fmt.Errorf("commit upsert namespace: %w", err)
-	}
-	return n, classifyOutcome(inserted, businessChanged), nil
-}
-
-// CreatePod inserts a new pod.
-//
-//nolint:gocritic // hugeParam: Store interface requires value param
-func (p *PG) CreatePod(ctx context.Context, in api.PodCreate) (api.Pod, error) {
-	id := uuid.New()
-	now := time.Now().UTC()
-
-	labelsJSON, err := marshalLabels(in.Labels)
-	if err != nil {
-		return api.Pod{}, err
-	}
-	containersJSON, err := marshalPorts(in.Containers)
-	if err != nil {
-		return api.Pod{}, err
-	}
-
-	const q = `
-		INSERT INTO pods (
-			id, namespace_id, name, phase, node_name, pod_ip,
-			workload_id, containers, labels, created_at, updated_at
-		) VALUES ($1, $2, $3, $4, $5, $6, $7, $8, $9, $10, $10)
-	`
-	_, err = p.pool.Exec(ctx, q,
-		id, in.NamespaceId, in.Name, in.Phase, in.NodeName, in.PodIp,
-		in.WorkloadId, containersJSON, labelsJSON, now,
-	)
-	if err != nil {
-		if pErr := classifyPodFKError(err, in.NamespaceId, in.WorkloadId); pErr != nil {
-			return api.Pod{}, pErr
-		}
-		var pgErr *pgconn.PgError
-		if errors.As(err, &pgErr) && pgErr.Code == "23505" {
-			return api.Pod{}, fmt.Errorf("pod %q in namespace %s already exists: %w", in.Name, in.NamespaceId, api.ErrConflict)
-		}
-		return api.Pod{}, fmt.Errorf("insert pod: %w", err)
-	}
-
-	return api.Pod{
-		Id:          &id,
-		NamespaceId: in.NamespaceId,
-		Name:        in.Name,
-		Phase:       in.Phase,
-		NodeName:    in.NodeName,
-		PodIp:       in.PodIp,
-		WorkloadId:  in.WorkloadId,
-		Containers:  in.Containers,
-		Labels:      in.Labels,
-		CreatedAt:   &now,
-		UpdatedAt:   &now,
-	}, nil
-}
-
-// classifyPodFKError disambiguates 23503 foreign-key violations on the pods
-// table into namespace vs workload misses, so the handler can return an
-// accurate 404 message. PG auto-names FK constraints <table>_<column>_fkey;
-// we match on the column name in pgErr.ConstraintName.
-func classifyPodFKError(err error, namespaceID uuid.UUID, workloadID *uuid.UUID) error {
-	var pgErr *pgconn.PgError
-	if !errors.As(err, &pgErr) || pgErr.Code != "23503" {
-		return nil
-	}
-	if strings.Contains(pgErr.ConstraintName, "workload_id") {
-		target := "<nil>"
-		if workloadID != nil {
-			target = workloadID.String()
-		}
-		return fmt.Errorf("workload %s does not exist: %w", target, api.ErrNotFound)
-	}
-	return fmt.Errorf("namespace %s does not exist: %w", namespaceID, api.ErrNotFound)
-}
-
-// GetPod fetches a pod by id.
-// Pod read-side projection with parent-name JOINs (ADR-0027). Pods also
-// LEFT JOIN workloads to expose `workload_name`.
-const podSelectColumns = `p.id, p.namespace_id, p.name, p.phase, p.node_name, p.pod_ip,
-	p.workload_id, p.containers, p.labels, p.created_at, p.updated_at,
-	n.name AS namespace_name, n.cluster_id AS namespace_cluster_id, c.name AS cluster_name,
-	w.name AS workload_name`
-
-const podFromJoined = `FROM pods p
-	LEFT JOIN namespaces n ON n.id = p.namespace_id
-	LEFT JOIN clusters   c ON c.id = n.cluster_id
-	LEFT JOIN workloads  w ON w.id = p.workload_id`
-
-// GetPod fetches a pod by id, including the denormalized namespace_name,
-// cluster_id, cluster_name, and workload_name (ADR-0027).
-func (p *PG) GetPod(ctx context.Context, id uuid.UUID) (api.Pod, error) {
-	q := `SELECT ` + podSelectColumns + ` ` + podFromJoined + ` WHERE p.id = $1`
-	row := p.pool.QueryRow(ctx, q, id)
-	pod, err := scanPod(row)
-	if err != nil {
-		if errors.Is(err, pgx.ErrNoRows) {
-			return api.Pod{}, api.ErrNotFound
-		}
-		return api.Pod{}, fmt.Errorf("select pod: %w", err)
-	}
-	return pod, nil
-}
-
-// ListPods returns up to limit pods, optionally filtered by namespace.
-//
-//nolint:gocyclo // cursor-paginated query builder with optional filters
-func (p *PG) ListPods(ctx context.Context, filter api.PodListFilter, limit int, cursor string) ([]api.Pod, string, error) {
-	if limit <= 0 {
-		limit = 50
-	}
-	if limit > 200 {
-		limit = 200
-	}
-
-	sb := strings.Builder{}
-	sb.WriteString(`SELECT `)
-	sb.WriteString(podSelectColumns)
-	sb.WriteString(` `)
-	sb.WriteString(podFromJoined)
-	args := make([]any, 0, 6)
-	conds := make([]string, 0, 4)
-
-	if filter.NamespaceID != nil {
-		args = append(args, *filter.NamespaceID)
-		conds = append(conds, fmt.Sprintf("p.namespace_id = $%d", len(args)))
-	}
-	if filter.NodeName != nil {
-		args = append(args, *filter.NodeName)
-		conds = append(conds, fmt.Sprintf("p.node_name = $%d", len(args)))
-	}
-	if filter.WorkloadID != nil {
-		args = append(args, *filter.WorkloadID)
-		conds = append(conds, fmt.Sprintf("p.workload_id = $%d", len(args)))
-	}
-	if filter.ImageSubstring != nil && *filter.ImageSubstring != "" {
-		// Case-insensitive substring match against any container's `image`
-		// string. jsonb_array_elements unpacks the array so we can compare
-		// ->>'image' element-by-element; EXISTS short-circuits on first hit.
-		// The ILIKE pattern is built with %…% wrapping.
-		args = append(args, "%"+*filter.ImageSubstring+"%")
-		conds = append(conds, fmt.Sprintf(
-			"EXISTS (SELECT 1 FROM jsonb_array_elements(p.containers) elem WHERE elem->>'image' ILIKE $%d)",
-			len(args),
-		))
-	}
-	if cursor != "" {
-		ts, cid, err := decodeCursor(cursor)
-		if err != nil {
-			return nil, "", err
-		}
-		args = append(args, ts)
-		tsIdx := len(args)
-		args = append(args, cid)
-		idIdx := len(args)
-		conds = append(conds, fmt.Sprintf("(p.created_at, p.id) < ($%d, $%d)", tsIdx, idIdx))
-	}
-	if len(conds) > 0 {
-		sb.WriteString(" WHERE ")
-		sb.WriteString(strings.Join(conds, " AND "))
-	}
-	args = append(args, limit+1)
-	fmt.Fprintf(&sb, " ORDER BY p.created_at DESC, p.id DESC LIMIT $%d", len(args))
-
-	rows, err := p.pool.Query(ctx, sb.String(), args...)
-	if err != nil {
-		return nil, "", fmt.Errorf("query pods: %w", err)
-	}
-	defer rows.Close()
-
-	items := make([]api.Pod, 0, limit)
-	for rows.Next() {
-		pod, err := scanPod(rows)
-		if err != nil {
-			return nil, "", fmt.Errorf("scan pod: %w", err)
-		}
-		items = append(items, pod)
-	}
-	if err := rows.Err(); err != nil {
-		return nil, "", fmt.Errorf("iterate pods: %w", err)
-	}
-
-	var next string
-	if len(items) > limit {
-		last := items[limit-1]
-		if last.CreatedAt != nil && last.Id != nil {
-			next = encodeCursor(*last.CreatedAt, *last.Id)
-		}
-		items = items[:limit]
-	}
-	return items, next, nil
-}
-
-// UpdatePod applies merge-patch semantics on mutable fields.
-//
-//nolint:gocyclo // merge-patch nil checks are inherently repetitive
-func (p *PG) UpdatePod(ctx context.Context, id uuid.UUID, in api.PodUpdate) (api.Pod, error) {
-	sets := make([]string, 0, 5)
-	args := make([]any, 0, 7)
-	idx := 1
-	appendSet := func(column string, value any) {
-		sets = append(sets, fmt.Sprintf("%s=$%d", column, idx))
-		args = append(args, value)
-		idx++
-	}
-
-	if in.Phase != nil {
-		appendSet("phase", *in.Phase)
-	}
-	if in.NodeName != nil {
-		appendSet("node_name", *in.NodeName)
-	}
-	if in.PodIp != nil {
-		appendSet("pod_ip", *in.PodIp)
-	}
-	if in.WorkloadId != nil {
-		appendSet("workload_id", *in.WorkloadId)
-	}
-	if in.Containers != nil {
-		b, err := marshalPorts(in.Containers)
-		if err != nil {
-			return api.Pod{}, err
-		}
-		appendSet("containers", b)
-	}
-	if in.Labels != nil {
-		b, err := marshalLabels(in.Labels)
-		if err != nil {
-			return api.Pod{}, err
-		}
-		appendSet("labels", b)
-	}
-	appendSet("updated_at", time.Now().UTC())
-	args = append(args, id)
-
-	q := fmt.Sprintf("UPDATE pods SET %s WHERE id=$%d", strings.Join(sets, ", "), idx)
-	tag, err := p.pool.Exec(ctx, q, args...)
-	if err != nil {
-		if pErr := classifyPodFKError(err, uuid.Nil, in.WorkloadId); pErr != nil {
-			return api.Pod{}, pErr
-		}
-		return api.Pod{}, fmt.Errorf("update pod: %w", err)
-	}
-	if tag.RowsAffected() == 0 {
-		return api.Pod{}, api.ErrNotFound
-	}
-	return p.GetPod(ctx, id)
-}
-
-// DeletePod removes a pod by id.
-func (p *PG) DeletePod(ctx context.Context, id uuid.UUID) error {
-	tag, err := p.pool.Exec(ctx, "DELETE FROM pods WHERE id=$1", id)
-	if err != nil {
-		return fmt.Errorf("delete pod: %w", err)
-	}
-	if tag.RowsAffected() == 0 {
-		return api.ErrNotFound
-	}
-	return nil
-}
-
-// UpsertPod inserts-or-updates a pod keyed by (namespace_id, name).
-//
-//nolint:gocritic,gocyclo // hugeParam: Store interface requires value param; CTE drives branch count
-func (p *PG) UpsertPod(ctx context.Context, in api.PodCreate) (api.Pod, api.UpsertOutcome, error) {
-	id := uuid.New()
-	now := time.Now().UTC()
-
-	labelsJSON, err := marshalLabels(in.Labels)
-	if err != nil {
-		return api.Pod{}, api.OutcomeNoChange, err
-	}
-	containersJSON, err := marshalPorts(in.Containers)
-	if err != nil {
-		return api.Pod{}, api.OutcomeNoChange, err
-	}
-
-	// AUDIT_BUSINESS_FIELDS: phase, node_name, pod_ip, workload_id,
-	// containers, labels. updated_at is a clock field — excluded from
-	// business_changed detection so reconcile-only ticks turn into NoChange.
-	const q = `
-		WITH old AS (
-		  SELECT phase, node_name, pod_ip, workload_id, containers, labels
-		    FROM pods WHERE namespace_id = $2 AND name = $3
-		),
-		upserted AS (
-		  INSERT INTO pods (
-		      id, namespace_id, name, phase, node_name, pod_ip,
-		      workload_id, containers, labels, created_at, updated_at
-		  ) VALUES ($1, $2, $3, $4, $5, $6, $7, $8, $9, $10, $10)
-		  ON CONFLICT (namespace_id, name) DO UPDATE SET
-		      phase       = EXCLUDED.phase,
-		      node_name   = EXCLUDED.node_name,
-		      pod_ip      = EXCLUDED.pod_ip,
-		      workload_id = EXCLUDED.workload_id,
-		      containers  = EXCLUDED.containers,
-		      labels      = EXCLUDED.labels,
-		      updated_at  = EXCLUDED.updated_at
-		  RETURNING id, namespace_id, name, phase, node_name, pod_ip,
-		            workload_id, containers, labels, created_at, updated_at,
-		            xmax
-		)
-		SELECT u.id, u.namespace_id, u.name, u.phase, u.node_name, u.pod_ip,
-		       u.workload_id, u.containers, u.labels, u.created_at, u.updated_at,
-		       (u.xmax = 0) AS inserted,
-		       (u.xmax <> 0 AND (
-		           o.phase       IS DISTINCT FROM u.phase       OR
-		           o.node_name   IS DISTINCT FROM u.node_name   OR
-		           o.pod_ip      IS DISTINCT FROM u.pod_ip      OR
-		           o.workload_id IS DISTINCT FROM u.workload_id OR
-		           o.containers  IS DISTINCT FROM u.containers  OR
-		           o.labels      IS DISTINCT FROM u.labels
-		       )) AS business_changed
-		  FROM upserted u LEFT JOIN old o ON true
-	`
-	row := p.pool.QueryRow(ctx, q,
-		id, in.NamespaceId, in.Name, in.Phase, in.NodeName, in.PodIp,
-		in.WorkloadId, containersJSON, labelsJSON, now,
-	)
-
-	var (
-		pod             api.Pod
-		podID           uuid.UUID
-		namespaceID     uuid.UUID
-		createdAt       time.Time
-		updatedAt       time.Time
-		phase           sql.NullString
-		nodeName        sql.NullString
-		podIP           sql.NullString
-		workloadID      *uuid.UUID
-		containersOut   []byte
-		labelsOut       []byte
-		inserted        bool
-		businessChanged bool
-	)
-	if err := row.Scan(
-		&podID, &namespaceID, &pod.Name,
-		&phase, &nodeName, &podIP,
-		&workloadID, &containersOut, &labelsOut,
-		&createdAt, &updatedAt,
-		&inserted, &businessChanged,
-	); err != nil {
-		if pErr := classifyPodFKError(err, in.NamespaceId, in.WorkloadId); pErr != nil {
-			return api.Pod{}, api.OutcomeNoChange, pErr
-		}
-		return api.Pod{}, api.OutcomeNoChange, fmt.Errorf("upsert pod: %w", err)
-	}
-	pod.Id = &podID
-	pod.NamespaceId = namespaceID
-	pod.CreatedAt = &createdAt
-	pod.UpdatedAt = &updatedAt
-	pod.Phase = nullableString(phase)
-	pod.NodeName = nullableString(nodeName)
-	pod.PodIp = nullableString(podIP)
-	if workloadID != nil {
-		pod.WorkloadId = workloadID
-	}
-	if cs, err := unmarshalContainers(containersOut); err != nil {
-		return api.Pod{}, api.OutcomeNoChange, fmt.Errorf("unmarshal pod containers: %w", err)
-	} else if cs != nil {
-		pod.Containers = cs
-	}
-	if len(labelsOut) > 0 {
-		var labels map[string]string
-		if err := json.Unmarshal(labelsOut, &labels); err != nil {
-			return api.Pod{}, api.OutcomeNoChange, fmt.Errorf("unmarshal pod labels: %w", err)
-		}
-		if len(labels) > 0 {
-			pod.Labels = &labels
-		}
-	}
-	return pod, classifyOutcome(inserted, businessChanged), nil
-}
-
-// DeletePodsNotIn removes every pod in the given namespace whose name is not
-// in keepNames. Same COALESCE safety against pgx encoding a nil slice as NULL.
-func (p *PG) DeletePodsNotIn(ctx context.Context, namespaceID uuid.UUID, keepNames []string) (int64, error) {
-	tag, err := p.pool.Exec(ctx,
-		`DELETE FROM pods
-		 WHERE namespace_id = $1
-		   AND name <> ALL(COALESCE($2::text[], ARRAY[]::text[]))`,
-		namespaceID, keepNames,
-	)
-	if err != nil {
-		return 0, fmt.Errorf("delete pods not in: %w", err)
-	}
-	return tag.RowsAffected(), nil
-}
-
-// CreateWorkload inserts a new workload.
-//
-//nolint:gocritic // hugeParam: Store interface requires value param
-func (p *PG) CreateWorkload(ctx context.Context, in api.WorkloadCreate) (api.Workload, error) {
-	id := uuid.New()
-	now := time.Now().UTC()
-
-	labelsJSON, err := marshalLabels(in.Labels)
-	if err != nil {
-		return api.Workload{}, err
-	}
-	specJSON, err := marshalSpec(in.Spec)
-	if err != nil {
-		return api.Workload{}, err
-	}
-	containersJSON, err := marshalPorts(in.Containers)
-	if err != nil {
-		return api.Workload{}, err
-	}
-
-	const q = `
-		INSERT INTO workloads (
-			id, namespace_id, kind, name, replicas, ready_replicas,
-			containers, labels, spec, created_at, updated_at
-		) VALUES ($1, $2, $3, $4, $5, $6, $7, $8, $9, $10, $10)
-	`
-	_, err = p.pool.Exec(ctx, q,
-		id, in.NamespaceId, string(in.Kind), in.Name, in.Replicas, in.ReadyReplicas,
-		containersJSON, labelsJSON, specJSON, now,
-	)
-	if err != nil {
-		var pgErr *pgconn.PgError
-		if errors.As(err, &pgErr) {
-			switch pgErr.Code {
-			case "23505":
-				return api.Workload{}, fmt.Errorf(
-					"workload %s/%q in namespace %s already exists: %w",
-					in.Kind, in.Name, in.NamespaceId, api.ErrConflict,
-				)
-			case "23503":
-				return api.Workload{}, fmt.Errorf("namespace %s does not exist: %w", in.NamespaceId, api.ErrNotFound)
-			}
-		}
-		return api.Workload{}, fmt.Errorf("insert workload: %w", err)
-	}
-
-	return api.Workload{
-		Id:            &id,
-		NamespaceId:   in.NamespaceId,
-		Kind:          in.Kind,
-		Name:          in.Name,
-		Replicas:      in.Replicas,
-		ReadyReplicas: in.ReadyReplicas,
-		Containers:    in.Containers,
-		Labels:        in.Labels,
-		Spec:          in.Spec,
-		CreatedAt:     &now,
-		UpdatedAt:     &now,
-	}, nil
-}
-
-// Workload read-side projection with parent-name JOINs (ADR-0027) plus
-// the operator-curated application_id soft-pointer (ADR-0029).
-const workloadSelectColumns = `w.id, w.namespace_id, w.kind, w.name, w.replicas, w.ready_replicas,
-	w.containers, w.labels, w.spec, w.application_id, w.created_at, w.updated_at,
-	n.name AS namespace_name, n.cluster_id AS namespace_cluster_id, c.name AS cluster_name,
-	a.name AS application_name`
-
-const workloadFromJoined = `FROM workloads w
-	LEFT JOIN namespaces   n ON n.id = w.namespace_id
-	LEFT JOIN clusters     c ON c.id = n.cluster_id
-	LEFT JOIN applications a ON a.id = w.application_id`
-
-// GetWorkload fetches a workload by id.
-func (p *PG) GetWorkload(ctx context.Context, id uuid.UUID) (api.Workload, error) {
-	q := `SELECT ` + workloadSelectColumns + ` ` + workloadFromJoined + ` WHERE w.id = $1`
-	row := p.pool.QueryRow(ctx, q, id)
-	w, err := scanWorkload(row)
-	if err != nil {
-		if errors.Is(err, pgx.ErrNoRows) {
-			return api.Workload{}, api.ErrNotFound
-		}
-		return api.Workload{}, fmt.Errorf("select workload: %w", err)
-	}
-	return w, nil
-}
-
-// ListWorkloads returns up to limit workloads, optionally filtered by
-// namespace, kind, and/or container image substring.
-//
-//nolint:gocyclo // cursor-paginated query builder with optional filters
-func (p *PG) ListWorkloads(ctx context.Context, filter api.WorkloadListFilter, limit int, cursor string) ([]api.Workload, string, error) {
-	if limit <= 0 {
-		limit = 50
-	}
-	if limit > 200 {
-		limit = 200
-	}
-
-	sb := strings.Builder{}
-	sb.WriteString(`SELECT `)
-	sb.WriteString(workloadSelectColumns)
-	sb.WriteString(` `)
-	sb.WriteString(workloadFromJoined)
-	args := make([]any, 0, 6)
-	conds := make([]string, 0, 5)
-
-	if !filter.IncludeTerminated {
-		conds = append(conds, "w.terminated_at IS NULL")
-	}
-	if filter.NamespaceID != nil {
-		args = append(args, *filter.NamespaceID)
-		conds = append(conds, fmt.Sprintf("w.namespace_id = $%d", len(args)))
-	}
-	if filter.Kind != nil {
-		args = append(args, string(*filter.Kind))
-		conds = append(conds, fmt.Sprintf("w.kind = $%d", len(args)))
-	}
-	if filter.ImageSubstring != nil && *filter.ImageSubstring != "" {
-		args = append(args, "%"+*filter.ImageSubstring+"%")
-		conds = append(conds, fmt.Sprintf(
-			"EXISTS (SELECT 1 FROM jsonb_array_elements(w.containers) elem WHERE elem->>'image' ILIKE $%d)",
-			len(args),
-		))
-	}
-	// ADR-0029: link-aware filters. application_id wins on conflict with
-	// application_name (mirrors the cloud_account_id / cloud_account_name
-	// precedence from ADR-0019). application_name is normalised server-side
-	// to match the kebab-case stored in applications.name.
-	if filter.ApplicationID != nil {
-		args = append(args, *filter.ApplicationID)
-		conds = append(conds, fmt.Sprintf("w.application_id = $%d", len(args)))
-	} else if filter.ApplicationName != nil && *filter.ApplicationName != "" {
-		args = append(args, api.NormalizeApplicationName(*filter.ApplicationName))
-		conds = append(conds, fmt.Sprintf(
-			"w.application_id = (SELECT id FROM applications WHERE name = $%d)",
-			len(args),
-		))
-	}
-	if filter.Unlinked != nil && *filter.Unlinked {
-		conds = append(conds, "w.application_id IS NULL")
-	}
-	// ADR-0029 §2.4: substring match on linked application name (used by
-	// the Search endpoint). LIKE metacharacters are escaped; case-insensitive
-	// via LOWER on both sides.
-	if filter.ApplicationNameSubstring != nil && *filter.ApplicationNameSubstring != "" {
-		args = append(args, "%"+escapeLike(strings.ToLower(*filter.ApplicationNameSubstring))+"%")
-		conds = append(conds, fmt.Sprintf(
-			"w.application_id IN (SELECT id FROM applications WHERE LOWER(name) LIKE $%d ESCAPE '\\')",
-			len(args),
-		))
-	}
-	if cursor != "" {
-		ts, cid, err := decodeCursor(cursor)
-		if err != nil {
-			return nil, "", err
-		}
-		args = append(args, ts)
-		tsIdx := len(args)
-		args = append(args, cid)
-		idIdx := len(args)
-		conds = append(conds, fmt.Sprintf("(w.created_at, w.id) < ($%d, $%d)", tsIdx, idIdx))
-	}
-	if len(conds) > 0 {
-		sb.WriteString(" WHERE ")
-		sb.WriteString(strings.Join(conds, " AND "))
-	}
-	args = append(args, limit+1)
-	fmt.Fprintf(&sb, " ORDER BY created_at DESC, id DESC LIMIT $%d", len(args))
-
-	rows, err := p.pool.Query(ctx, sb.String(), args...)
-	if err != nil {
-		return nil, "", fmt.Errorf("query workloads: %w", err)
-	}
-	defer rows.Close()
-
-	items := make([]api.Workload, 0, limit)
-	for rows.Next() {
-		w, err := scanWorkload(rows)
-		if err != nil {
-			return nil, "", fmt.Errorf("scan workload: %w", err)
-		}
-		items = append(items, w)
-	}
-	if err := rows.Err(); err != nil {
-		return nil, "", fmt.Errorf("iterate workloads: %w", err)
-	}
-
-	var next string
-	if len(items) > limit {
-		last := items[limit-1]
-		if last.CreatedAt != nil && last.Id != nil {
-			next = encodeCursor(*last.CreatedAt, *last.Id)
-		}
-		items = items[:limit]
-	}
-	return items, next, nil
-}
-
-// UpdateWorkload applies merge-patch semantics on mutable fields.
-//
-//nolint:gocyclo // merge-patch nil checks are inherently repetitive
-func (p *PG) UpdateWorkload(ctx context.Context, id uuid.UUID, in api.WorkloadUpdate, clearApplication bool) (api.Workload, error) {
-	sets := make([]string, 0, 4)
-	args := make([]any, 0, 6)
-	idx := 1
-	appendSet := func(column string, value any) {
-		sets = append(sets, fmt.Sprintf("%s=$%d", column, idx))
-		args = append(args, value)
-		idx++
-	}
-
-	if in.Replicas != nil {
-		appendSet("replicas", *in.Replicas)
-	}
-	if in.ReadyReplicas != nil {
-		appendSet("ready_replicas", *in.ReadyReplicas)
-	}
-	if in.Containers != nil {
-		b, err := marshalPorts(in.Containers)
-		if err != nil {
-			return api.Workload{}, err
-		}
-		appendSet("containers", b)
-	}
-	if in.Labels != nil {
-		b, err := marshalLabels(in.Labels)
-		if err != nil {
-			return api.Workload{}, err
-		}
-		appendSet("labels", b)
-	}
-	if in.Spec != nil {
-		b, err := marshalSpec(in.Spec)
-		if err != nil {
-			return api.Workload{}, err
-		}
-		appendSet("spec", b)
-	}
-	// ADR-0029: operator-curated soft-pointer with three-state merge-patch
-	// semantics (RFC 7396). The codegen-generated field is *uuid.UUID, which
-	// can't distinguish absent from JSON null, so the handler detects an
-	// explicit `"application_id": null` from the raw body and threads it here
-	// via clearApplication:
-	//   - clearApplication=true  → SET application_id = NULL (unlink),
-	//   - in.ApplicationId != nil → SET application_id = <value> (link/relink),
-	//   - otherwise               → leave the column unchanged.
-	// id wins over null: an explicit value alongside null still links.
-	switch {
-	case in.ApplicationId != nil:
-		appendSet("application_id", *in.ApplicationId)
-	case clearApplication:
-		appendSet("application_id", nil)
-	}
-	appendSet("updated_at", time.Now().UTC())
-	args = append(args, id)
-
-	tx, txErr := p.pool.Begin(ctx)
-	if txErr != nil {
-		return api.Workload{}, fmt.Errorf("begin update workload: %w", txErr)
-	}
-	defer func() { _ = tx.Rollback(ctx) }()
-
-	prev, _ := workloadRowMap(ctx, tx, id) // FOR UPDATE
-
-	q := fmt.Sprintf("UPDATE workloads SET %s WHERE id=$%d", strings.Join(sets, ", "), idx)
-	tag, err := tx.Exec(ctx, q, args...)
-	if err != nil {
-		return api.Workload{}, fmt.Errorf("update workload: %w", err)
-	}
-	if tag.RowsAffected() == 0 {
-		return api.Workload{}, api.ErrNotFound
-	}
-
-	if prev != nil {
-		if next, err := workloadRowMapNoLock(ctx, tx, id); err == nil {
-			actor := timetravel.ActorFromContext(ctx)
-			_ = timetravel.Capture(ctx, tx, timetravel.KindWorkload, id, prev, next, changeTypeUpdate, actor)
-		}
-	}
-
-	if err := tx.Commit(ctx); err != nil {
-		return api.Workload{}, fmt.Errorf("commit update workload: %w", err)
-	}
-	return p.GetWorkload(ctx, id)
-}
-
-// DeleteWorkload removes a workload by id.
-func (p *PG) DeleteWorkload(ctx context.Context, id uuid.UUID) error {
-	tag, err := p.pool.Exec(ctx, "DELETE FROM workloads WHERE id=$1", id)
-	if err != nil {
-		return fmt.Errorf("delete workload: %w", err)
-	}
-	if tag.RowsAffected() == 0 {
-		return api.ErrNotFound
-	}
-	return nil
-}
-
-// UpsertWorkload inserts-or-updates a workload keyed by (namespace_id, kind, name).
-//
-//nolint:gocritic,gocyclo // hugeParam: Store interface requires value param; history capture adds branches
-func (p *PG) UpsertWorkload(ctx context.Context, in api.WorkloadCreate) (api.Workload, api.UpsertOutcome, error) {
-	id := uuid.New()
-	now := time.Now().UTC()
-
-	labelsJSON, err := marshalLabels(in.Labels)
-	if err != nil {
-		return api.Workload{}, api.OutcomeNoChange, err
-	}
-	specJSON, err := marshalSpec(in.Spec)
-	if err != nil {
-		return api.Workload{}, api.OutcomeNoChange, err
-	}
-	containersJSON, err := marshalPorts(in.Containers)
-	if err != nil {
-		return api.Workload{}, api.OutcomeNoChange, err
-	}
-
-	tx, err := p.pool.Begin(ctx)
-	if err != nil {
-		return api.Workload{}, api.OutcomeNoChange, fmt.Errorf("begin upsert workload: %w", err)
-	}
-	defer func() { _ = tx.Rollback(ctx) }()
-
-	var prevTerminatedAt *time.Time
-	var prevWLID *uuid.UUID
-	_ = tx.QueryRow(ctx,
-		`SELECT id, terminated_at FROM workloads WHERE namespace_id=$1 AND kind=$2 AND name=$3`,
-		in.NamespaceId, string(in.Kind), in.Name,
-	).Scan(&prevWLID, &prevTerminatedAt)
-	isCreate := prevWLID == nil
-	isRestore := prevWLID != nil && prevTerminatedAt != nil
-
-	// AUDIT_BUSINESS_FIELDS: replicas, ready_replicas, containers, labels, spec,
-	// terminated_at (restore flips it). updated_at is a clock field — excluded.
-	//
-	// COLLECTOR INVARIANT (ADR-0029 §7): application_id is operator-curated and
-	// MUST NOT be touched by collector ticks. The upsert deliberately omits it
-	// from both the INSERT column list and the ON CONFLICT SET list so that
-	// re-running reconcile preserves the curator's link. The PATCH path
-	// (UpdateWorkload) is the only place that writes this column.
-	const q = `
-		WITH old AS (
-		  SELECT replicas, ready_replicas, containers, labels, spec, terminated_at
-		    FROM workloads WHERE namespace_id=$2 AND kind=$3 AND name=$4
-		),
-		upserted AS (
-		  INSERT INTO workloads (
-		      id, namespace_id, kind, name, replicas, ready_replicas,
-		      containers, labels, spec, created_at, updated_at
-		  ) VALUES ($1, $2, $3, $4, $5, $6, $7, $8, $9, $10, $10)
-		  ON CONFLICT (namespace_id, kind, name) DO UPDATE SET
-		      replicas       = EXCLUDED.replicas,
-		      ready_replicas = EXCLUDED.ready_replicas,
-		      containers     = EXCLUDED.containers,
-		      labels         = EXCLUDED.labels,
-		      spec           = EXCLUDED.spec,
-		      terminated_at  = NULL,
-		      updated_at     = EXCLUDED.updated_at
-		  RETURNING id, namespace_id, kind, name, replicas, ready_replicas,
-		            containers, labels, spec, created_at, updated_at,
-		            terminated_at, xmax
-		)
-		SELECT u.id, u.namespace_id, u.kind, u.name, u.replicas, u.ready_replicas,
-		       u.containers, u.labels, u.spec, u.created_at, u.updated_at,
-		       (u.xmax = 0) AS inserted,
-		       (u.xmax <> 0 AND (
-		           o.replicas       IS DISTINCT FROM u.replicas       OR
-		           o.ready_replicas IS DISTINCT FROM u.ready_replicas OR
-		           o.containers     IS DISTINCT FROM u.containers     OR
-		           o.labels         IS DISTINCT FROM u.labels         OR
-		           o.spec           IS DISTINCT FROM u.spec           OR
-		           o.terminated_at  IS DISTINCT FROM u.terminated_at
-		       )) AS business_changed
-		  FROM upserted u LEFT JOIN old o ON true
-	`
-	row := tx.QueryRow(ctx, q,
-		id, in.NamespaceId, string(in.Kind), in.Name, in.Replicas, in.ReadyReplicas,
-		containersJSON, labelsJSON, specJSON, now,
-	)
-
-	var (
-		w               api.Workload
-		wID             uuid.UUID
-		namespaceID     uuid.UUID
-		kind            string
-		replicas        sql.NullInt32
-		readyReplicas   sql.NullInt32
-		createdAt       time.Time
-		updatedAt       time.Time
-		containersOut   []byte
-		labelsOut       []byte
-		specOut         []byte
-		inserted        bool
-		businessChanged bool
-	)
-	if err := row.Scan(
-		&wID, &namespaceID, &kind, &w.Name,
-		&replicas, &readyReplicas,
-		&containersOut, &labelsOut, &specOut,
-		&createdAt, &updatedAt,
-		&inserted, &businessChanged,
-	); err != nil {
-		var pgErr *pgconn.PgError
-		if errors.As(err, &pgErr) && pgErr.Code == "23503" {
-			return api.Workload{}, api.OutcomeNoChange, fmt.Errorf("namespace %s does not exist: %w", in.NamespaceId, api.ErrNotFound)
-		}
-		return api.Workload{}, api.OutcomeNoChange, fmt.Errorf("upsert workload: %w", err)
-	}
-	w.Id = &wID
-	w.NamespaceId = namespaceID
-	w.Kind = api.WorkloadKind(kind)
-	w.CreatedAt = &createdAt
-	w.UpdatedAt = &updatedAt
-	if replicas.Valid {
-		v := int(replicas.Int32)
-		w.Replicas = &v
-	}
-	if readyReplicas.Valid {
-		v := int(readyReplicas.Int32)
-		w.ReadyReplicas = &v
-	}
-	if cs, err := unmarshalContainers(containersOut); err != nil {
-		return api.Workload{}, api.OutcomeNoChange, fmt.Errorf("unmarshal workload containers: %w", err)
-	} else if cs != nil {
-		w.Containers = cs
-	}
-	if len(labelsOut) > 0 {
-		var labels map[string]string
-		if err := json.Unmarshal(labelsOut, &labels); err != nil {
-			return api.Workload{}, api.OutcomeNoChange, fmt.Errorf("unmarshal workload labels: %w", err)
-		}
-		if len(labels) > 0 {
-			w.Labels = &labels
-		}
-	}
-	if len(specOut) > 0 {
-		var spec map[string]interface{}
-		if err := json.Unmarshal(specOut, &spec); err != nil {
-			return api.Workload{}, api.OutcomeNoChange, fmt.Errorf("unmarshal workload spec: %w", err)
-		}
-		if len(spec) > 0 {
-			w.Spec = &spec
-		}
-	}
-
-	actualID := *w.Id
-	if snap, err := workloadRowMapNoLock(ctx, tx, actualID); err == nil {
-		actor := timetravel.ActorFromContext(ctx)
-		changeType := changeTypeUpdate
-		if isCreate {
-			changeType = changeTypeCreate
-		} else if isRestore {
-			changeType = changeTypeRestore
-		}
-		_ = timetravel.Capture(ctx, tx, timetravel.KindWorkload, actualID, nil, snap, changeType, actor)
-	}
-
-	if err := tx.Commit(ctx); err != nil {
-		return api.Workload{}, api.OutcomeNoChange, fmt.Errorf("commit upsert workload: %w", err)
-	}
-	return w, classifyOutcome(inserted, businessChanged), nil
-}
-
-// DeleteWorkloadsNotIn soft-deletes workloads in the namespace whose
-// (kind, name) tuple is not in the parallel keep arrays and that are not
-// already terminated. Per ADR-0021 §5; same semantics as DeleteNodesNotIn.
-// COALESCE guards against pgx encoding nil slices as SQL NULL.
-func (p *PG) DeleteWorkloadsNotIn(ctx context.Context, namespaceID uuid.UUID, keepKinds, keepNames []string) (int64, error) {
-	tx, err := p.pool.Begin(ctx)
-	if err != nil {
-		return 0, fmt.Errorf("begin delete workloads not in: %w", err)
-	}
-	defer func() { _ = tx.Rollback(ctx) }()
-
-	toDeleteRows, err := tx.Query(ctx,
-		`SELECT id FROM workloads
-		  WHERE namespace_id = $1
-		    AND (kind, name) NOT IN (
-		      SELECT k, n FROM UNNEST(
-		        COALESCE($2::text[], ARRAY[]::text[]),
-		        COALESCE($3::text[], ARRAY[]::text[])
-		      ) AS t(k, n)
-		    )
-		    AND terminated_at IS NULL`,
-		namespaceID, keepKinds, keepNames)
-	if err != nil {
-		return 0, fmt.Errorf("list workloads to soft-delete: %w", err)
-	}
-	toDelete, err := scanUUIDs(toDeleteRows)
-	if err != nil {
-		return 0, fmt.Errorf("scan workloads to soft-delete: %w", err)
-	}
-
-	tag, err := tx.Exec(ctx,
-		`UPDATE workloads
-		    SET terminated_at = NOW(), updated_at = NOW()
-		  WHERE namespace_id = $1
-		    AND (kind, name) NOT IN (
-		      SELECT k, n FROM UNNEST(
-		        COALESCE($2::text[], ARRAY[]::text[]),
-		        COALESCE($3::text[], ARRAY[]::text[])
-		      ) AS t(k, n)
-		    )
-		    AND terminated_at IS NULL`,
-		namespaceID, keepKinds, keepNames,
-	)
-	if err != nil {
-		return 0, fmt.Errorf("soft-delete workloads not in: %w", err)
-	}
-
-	actor := timetravel.ActorFromContext(ctx)
-	for _, wlID := range toDelete {
-		if snap, err := workloadRowMapNoLock(ctx, tx, wlID); err == nil {
-			_ = timetravel.Capture(ctx, tx, timetravel.KindWorkload, wlID, nil, snap, changeTypeSoftDelete, actor)
-		}
-	}
-
-	if err := tx.Commit(ctx); err != nil {
-		return 0, fmt.Errorf("commit delete workloads not in: %w", err)
-	}
-	return tag.RowsAffected(), nil
-}
-
-func marshalSpec(spec *map[string]interface{}) ([]byte, error) { //nolint:gocritic // ptrToRefParam: callers pass *map from generated API types
-	if spec == nil {
-		return []byte("{}"), nil
-	}
-	b, err := json.Marshal(*spec)
-	if err != nil {
-		return nil, fmt.Errorf("marshal spec: %w", err)
-	}
-	return b, nil
-}
-
-//nolint:gocyclo // JSONB unmarshal branches add inherent cyclomatic complexity
-func scanWorkload(row pgx.Row) (api.Workload, error) {
-	var (
-		w                  api.Workload
-		id                 uuid.UUID
-		namespaceID        uuid.UUID
-		kind               string
-		replicas           sql.NullInt32
-		readyReplicas      sql.NullInt32
-		createdAt          time.Time
-		updatedAt          time.Time
-		containersJSON     []byte
-		labelsJSON         []byte
-		specJSON           []byte
-		applicationID      uuid.NullUUID
-		namespaceName      sql.NullString
-		namespaceClusterID uuid.NullUUID
-		clusterName        sql.NullString
-		applicationName    sql.NullString
-	)
-	if err := row.Scan(
-		&id, &namespaceID, &kind, &w.Name,
-		&replicas, &readyReplicas,
-		&containersJSON, &labelsJSON, &specJSON, &applicationID,
-		&createdAt, &updatedAt,
-		&namespaceName, &namespaceClusterID, &clusterName,
-		&applicationName,
-	); err != nil {
-		return api.Workload{}, fmt.Errorf("scan workload: %w", err)
-	}
-	w.Id = &id
-	w.NamespaceId = namespaceID
-	w.Kind = api.WorkloadKind(kind)
-	w.CreatedAt = &createdAt
-	w.UpdatedAt = &updatedAt
-	w.NamespaceName = nullableString(namespaceName)
-	if namespaceClusterID.Valid {
-		v := namespaceClusterID.UUID
-		w.ClusterId = &v
-	}
-	w.ClusterName = nullableString(clusterName)
-	if applicationID.Valid {
-		v := applicationID.UUID
-		w.ApplicationId = &v
-	}
-	w.ApplicationName = nullableString(applicationName)
-	if replicas.Valid {
-		v := int(replicas.Int32)
-		w.Replicas = &v
-	}
-	if readyReplicas.Valid {
-		v := int(readyReplicas.Int32)
-		w.ReadyReplicas = &v
-	}
-	if cs, err := unmarshalContainers(containersJSON); err != nil {
-		return api.Workload{}, fmt.Errorf("unmarshal workload containers: %w", err)
-	} else if cs != nil {
-		w.Containers = cs
-	}
-	if len(labelsJSON) > 0 {
-		var labels map[string]string
-		if err := json.Unmarshal(labelsJSON, &labels); err != nil {
-			return api.Workload{}, fmt.Errorf("unmarshal workload labels: %w", err)
-		}
-		if len(labels) > 0 {
-			w.Labels = &labels
-		}
-	}
-	if len(specJSON) > 0 {
-		var spec map[string]interface{}
-		if err := json.Unmarshal(specJSON, &spec); err != nil {
-			return api.Workload{}, fmt.Errorf("unmarshal workload spec: %w", err)
-		}
-		if len(spec) > 0 {
-			w.Spec = &spec
-		}
-	}
-	return w, nil
-}
-
-// CreateService inserts a new service.
-// serviceColumns — same pattern as nodeColumns / ingressColumns.
-const serviceColumns = `id, namespace_id, name, type, cluster_ip,
-	selector, ports, load_balancer, labels, created_at, updated_at`
-
-// Read-side projection with parent-name JOINs (ADR-0027).
-const serviceSelectColumns = `s.id, s.namespace_id, s.name, s.type, s.cluster_ip,
-	s.selector, s.ports, s.load_balancer, s.labels, s.created_at, s.updated_at,
-	n.name AS namespace_name, n.cluster_id AS namespace_cluster_id, c.name AS cluster_name`
-
-const serviceFromJoined = `FROM services s
-	LEFT JOIN namespaces n ON n.id = s.namespace_id
-	LEFT JOIN clusters   c ON c.id = n.cluster_id`
-
-// CreateService inserts a new service into the given namespace.
-//
-//nolint:gocritic // hugeParam: Store interface requires value param
-func (p *PG) CreateService(ctx context.Context, in api.ServiceCreate) (api.Service, error) {
-	id := uuid.New()
-	now := time.Now().UTC()
-
-	labelsJSON, err := marshalLabels(in.Labels)
-	if err != nil {
-		return api.Service{}, err
-	}
-	selectorJSON, err := marshalLabels(in.Selector)
-	if err != nil {
-		return api.Service{}, err
-	}
-	portsJSON, err := marshalPorts(in.Ports)
-	if err != nil {
-		return api.Service{}, err
-	}
-	lbJSON, err := marshalPorts(in.LoadBalancer)
-	if err != nil {
-		return api.Service{}, err
-	}
-
-	var svcType *string
-	if in.Type != nil {
-		t := string(*in.Type)
-		svcType = &t
-	}
-
-	q := `INSERT INTO services (` + serviceColumns + `) VALUES ($1,$2,$3,$4,$5,$6,$7,$8,$9,$10,$10)`
-	_, err = p.pool.Exec(ctx, q,
-		id, in.NamespaceId, in.Name, svcType, in.ClusterIp,
-		selectorJSON, portsJSON, lbJSON, labelsJSON, now,
-	)
-	if err != nil {
-		var pgErr *pgconn.PgError
-		if errors.As(err, &pgErr) {
-			switch pgErr.Code {
-			case "23505":
-				return api.Service{}, fmt.Errorf("service %q in namespace %s already exists: %w", in.Name, in.NamespaceId, api.ErrConflict)
-			case "23503":
-				return api.Service{}, fmt.Errorf("namespace %s does not exist: %w", in.NamespaceId, api.ErrNotFound)
-			}
-		}
-		return api.Service{}, fmt.Errorf("insert service: %w", err)
-	}
-
-	return api.Service{
-		Id:           &id,
-		NamespaceId:  in.NamespaceId,
-		Name:         in.Name,
-		Type:         in.Type,
-		ClusterIp:    in.ClusterIp,
-		Selector:     in.Selector,
-		Ports:        in.Ports,
-		LoadBalancer: in.LoadBalancer,
-		Labels:       in.Labels,
-		CreatedAt:    &now,
-		UpdatedAt:    &now,
-	}, nil
-}
-
-// GetService fetches a service by id.
-func (p *PG) GetService(ctx context.Context, id uuid.UUID) (api.Service, error) {
-	q := `SELECT ` + serviceSelectColumns + ` ` + serviceFromJoined + ` WHERE s.id = $1`
-	row := p.pool.QueryRow(ctx, q, id)
-	s, err := scanService(row)
-	if err != nil {
-		if errors.Is(err, pgx.ErrNoRows) {
-			return api.Service{}, api.ErrNotFound
-		}
-		return api.Service{}, fmt.Errorf("select service: %w", err)
-	}
-	return s, nil
-}
-
-// ListServices returns up to limit services, optionally filtered by namespace.
-//
-//nolint:gocyclo // cursor-paginated query builder with optional filters
-func (p *PG) ListServices(ctx context.Context, namespaceID *uuid.UUID, limit int, cursor string) ([]api.Service, string, error) {
-	if limit <= 0 {
-		limit = 50
-	}
-	if limit > 200 {
-		limit = 200
-	}
-
-	sb := strings.Builder{}
-	sb.WriteString(`SELECT `)
-	sb.WriteString(serviceSelectColumns)
-	sb.WriteString(` `)
-	sb.WriteString(serviceFromJoined)
-	args := make([]any, 0, 4)
-	conds := make([]string, 0, 2)
-
-	if namespaceID != nil {
-		args = append(args, *namespaceID)
-		conds = append(conds, fmt.Sprintf("s.namespace_id = $%d", len(args)))
-	}
-	if cursor != "" {
-		ts, cid, err := decodeCursor(cursor)
-		if err != nil {
-			return nil, "", err
-		}
-		args = append(args, ts)
-		tsIdx := len(args)
-		args = append(args, cid)
-		idIdx := len(args)
-		conds = append(conds, fmt.Sprintf("(s.created_at, s.id) < ($%d, $%d)", tsIdx, idIdx))
-	}
-	if len(conds) > 0 {
-		sb.WriteString(" WHERE ")
-		sb.WriteString(strings.Join(conds, " AND "))
-	}
-	args = append(args, limit+1)
-	fmt.Fprintf(&sb, " ORDER BY s.created_at DESC, s.id DESC LIMIT $%d", len(args))
-
-	rows, err := p.pool.Query(ctx, sb.String(), args...)
-	if err != nil {
-		return nil, "", fmt.Errorf("query services: %w", err)
-	}
-	defer rows.Close()
-
-	items := make([]api.Service, 0, limit)
-	for rows.Next() {
-		s, err := scanService(rows)
-		if err != nil {
-			return nil, "", fmt.Errorf("scan service: %w", err)
-		}
-		items = append(items, s)
-	}
-	if err := rows.Err(); err != nil {
-		return nil, "", fmt.Errorf("iterate services: %w", err)
-	}
-
-	var next string
-	if len(items) > limit {
-		last := items[limit-1]
-		if last.CreatedAt != nil && last.Id != nil {
-			next = encodeCursor(*last.CreatedAt, *last.Id)
-		}
-		items = items[:limit]
-	}
-	return items, next, nil
-}
-
-// UpdateService applies merge-patch semantics on mutable fields.
-//
-//nolint:gocyclo // merge-patch nil checks are inherently repetitive
-func (p *PG) UpdateService(ctx context.Context, id uuid.UUID, in api.ServiceUpdate) (api.Service, error) {
-	sets := make([]string, 0, 5)
-	args := make([]any, 0, 7)
-	idx := 1
-	appendSet := func(column string, value any) {
-		sets = append(sets, fmt.Sprintf("%s=$%d", column, idx))
-		args = append(args, value)
-		idx++
-	}
-
-	if in.Type != nil {
-		appendSet("type", string(*in.Type))
-	}
-	if in.ClusterIp != nil {
-		appendSet("cluster_ip", *in.ClusterIp)
-	}
-	if in.Selector != nil {
-		b, err := marshalLabels(in.Selector)
-		if err != nil {
-			return api.Service{}, err
-		}
-		appendSet("selector", b)
-	}
-	if in.Ports != nil {
-		b, err := marshalPorts(in.Ports)
-		if err != nil {
-			return api.Service{}, err
-		}
-		appendSet("ports", b)
-	}
-	if in.LoadBalancer != nil {
-		b, err := marshalPorts(in.LoadBalancer)
-		if err != nil {
-			return api.Service{}, err
-		}
-		appendSet("load_balancer", b)
-	}
-	if in.Labels != nil {
-		b, err := marshalLabels(in.Labels)
-		if err != nil {
-			return api.Service{}, err
-		}
-		appendSet("labels", b)
-	}
-	appendSet("updated_at", time.Now().UTC())
-	args = append(args, id)
-
-	q := fmt.Sprintf("UPDATE services SET %s WHERE id=$%d", strings.Join(sets, ", "), idx)
-	tag, err := p.pool.Exec(ctx, q, args...)
-	if err != nil {
-		return api.Service{}, fmt.Errorf("update service: %w", err)
-	}
-	if tag.RowsAffected() == 0 {
-		return api.Service{}, api.ErrNotFound
-	}
-	return p.GetService(ctx, id)
-}
-
-// DeleteService removes a service by id.
-func (p *PG) DeleteService(ctx context.Context, id uuid.UUID) error {
-	tag, err := p.pool.Exec(ctx, "DELETE FROM services WHERE id=$1", id)
-	if err != nil {
-		return fmt.Errorf("delete service: %w", err)
-	}
-	if tag.RowsAffected() == 0 {
-		return api.ErrNotFound
-	}
-	return nil
-}
-
-// UpsertService inserts-or-updates a service keyed by (namespace_id, name).
-//
-//nolint:gocritic,gocyclo // hugeParam: Store interface requires value param; CTE drives branch count
-func (p *PG) UpsertService(ctx context.Context, in api.ServiceCreate) (api.Service, api.UpsertOutcome, error) {
-	id := uuid.New()
-	now := time.Now().UTC()
-
-	labelsJSON, err := marshalLabels(in.Labels)
-	if err != nil {
-		return api.Service{}, api.OutcomeNoChange, err
-	}
-	selectorJSON, err := marshalLabels(in.Selector)
-	if err != nil {
-		return api.Service{}, api.OutcomeNoChange, err
-	}
-	portsJSON, err := marshalPorts(in.Ports)
-	if err != nil {
-		return api.Service{}, api.OutcomeNoChange, err
-	}
-	lbJSON, err := marshalPorts(in.LoadBalancer)
-	if err != nil {
-		return api.Service{}, api.OutcomeNoChange, err
-	}
-
-	var svcType *string
-	if in.Type != nil {
-		t := string(*in.Type)
-		svcType = &t
-	}
-
-	// AUDIT_BUSINESS_FIELDS: type, cluster_ip, selector, ports, load_balancer,
-	// labels. updated_at is a clock field — excluded from business_changed
-	// detection so reconcile-only ticks turn into NoChange.
-	const q = `
-		WITH old AS (
-		  SELECT type, cluster_ip, selector, ports, load_balancer, labels
-		    FROM services WHERE namespace_id = $2 AND name = $3
-		),
-		upserted AS (
-		  INSERT INTO services (
-		      id, namespace_id, name, type, cluster_ip,
-		      selector, ports, load_balancer, labels, created_at, updated_at
-		  ) VALUES ($1,$2,$3,$4,$5,$6,$7,$8,$9,$10,$10)
-		  ON CONFLICT (namespace_id, name) DO UPDATE SET
-		      type          = EXCLUDED.type,
-		      cluster_ip    = EXCLUDED.cluster_ip,
-		      selector      = EXCLUDED.selector,
-		      ports         = EXCLUDED.ports,
-		      load_balancer = EXCLUDED.load_balancer,
-		      labels        = EXCLUDED.labels,
-		      updated_at    = EXCLUDED.updated_at
-		  RETURNING id, namespace_id, name, type, cluster_ip,
-		            selector, ports, load_balancer, labels, created_at, updated_at,
-		            xmax
-		)
-		SELECT u.id, u.namespace_id, u.name, u.type, u.cluster_ip,
-		       u.selector, u.ports, u.load_balancer, u.labels,
-		       u.created_at, u.updated_at,
-		       (u.xmax = 0) AS inserted,
-		       (u.xmax <> 0 AND (
-		           o.type          IS DISTINCT FROM u.type          OR
-		           o.cluster_ip    IS DISTINCT FROM u.cluster_ip    OR
-		           o.selector      IS DISTINCT FROM u.selector      OR
-		           o.ports         IS DISTINCT FROM u.ports         OR
-		           o.load_balancer IS DISTINCT FROM u.load_balancer OR
-		           o.labels        IS DISTINCT FROM u.labels
-		       )) AS business_changed
-		  FROM upserted u LEFT JOIN old o ON true
-	`
-	row := p.pool.QueryRow(ctx, q,
-		id, in.NamespaceId, in.Name, svcType, in.ClusterIp,
-		selectorJSON, portsJSON, lbJSON, labelsJSON, now,
-	)
-
-	var (
-		s               api.Service
-		svcID           uuid.UUID
-		namespaceID     uuid.UUID
-		createdAt       time.Time
-		updatedAt       time.Time
-		svcTypeOut      sql.NullString
-		clusterIP       sql.NullString
-		selectorOut     []byte
-		portsOut        []byte
-		lbOut           []byte
-		labelsOut       []byte
-		inserted        bool
-		businessChanged bool
-	)
-	if err := row.Scan(
-		&svcID, &namespaceID, &s.Name,
-		&svcTypeOut, &clusterIP,
-		&selectorOut, &portsOut, &lbOut, &labelsOut,
-		&createdAt, &updatedAt,
-		&inserted, &businessChanged,
-	); err != nil {
-		var pgErr *pgconn.PgError
-		if errors.As(err, &pgErr) && pgErr.Code == "23503" {
-			return api.Service{}, api.OutcomeNoChange, fmt.Errorf("namespace %s does not exist: %w", in.NamespaceId, api.ErrNotFound)
-		}
-		return api.Service{}, api.OutcomeNoChange, fmt.Errorf("upsert service: %w", err)
-	}
-	s.Id = &svcID
-	s.NamespaceId = namespaceID
-	s.CreatedAt = &createdAt
-	s.UpdatedAt = &updatedAt
-	s.ClusterIp = nullableString(clusterIP)
-	if svcTypeOut.Valid {
-		t := api.ServiceType(svcTypeOut.String)
-		s.Type = &t
-	}
-	if len(selectorOut) > 0 {
-		var sel map[string]string
-		if err := json.Unmarshal(selectorOut, &sel); err != nil {
-			return api.Service{}, api.OutcomeNoChange, fmt.Errorf("unmarshal service selector: %w", err)
-		}
-		if len(sel) > 0 {
-			s.Selector = &sel
-		}
-	}
-	if len(portsOut) > 0 {
-		var ports []map[string]interface{}
-		if err := json.Unmarshal(portsOut, &ports); err != nil {
-			return api.Service{}, api.OutcomeNoChange, fmt.Errorf("unmarshal service ports: %w", err)
-		}
-		if len(ports) > 0 {
-			s.Ports = &ports
-		}
-	}
-	if lb, err := unmarshalMapArray(lbOut); err != nil {
-		return api.Service{}, api.OutcomeNoChange, fmt.Errorf("unmarshal service load_balancer: %w", err)
-	} else {
-		s.LoadBalancer = lb
-	}
-	if len(labelsOut) > 0 {
-		var labels map[string]string
-		if err := json.Unmarshal(labelsOut, &labels); err != nil {
-			return api.Service{}, api.OutcomeNoChange, fmt.Errorf("unmarshal service labels: %w", err)
-		}
-		if len(labels) > 0 {
-			s.Labels = &labels
-		}
-	}
-	return s, classifyOutcome(inserted, businessChanged), nil
-}
-
-// DeleteServicesNotIn mirrors DeletePodsNotIn.
-func (p *PG) DeleteServicesNotIn(ctx context.Context, namespaceID uuid.UUID, keepNames []string) (int64, error) {
-	tag, err := p.pool.Exec(ctx,
-		`DELETE FROM services
-		 WHERE namespace_id = $1
-		   AND name <> ALL(COALESCE($2::text[], ARRAY[]::text[]))`,
-		namespaceID, keepNames,
-	)
-	if err != nil {
-		return 0, fmt.Errorf("delete services not in: %w", err)
-	}
-	return tag.RowsAffected(), nil
-}
-
 func marshalPorts(ports *[]map[string]interface{}) ([]byte, error) {
 	if ports == nil {
 		return []byte("[]"), nil
@@ -3087,592 +170,6 @@ func marshalPorts(ports *[]map[string]interface{}) ([]byte, error) {
 		return nil, fmt.Errorf("marshal ports: %w", err)
 	}
 	return b, nil
-}
-
-// CreateIngress inserts a new ingress.
-// ingressColumns is the full INSERT / SELECT / RETURNING column list for
-// the ingresses table. Kept as a single const so the several SQL paths
-// stay in sync; the load_balancer JSONB sits between tls and labels.
-const ingressColumns = `id, namespace_id, name, ingress_class_name,
-	rules, tls, load_balancer, labels, created_at, updated_at`
-
-// ingressSelectColumns is the read-side projection: every column from the
-// ingresses row (aliased to `i.`) plus the denormalized parent names
-// joined in via LEFT JOIN (ADR-0027). LEFT JOIN keeps orphan rows
-// visible with NULL names — the UI renders that as an explicit badge.
-const ingressSelectColumns = `i.id, i.namespace_id, i.name, i.ingress_class_name,
-	i.rules, i.tls, i.load_balancer, i.labels, i.created_at, i.updated_at,
-	n.name AS namespace_name, n.cluster_id AS namespace_cluster_id, c.name AS cluster_name`
-
-const ingressFromJoined = `FROM ingresses i
-	LEFT JOIN namespaces n ON n.id = i.namespace_id
-	LEFT JOIN clusters   c ON c.id = n.cluster_id`
-
-// CreateIngress inserts a new ingress into the given namespace.
-func (p *PG) CreateIngress(ctx context.Context, in api.IngressCreate) (api.Ingress, error) {
-	id := uuid.New()
-	now := time.Now().UTC()
-
-	labelsJSON, err := marshalLabels(in.Labels)
-	if err != nil {
-		return api.Ingress{}, err
-	}
-	rulesJSON, err := marshalPorts(in.Rules)
-	if err != nil {
-		return api.Ingress{}, err
-	}
-	tlsJSON, err := marshalPorts(in.Tls)
-	if err != nil {
-		return api.Ingress{}, err
-	}
-	lbJSON, err := marshalPorts(in.LoadBalancer)
-	if err != nil {
-		return api.Ingress{}, err
-	}
-
-	q := `INSERT INTO ingresses (` + ingressColumns + `) VALUES ($1,$2,$3,$4,$5,$6,$7,$8,$9,$9)`
-	_, err = p.pool.Exec(ctx, q,
-		id, in.NamespaceId, in.Name, in.IngressClassName,
-		rulesJSON, tlsJSON, lbJSON, labelsJSON, now,
-	)
-	if err != nil {
-		var pgErr *pgconn.PgError
-		if errors.As(err, &pgErr) {
-			switch pgErr.Code {
-			case "23505":
-				return api.Ingress{}, fmt.Errorf("ingress %q in namespace %s already exists: %w", in.Name, in.NamespaceId, api.ErrConflict)
-			case "23503":
-				return api.Ingress{}, fmt.Errorf("namespace %s does not exist: %w", in.NamespaceId, api.ErrNotFound)
-			}
-		}
-		return api.Ingress{}, fmt.Errorf("insert ingress: %w", err)
-	}
-
-	return api.Ingress{
-		Id:               &id,
-		NamespaceId:      in.NamespaceId,
-		Name:             in.Name,
-		IngressClassName: in.IngressClassName,
-		Rules:            in.Rules,
-		Tls:              in.Tls,
-		LoadBalancer:     in.LoadBalancer,
-		Labels:           in.Labels,
-		CreatedAt:        &now,
-		UpdatedAt:        &now,
-	}, nil
-}
-
-// GetIngress fetches an ingress by id.
-func (p *PG) GetIngress(ctx context.Context, id uuid.UUID) (api.Ingress, error) {
-	q := `SELECT ` + ingressSelectColumns + ` ` + ingressFromJoined + ` WHERE i.id = $1`
-	row := p.pool.QueryRow(ctx, q, id)
-	ing, err := scanIngress(row)
-	if err != nil {
-		if errors.Is(err, pgx.ErrNoRows) {
-			return api.Ingress{}, api.ErrNotFound
-		}
-		return api.Ingress{}, fmt.Errorf("select ingress: %w", err)
-	}
-	return ing, nil
-}
-
-// ListIngresses returns up to limit ingresses, optionally filtered by namespace.
-//
-//nolint:gocyclo // cursor-paginated query builder with optional filters
-func (p *PG) ListIngresses(ctx context.Context, namespaceID *uuid.UUID, limit int, cursor string) ([]api.Ingress, string, error) {
-	if limit <= 0 {
-		limit = 50
-	}
-	if limit > 200 {
-		limit = 200
-	}
-
-	sb := strings.Builder{}
-	sb.WriteString(`SELECT `)
-	sb.WriteString(ingressSelectColumns)
-	sb.WriteString(` `)
-	sb.WriteString(ingressFromJoined)
-	args := make([]any, 0, 4)
-	conds := make([]string, 0, 2)
-
-	if namespaceID != nil {
-		args = append(args, *namespaceID)
-		conds = append(conds, fmt.Sprintf("i.namespace_id = $%d", len(args)))
-	}
-	if cursor != "" {
-		ts, cid, err := decodeCursor(cursor)
-		if err != nil {
-			return nil, "", err
-		}
-		args = append(args, ts)
-		tsIdx := len(args)
-		args = append(args, cid)
-		idIdx := len(args)
-		conds = append(conds, fmt.Sprintf("(i.created_at, i.id) < ($%d, $%d)", tsIdx, idIdx))
-	}
-	if len(conds) > 0 {
-		sb.WriteString(" WHERE ")
-		sb.WriteString(strings.Join(conds, " AND "))
-	}
-	args = append(args, limit+1)
-	fmt.Fprintf(&sb, " ORDER BY i.created_at DESC, i.id DESC LIMIT $%d", len(args))
-
-	rows, err := p.pool.Query(ctx, sb.String(), args...)
-	if err != nil {
-		return nil, "", fmt.Errorf("query ingresses: %w", err)
-	}
-	defer rows.Close()
-
-	items := make([]api.Ingress, 0, limit)
-	for rows.Next() {
-		ing, err := scanIngress(rows)
-		if err != nil {
-			return nil, "", fmt.Errorf("scan ingress: %w", err)
-		}
-		items = append(items, ing)
-	}
-	if err := rows.Err(); err != nil {
-		return nil, "", fmt.Errorf("iterate ingresses: %w", err)
-	}
-
-	var next string
-	if len(items) > limit {
-		last := items[limit-1]
-		if last.CreatedAt != nil && last.Id != nil {
-			next = encodeCursor(*last.CreatedAt, *last.Id)
-		}
-		items = items[:limit]
-	}
-	return items, next, nil
-}
-
-// UpdateIngress applies merge-patch semantics on mutable fields.
-//
-//nolint:gocyclo // merge-patch nil checks are inherently repetitive
-func (p *PG) UpdateIngress(ctx context.Context, id uuid.UUID, in api.IngressUpdate) (api.Ingress, error) {
-	sets := make([]string, 0, 4)
-	args := make([]any, 0, 6)
-	idx := 1
-	appendSet := func(column string, value any) {
-		sets = append(sets, fmt.Sprintf("%s=$%d", column, idx))
-		args = append(args, value)
-		idx++
-	}
-
-	if in.IngressClassName != nil {
-		appendSet("ingress_class_name", *in.IngressClassName)
-	}
-	if in.Rules != nil {
-		b, err := marshalPorts(in.Rules)
-		if err != nil {
-			return api.Ingress{}, err
-		}
-		appendSet("rules", b)
-	}
-	if in.Tls != nil {
-		b, err := marshalPorts(in.Tls)
-		if err != nil {
-			return api.Ingress{}, err
-		}
-		appendSet("tls", b)
-	}
-	if in.LoadBalancer != nil {
-		b, err := marshalPorts(in.LoadBalancer)
-		if err != nil {
-			return api.Ingress{}, err
-		}
-		appendSet("load_balancer", b)
-	}
-	if in.Labels != nil {
-		b, err := marshalLabels(in.Labels)
-		if err != nil {
-			return api.Ingress{}, err
-		}
-		appendSet("labels", b)
-	}
-	appendSet("updated_at", time.Now().UTC())
-	args = append(args, id)
-
-	q := fmt.Sprintf("UPDATE ingresses SET %s WHERE id=$%d", strings.Join(sets, ", "), idx)
-	tag, err := p.pool.Exec(ctx, q, args...)
-	if err != nil {
-		return api.Ingress{}, fmt.Errorf("update ingress: %w", err)
-	}
-	if tag.RowsAffected() == 0 {
-		return api.Ingress{}, api.ErrNotFound
-	}
-	return p.GetIngress(ctx, id)
-}
-
-// DeleteIngress removes an ingress by id.
-func (p *PG) DeleteIngress(ctx context.Context, id uuid.UUID) error {
-	tag, err := p.pool.Exec(ctx, "DELETE FROM ingresses WHERE id=$1", id)
-	if err != nil {
-		return fmt.Errorf("delete ingress: %w", err)
-	}
-	if tag.RowsAffected() == 0 {
-		return api.ErrNotFound
-	}
-	return nil
-}
-
-// UpsertIngress inserts-or-updates an ingress keyed by (namespace_id, name).
-//
-//nolint:gocyclo // CTE drives branch count
-func (p *PG) UpsertIngress(ctx context.Context, in api.IngressCreate) (api.Ingress, api.UpsertOutcome, error) {
-	id := uuid.New()
-	now := time.Now().UTC()
-
-	labelsJSON, err := marshalLabels(in.Labels)
-	if err != nil {
-		return api.Ingress{}, api.OutcomeNoChange, err
-	}
-	rulesJSON, err := marshalPorts(in.Rules)
-	if err != nil {
-		return api.Ingress{}, api.OutcomeNoChange, err
-	}
-	tlsJSON, err := marshalPorts(in.Tls)
-	if err != nil {
-		return api.Ingress{}, api.OutcomeNoChange, err
-	}
-	lbJSON, err := marshalPorts(in.LoadBalancer)
-	if err != nil {
-		return api.Ingress{}, api.OutcomeNoChange, err
-	}
-
-	// AUDIT_BUSINESS_FIELDS: ingress_class_name, rules, tls, load_balancer,
-	// labels. updated_at is a clock field — excluded from business_changed.
-	const q = `
-		WITH old AS (
-		  SELECT ingress_class_name, rules, tls, load_balancer, labels
-		    FROM ingresses WHERE namespace_id = $2 AND name = $3
-		),
-		upserted AS (
-		  INSERT INTO ingresses (
-		      id, namespace_id, name, ingress_class_name,
-		      rules, tls, load_balancer, labels, created_at, updated_at
-		  ) VALUES ($1,$2,$3,$4,$5,$6,$7,$8,$9,$9)
-		  ON CONFLICT (namespace_id, name) DO UPDATE SET
-		      ingress_class_name = EXCLUDED.ingress_class_name,
-		      rules              = EXCLUDED.rules,
-		      tls                = EXCLUDED.tls,
-		      load_balancer      = EXCLUDED.load_balancer,
-		      labels             = EXCLUDED.labels,
-		      updated_at         = EXCLUDED.updated_at
-		  RETURNING id, namespace_id, name, ingress_class_name,
-		            rules, tls, load_balancer, labels, created_at, updated_at,
-		            xmax
-		)
-		SELECT u.id, u.namespace_id, u.name, u.ingress_class_name,
-		       u.rules, u.tls, u.load_balancer, u.labels,
-		       u.created_at, u.updated_at,
-		       (u.xmax = 0) AS inserted,
-		       (u.xmax <> 0 AND (
-		           o.ingress_class_name IS DISTINCT FROM u.ingress_class_name OR
-		           o.rules              IS DISTINCT FROM u.rules              OR
-		           o.tls                IS DISTINCT FROM u.tls                OR
-		           o.load_balancer      IS DISTINCT FROM u.load_balancer      OR
-		           o.labels             IS DISTINCT FROM u.labels
-		       )) AS business_changed
-		  FROM upserted u LEFT JOIN old o ON true
-	`
-
-	row := p.pool.QueryRow(ctx, q,
-		id, in.NamespaceId, in.Name, in.IngressClassName,
-		rulesJSON, tlsJSON, lbJSON, labelsJSON, now,
-	)
-
-	var (
-		i                api.Ingress
-		ingID            uuid.UUID
-		namespaceID      uuid.UUID
-		createdAt        time.Time
-		updatedAt        time.Time
-		ingressClassName sql.NullString
-		rulesOut         []byte
-		tlsOut           []byte
-		lbOut            []byte
-		labelsOut        []byte
-		inserted         bool
-		businessChanged  bool
-	)
-	if err := row.Scan(
-		&ingID, &namespaceID, &i.Name,
-		&ingressClassName,
-		&rulesOut, &tlsOut, &lbOut, &labelsOut,
-		&createdAt, &updatedAt,
-		&inserted, &businessChanged,
-	); err != nil {
-		var pgErr *pgconn.PgError
-		if errors.As(err, &pgErr) && pgErr.Code == "23503" {
-			return api.Ingress{}, api.OutcomeNoChange, fmt.Errorf("namespace %s does not exist: %w", in.NamespaceId, api.ErrNotFound)
-		}
-		return api.Ingress{}, api.OutcomeNoChange, fmt.Errorf("upsert ingress: %w", err)
-	}
-	i.Id = &ingID
-	i.NamespaceId = namespaceID
-	i.CreatedAt = &createdAt
-	i.UpdatedAt = &updatedAt
-	i.IngressClassName = nullableString(ingressClassName)
-	if len(rulesOut) > 0 {
-		var rules []map[string]interface{}
-		if err := json.Unmarshal(rulesOut, &rules); err != nil {
-			return api.Ingress{}, api.OutcomeNoChange, fmt.Errorf("unmarshal ingress rules: %w", err)
-		}
-		if len(rules) > 0 {
-			i.Rules = &rules
-		}
-	}
-	if len(tlsOut) > 0 {
-		var tls []map[string]interface{}
-		if err := json.Unmarshal(tlsOut, &tls); err != nil {
-			return api.Ingress{}, api.OutcomeNoChange, fmt.Errorf("unmarshal ingress tls: %w", err)
-		}
-		if len(tls) > 0 {
-			i.Tls = &tls
-		}
-	}
-	if lb, err := unmarshalMapArray(lbOut); err != nil {
-		return api.Ingress{}, api.OutcomeNoChange, fmt.Errorf("unmarshal ingress load_balancer: %w", err)
-	} else {
-		i.LoadBalancer = lb
-	}
-	if len(labelsOut) > 0 {
-		var labels map[string]string
-		if err := json.Unmarshal(labelsOut, &labels); err != nil {
-			return api.Ingress{}, api.OutcomeNoChange, fmt.Errorf("unmarshal ingress labels: %w", err)
-		}
-		if len(labels) > 0 {
-			i.Labels = &labels
-		}
-	}
-
-	return i, classifyOutcome(inserted, businessChanged), nil
-}
-
-// DeleteIngressesNotIn mirrors DeleteServicesNotIn.
-func (p *PG) DeleteIngressesNotIn(ctx context.Context, namespaceID uuid.UUID, keepNames []string) (int64, error) {
-	tag, err := p.pool.Exec(ctx,
-		`DELETE FROM ingresses
-		 WHERE namespace_id = $1
-		   AND name <> ALL(COALESCE($2::text[], ARRAY[]::text[]))`,
-		namespaceID, keepNames,
-	)
-	if err != nil {
-		return 0, fmt.Errorf("delete ingresses not in: %w", err)
-	}
-	return tag.RowsAffected(), nil
-}
-
-//nolint:gocyclo // JSONB unmarshal branches add inherent cyclomatic complexity
-func scanIngress(row pgx.Row) (api.Ingress, error) {
-	var (
-		i                  api.Ingress
-		id                 uuid.UUID
-		namespaceID        uuid.UUID
-		createdAt          time.Time
-		updatedAt          time.Time
-		ingressClassName   sql.NullString
-		rulesJSON          []byte
-		tlsJSON            []byte
-		lbJSON             []byte
-		labelsJSON         []byte
-		namespaceName      sql.NullString
-		namespaceClusterID uuid.NullUUID
-		clusterName        sql.NullString
-	)
-	if err := row.Scan(
-		&id, &namespaceID, &i.Name,
-		&ingressClassName,
-		&rulesJSON, &tlsJSON, &lbJSON, &labelsJSON,
-		&createdAt, &updatedAt,
-		&namespaceName, &namespaceClusterID, &clusterName,
-	); err != nil {
-		return api.Ingress{}, fmt.Errorf("scan ingress: %w", err)
-	}
-	i.Id = &id
-	i.NamespaceId = namespaceID
-	i.CreatedAt = &createdAt
-	i.UpdatedAt = &updatedAt
-	i.IngressClassName = nullableString(ingressClassName)
-	i.NamespaceName = nullableString(namespaceName)
-	if namespaceClusterID.Valid {
-		v := namespaceClusterID.UUID
-		i.ClusterId = &v
-	}
-	i.ClusterName = nullableString(clusterName)
-	if len(rulesJSON) > 0 {
-		var rules []map[string]interface{}
-		if err := json.Unmarshal(rulesJSON, &rules); err != nil {
-			return api.Ingress{}, fmt.Errorf("unmarshal ingress rules: %w", err)
-		}
-		if len(rules) > 0 {
-			i.Rules = &rules
-		}
-	}
-	if len(tlsJSON) > 0 {
-		var tls []map[string]interface{}
-		if err := json.Unmarshal(tlsJSON, &tls); err != nil {
-			return api.Ingress{}, fmt.Errorf("unmarshal ingress tls: %w", err)
-		}
-		if len(tls) > 0 {
-			i.Tls = &tls
-		}
-	}
-	if lb, err := unmarshalMapArray(lbJSON); err != nil {
-		return api.Ingress{}, fmt.Errorf("unmarshal ingress load_balancer: %w", err)
-	} else {
-		i.LoadBalancer = lb
-	}
-	if len(labelsJSON) > 0 {
-		var labels map[string]string
-		if err := json.Unmarshal(labelsJSON, &labels); err != nil {
-			return api.Ingress{}, fmt.Errorf("unmarshal ingress labels: %w", err)
-		}
-		if len(labels) > 0 {
-			i.Labels = &labels
-		}
-	}
-	return i, nil
-}
-
-//nolint:gocyclo // JSONB unmarshal branches add inherent cyclomatic complexity
-func scanService(row pgx.Row) (api.Service, error) {
-	var (
-		s                  api.Service
-		id                 uuid.UUID
-		namespaceID        uuid.UUID
-		createdAt          time.Time
-		updatedAt          time.Time
-		svcType            sql.NullString
-		clusterIP          sql.NullString
-		selectorJSON       []byte
-		portsJSON          []byte
-		lbJSON             []byte
-		labelsJSON         []byte
-		namespaceName      sql.NullString
-		namespaceClusterID uuid.NullUUID
-		clusterName        sql.NullString
-	)
-	if err := row.Scan(
-		&id, &namespaceID, &s.Name,
-		&svcType, &clusterIP,
-		&selectorJSON, &portsJSON, &lbJSON, &labelsJSON,
-		&createdAt, &updatedAt,
-		&namespaceName, &namespaceClusterID, &clusterName,
-	); err != nil {
-		return api.Service{}, fmt.Errorf("scan service: %w", err)
-	}
-	s.Id = &id
-	s.NamespaceId = namespaceID
-	s.CreatedAt = &createdAt
-	s.UpdatedAt = &updatedAt
-	s.ClusterIp = nullableString(clusterIP)
-	s.NamespaceName = nullableString(namespaceName)
-	if namespaceClusterID.Valid {
-		v := namespaceClusterID.UUID
-		s.ClusterId = &v
-	}
-	s.ClusterName = nullableString(clusterName)
-	if svcType.Valid {
-		t := api.ServiceType(svcType.String)
-		s.Type = &t
-	}
-	if len(selectorJSON) > 0 {
-		var sel map[string]string
-		if err := json.Unmarshal(selectorJSON, &sel); err != nil {
-			return api.Service{}, fmt.Errorf("unmarshal service selector: %w", err)
-		}
-		if len(sel) > 0 {
-			s.Selector = &sel
-		}
-	}
-	if len(portsJSON) > 0 {
-		var ports []map[string]interface{}
-		if err := json.Unmarshal(portsJSON, &ports); err != nil {
-			return api.Service{}, fmt.Errorf("unmarshal service ports: %w", err)
-		}
-		if len(ports) > 0 {
-			s.Ports = &ports
-		}
-	}
-	if lb, err := unmarshalMapArray(lbJSON); err != nil {
-		return api.Service{}, fmt.Errorf("unmarshal service load_balancer: %w", err)
-	} else {
-		s.LoadBalancer = lb
-	}
-	if len(labelsJSON) > 0 {
-		var labels map[string]string
-		if err := json.Unmarshal(labelsJSON, &labels); err != nil {
-			return api.Service{}, fmt.Errorf("unmarshal service labels: %w", err)
-		}
-		if len(labels) > 0 {
-			s.Labels = &labels
-		}
-	}
-	return s, nil
-}
-
-func scanPod(row pgx.Row) (api.Pod, error) {
-	var (
-		p                  api.Pod
-		id                 uuid.UUID
-		namespaceID        uuid.UUID
-		createdAt          time.Time
-		updatedAt          time.Time
-		phase              sql.NullString
-		nodeName           sql.NullString
-		podIP              sql.NullString
-		workloadID         *uuid.UUID
-		containersJSON     []byte
-		labelsJSON         []byte
-		namespaceName      sql.NullString
-		namespaceClusterID uuid.NullUUID
-		clusterName        sql.NullString
-		workloadName       sql.NullString
-	)
-	if err := row.Scan(
-		&id, &namespaceID, &p.Name,
-		&phase, &nodeName, &podIP,
-		&workloadID, &containersJSON, &labelsJSON,
-		&createdAt, &updatedAt,
-		&namespaceName, &namespaceClusterID, &clusterName,
-		&workloadName,
-	); err != nil {
-		return api.Pod{}, fmt.Errorf("scan pod: %w", err)
-	}
-	p.Id = &id
-	p.NamespaceId = namespaceID
-	p.CreatedAt = &createdAt
-	p.UpdatedAt = &updatedAt
-	p.Phase = nullableString(phase)
-	p.NodeName = nullableString(nodeName)
-	p.PodIp = nullableString(podIP)
-	if workloadID != nil {
-		p.WorkloadId = workloadID
-	}
-	p.NamespaceName = nullableString(namespaceName)
-	if namespaceClusterID.Valid {
-		v := namespaceClusterID.UUID
-		p.ClusterId = &v
-	}
-	p.ClusterName = nullableString(clusterName)
-	p.WorkloadName = nullableString(workloadName)
-	if cs, err := unmarshalContainers(containersJSON); err != nil {
-		return api.Pod{}, fmt.Errorf("unmarshal pod containers: %w", err)
-	} else if cs != nil {
-		p.Containers = cs
-	}
-	if len(labelsJSON) > 0 {
-		var labels map[string]string
-		if err := json.Unmarshal(labelsJSON, &labels); err != nil {
-			return api.Pod{}, fmt.Errorf("unmarshal pod labels: %w", err)
-		}
-		if len(labels) > 0 {
-			p.Labels = &labels
-		}
-	}
-	return p, nil
 }
 
 // unmarshalContainers decodes a JSONB array into the shared ContainerList
@@ -3691,353 +188,6 @@ func unmarshalContainers(b []byte) (*api.ContainerList, error) {
 		return nil, nil
 	}
 	return &cs, nil
-}
-
-func scanNamespace(row pgx.Row) (api.Namespace, error) {
-	var (
-		n               api.Namespace
-		id              uuid.UUID
-		clusterID       uuid.UUID
-		createdAt       time.Time
-		updatedAt       time.Time
-		displayName     sql.NullString
-		phase           sql.NullString
-		labelsJSON      []byte
-		owner           sql.NullString
-		criticality     sql.NullString
-		notes           sql.NullString
-		runbookURL      sql.NullString
-		annotationsJSON []byte
-		clusterName     sql.NullString
-	)
-	if err := row.Scan(
-		&id, &clusterID, &n.Name,
-		&displayName, &phase, &labelsJSON,
-		&owner, &criticality, &notes, &runbookURL, &annotationsJSON,
-		&createdAt, &updatedAt,
-		&clusterName,
-	); err != nil {
-		return api.Namespace{}, fmt.Errorf("scan namespace: %w", err)
-	}
-	n.Id = &id
-	n.ClusterId = clusterID
-	n.CreatedAt = &createdAt
-	n.UpdatedAt = &updatedAt
-	n.DisplayName = nullableString(displayName)
-	n.Phase = nullableString(phase)
-	n.Owner = nullableString(owner)
-	n.Criticality = nullableString(criticality)
-	n.Notes = nullableString(notes)
-	n.RunbookUrl = nullableString(runbookURL)
-	n.ClusterName = nullableString(clusterName)
-	if len(labelsJSON) > 0 {
-		var labels map[string]string
-		if err := json.Unmarshal(labelsJSON, &labels); err != nil {
-			return api.Namespace{}, fmt.Errorf("unmarshal namespace labels: %w", err)
-		}
-		if len(labels) > 0 {
-			n.Labels = &labels
-		}
-	}
-	if len(annotationsJSON) > 0 {
-		var annotations map[string]string
-		if err := json.Unmarshal(annotationsJSON, &annotations); err != nil {
-			return api.Namespace{}, fmt.Errorf("unmarshal namespace annotations: %w", err)
-		}
-		if len(annotations) > 0 {
-			n.Annotations = &annotations
-		}
-	}
-	return n, nil
-}
-
-// detectNodeUpsertChangeType inspects the existing node row (if any) keyed
-// by (clusterID, name) and returns the timetravel change type to record:
-// create when no row exists, restore when a soft-deleted row will be revived,
-// update otherwise. Errors from the lookup are swallowed — capture is best-
-// effort and the caller treats a missing row as "create".
-func detectNodeUpsertChangeType(ctx context.Context, tx pgx.Tx, clusterID uuid.UUID, name string) string {
-	var prevTerminatedAt *time.Time
-	var prevNodeID *uuid.UUID
-	_ = tx.QueryRow(ctx,
-		`SELECT id, terminated_at FROM nodes WHERE cluster_id=$1 AND name=$2`,
-		clusterID, name,
-	).Scan(&prevNodeID, &prevTerminatedAt)
-	switch {
-	case prevNodeID == nil:
-		return changeTypeCreate
-	case prevTerminatedAt != nil:
-		return changeTypeRestore
-	default:
-		return changeTypeUpdate
-	}
-}
-
-// UpsertNode inserts-or-updates a node keyed by (cluster_id, name). The
-// unique index on (cluster_id, name) drives the ON CONFLICT target. On
-// conflict only mutable columns are overwritten so created_at is preserved.
-//
-//nolint:gocritic // hugeParam: Store interface requires value param
-func (p *PG) UpsertNode(ctx context.Context, in api.NodeCreate) (api.Node, api.UpsertOutcome, error) {
-	id := uuid.New()
-	now := time.Now().UTC()
-
-	values, err := nodeInsertValues(&in, id, now)
-	if err != nil {
-		return api.Node{}, api.OutcomeNoChange, err
-	}
-
-	tx, err := p.pool.Begin(ctx)
-	if err != nil {
-		return api.Node{}, api.OutcomeNoChange, fmt.Errorf("begin upsert node: %w", err)
-	}
-	defer func() { _ = tx.Rollback(ctx) }()
-
-	changeType := detectNodeUpsertChangeType(ctx, tx, in.ClusterId, in.Name)
-
-	// AUDIT_BUSINESS_FIELDS: display_name, role, kubelet_version, kube_proxy_version,
-	// container_runtime_version, os_image, operating_system, kernel_version, architecture,
-	// internal_ip, external_ip, pod_cidr, provider_id, instance_type, zone,
-	// capacity_cpu, capacity_memory, capacity_pods, capacity_ephemeral_storage,
-	// allocatable_cpu, allocatable_memory, allocatable_pods, allocatable_ephemeral_storage,
-	// conditions, taints, unschedulable, ready, labels, terminated_at (restore flips it).
-	// updated_at is a clock field — excluded. Curator-only fields
-	// (owner/criticality/notes/runbook_url/annotations/hardware_model) are not touched
-	// by the collector upsert and excluded from the OR-chain.
-	const q = `
-		WITH old AS (
-		  SELECT display_name, role, kubelet_version, kube_proxy_version,
-		         container_runtime_version, os_image, operating_system, kernel_version,
-		         architecture, internal_ip, external_ip, pod_cidr,
-		         provider_id, instance_type, zone,
-		         capacity_cpu, capacity_memory, capacity_pods, capacity_ephemeral_storage,
-		         allocatable_cpu, allocatable_memory, allocatable_pods, allocatable_ephemeral_storage,
-		         conditions, taints, unschedulable, ready, labels, terminated_at
-		    FROM nodes WHERE cluster_id=$2 AND name=$3
-		),
-		upserted AS (
-		  INSERT INTO nodes (` + nodeColumns + `)
-		    VALUES ($1,$2,$3,$4,$5,$6,$7,$8,$9,$10,$11,$12,$13,$14,$15,$16,$17,$18,$19,$20,$21,$22,$23,$24,$25,$26,$27,$28,$29,$30,$31,$32,$33,$34,$35,$36,$37,$38,$38)
-		  ON CONFLICT (cluster_id, name) DO UPDATE SET
-		      display_name                  = EXCLUDED.display_name,
-		      role                          = EXCLUDED.role,
-		      kubelet_version               = EXCLUDED.kubelet_version,
-		      kube_proxy_version            = EXCLUDED.kube_proxy_version,
-		      container_runtime_version     = EXCLUDED.container_runtime_version,
-		      os_image                      = EXCLUDED.os_image,
-		      operating_system              = EXCLUDED.operating_system,
-		      kernel_version                = EXCLUDED.kernel_version,
-		      architecture                  = EXCLUDED.architecture,
-		      internal_ip                   = EXCLUDED.internal_ip,
-		      external_ip                   = EXCLUDED.external_ip,
-		      pod_cidr                      = EXCLUDED.pod_cidr,
-		      provider_id                   = EXCLUDED.provider_id,
-		      instance_type                 = EXCLUDED.instance_type,
-		      zone                          = EXCLUDED.zone,
-		      capacity_cpu                  = EXCLUDED.capacity_cpu,
-		      capacity_memory               = EXCLUDED.capacity_memory,
-		      capacity_pods                 = EXCLUDED.capacity_pods,
-		      capacity_ephemeral_storage    = EXCLUDED.capacity_ephemeral_storage,
-		      allocatable_cpu               = EXCLUDED.allocatable_cpu,
-		      allocatable_memory            = EXCLUDED.allocatable_memory,
-		      allocatable_pods              = EXCLUDED.allocatable_pods,
-		      allocatable_ephemeral_storage = EXCLUDED.allocatable_ephemeral_storage,
-		      conditions                    = EXCLUDED.conditions,
-		      taints                        = EXCLUDED.taints,
-		      unschedulable                 = EXCLUDED.unschedulable,
-		      ready                         = EXCLUDED.ready,
-		      labels                        = EXCLUDED.labels,
-		      terminated_at                 = NULL,
-		      updated_at                    = EXCLUDED.updated_at
-		  RETURNING ` + nodeColumns + `, terminated_at, xmax
-		)
-		SELECT ` + nodeColumnsUAliased + `,
-		       (u.xmax = 0) AS inserted,
-		       (u.xmax <> 0 AND (
-		           o.display_name                  IS DISTINCT FROM u.display_name                  OR
-		           o.role                          IS DISTINCT FROM u.role                          OR
-		           o.kubelet_version               IS DISTINCT FROM u.kubelet_version               OR
-		           o.kube_proxy_version            IS DISTINCT FROM u.kube_proxy_version            OR
-		           o.container_runtime_version     IS DISTINCT FROM u.container_runtime_version     OR
-		           o.os_image                      IS DISTINCT FROM u.os_image                      OR
-		           o.operating_system              IS DISTINCT FROM u.operating_system              OR
-		           o.kernel_version                IS DISTINCT FROM u.kernel_version                OR
-		           o.architecture                  IS DISTINCT FROM u.architecture                  OR
-		           o.internal_ip                   IS DISTINCT FROM u.internal_ip                   OR
-		           o.external_ip                   IS DISTINCT FROM u.external_ip                   OR
-		           o.pod_cidr                      IS DISTINCT FROM u.pod_cidr                      OR
-		           o.provider_id                   IS DISTINCT FROM u.provider_id                   OR
-		           o.instance_type                 IS DISTINCT FROM u.instance_type                 OR
-		           o.zone                          IS DISTINCT FROM u.zone                          OR
-		           o.capacity_cpu                  IS DISTINCT FROM u.capacity_cpu                  OR
-		           o.capacity_memory               IS DISTINCT FROM u.capacity_memory               OR
-		           o.capacity_pods                 IS DISTINCT FROM u.capacity_pods                 OR
-		           o.capacity_ephemeral_storage    IS DISTINCT FROM u.capacity_ephemeral_storage    OR
-		           o.allocatable_cpu               IS DISTINCT FROM u.allocatable_cpu               OR
-		           o.allocatable_memory            IS DISTINCT FROM u.allocatable_memory            OR
-		           o.allocatable_pods              IS DISTINCT FROM u.allocatable_pods              OR
-		           o.allocatable_ephemeral_storage IS DISTINCT FROM u.allocatable_ephemeral_storage OR
-		           o.conditions                    IS DISTINCT FROM u.conditions                    OR
-		           o.taints                        IS DISTINCT FROM u.taints                        OR
-		           o.unschedulable                 IS DISTINCT FROM u.unschedulable                 OR
-		           o.ready                         IS DISTINCT FROM u.ready                         OR
-		           o.labels                        IS DISTINCT FROM u.labels                        OR
-		           o.terminated_at                 IS DISTINCT FROM u.terminated_at
-		       )) AS business_changed
-		  FROM upserted u
-		  LEFT JOIN old o      ON true
-		  LEFT JOIN clusters c ON c.id = u.cluster_id
-	`
-	row := tx.QueryRow(ctx, q, values...)
-	var inserted, businessChanged bool
-	n, err := scanNode(scanRowWith{row: row, extra: []any{&inserted, &businessChanged}})
-	if err != nil {
-		var pgErr *pgconn.PgError
-		if errors.As(err, &pgErr) && pgErr.Code == "23503" {
-			return api.Node{}, api.OutcomeNoChange, fmt.Errorf("cluster %s does not exist: %w", in.ClusterId, api.ErrNotFound)
-		}
-		return api.Node{}, api.OutcomeNoChange, fmt.Errorf("upsert node: %w", err)
-	}
-
-	actualID := *n.Id
-	if snap, err := nodeRowMapNoLock(ctx, tx, actualID); err == nil {
-		actor := timetravel.ActorFromContext(ctx)
-		_ = timetravel.Capture(ctx, tx, timetravel.KindNode, actualID, nil, snap, changeType, actor)
-	}
-
-	if err := tx.Commit(ctx); err != nil {
-		return api.Node{}, api.OutcomeNoChange, fmt.Errorf("commit upsert node: %w", err)
-	}
-	return n, classifyOutcome(inserted, businessChanged), nil
-}
-
-func scanNode(row pgx.Row) (api.Node, error) {
-	var (
-		n                       api.Node
-		id                      uuid.UUID
-		clusterID               uuid.UUID
-		createdAt               time.Time
-		updatedAt               time.Time
-		displayName             sql.NullString
-		role                    sql.NullString
-		kubeletVersion          sql.NullString
-		kubeProxyVersion        sql.NullString
-		containerRuntimeVersion sql.NullString
-		osImage                 sql.NullString
-		operatingSystem         sql.NullString
-		kernelVersion           sql.NullString
-		architecture            sql.NullString
-		internalIP              sql.NullString
-		externalIP              sql.NullString
-		podCIDR                 sql.NullString
-		providerID              sql.NullString
-		instanceType            sql.NullString
-		zone                    sql.NullString
-		capacityCPU             sql.NullString
-		capacityMemory          sql.NullString
-		capacityPods            sql.NullString
-		capacityEphemeral       sql.NullString
-		allocatableCPU          sql.NullString
-		allocatableMemory       sql.NullString
-		allocatablePods         sql.NullString
-		allocatableEphemeral    sql.NullString
-		conditionsJSON          []byte
-		taintsJSON              []byte
-		unschedulable           bool
-		ready                   bool
-		labelsJSON              []byte
-		owner                   sql.NullString
-		criticality             sql.NullString
-		notes                   sql.NullString
-		runbookURL              sql.NullString
-		annotationsJSON         []byte
-		hardwareModel           sql.NullString
-		clusterName             sql.NullString
-	)
-	if err := row.Scan(
-		&id, &clusterID, &n.Name, &displayName, &role,
-		&kubeletVersion, &kubeProxyVersion, &containerRuntimeVersion,
-		&osImage, &operatingSystem, &kernelVersion, &architecture,
-		&internalIP, &externalIP, &podCIDR,
-		&providerID, &instanceType, &zone,
-		&capacityCPU, &capacityMemory, &capacityPods, &capacityEphemeral,
-		&allocatableCPU, &allocatableMemory, &allocatablePods, &allocatableEphemeral,
-		&conditionsJSON, &taintsJSON, &unschedulable, &ready,
-		&labelsJSON,
-		&owner, &criticality, &notes, &runbookURL, &annotationsJSON, &hardwareModel,
-		&createdAt, &updatedAt,
-		&clusterName,
-	); err != nil {
-		return api.Node{}, fmt.Errorf("scan node: %w", err)
-	}
-
-	n.Id = &id
-	n.ClusterId = clusterID
-	n.CreatedAt = &createdAt
-	n.UpdatedAt = &updatedAt
-	n.ClusterName = nullableString(clusterName)
-	n.DisplayName = nullableString(displayName)
-	n.Role = nullableString(role)
-	n.KubeletVersion = nullableString(kubeletVersion)
-	n.KubeProxyVersion = nullableString(kubeProxyVersion)
-	n.ContainerRuntimeVersion = nullableString(containerRuntimeVersion)
-	n.OsImage = nullableString(osImage)
-	n.OperatingSystem = nullableString(operatingSystem)
-	n.KernelVersion = nullableString(kernelVersion)
-	n.Architecture = nullableString(architecture)
-	n.InternalIp = nullableString(internalIP)
-	n.ExternalIp = nullableString(externalIP)
-	n.PodCidr = nullableString(podCIDR)
-	n.ProviderId = nullableString(providerID)
-	n.InstanceType = nullableString(instanceType)
-	n.Zone = nullableString(zone)
-	n.CapacityCpu = nullableString(capacityCPU)
-	n.CapacityMemory = nullableString(capacityMemory)
-	n.CapacityPods = nullableString(capacityPods)
-	n.CapacityEphemeralStorage = nullableString(capacityEphemeral)
-	n.AllocatableCpu = nullableString(allocatableCPU)
-	n.AllocatableMemory = nullableString(allocatableMemory)
-	n.AllocatablePods = nullableString(allocatablePods)
-	n.AllocatableEphemeralStorage = nullableString(allocatableEphemeral)
-	n.Unschedulable = &unschedulable
-	n.Ready = &ready
-
-	if cs, err := unmarshalMapArray(conditionsJSON); err != nil {
-		return api.Node{}, fmt.Errorf("unmarshal node conditions: %w", err)
-	} else {
-		n.Conditions = cs
-	}
-	if ts, err := unmarshalMapArray(taintsJSON); err != nil {
-		return api.Node{}, fmt.Errorf("unmarshal node taints: %w", err)
-	} else {
-		n.Taints = ts
-	}
-
-	if len(labelsJSON) > 0 {
-		var labels map[string]string
-		if err := json.Unmarshal(labelsJSON, &labels); err != nil {
-			return api.Node{}, fmt.Errorf("unmarshal node labels: %w", err)
-		}
-		if len(labels) > 0 {
-			n.Labels = &labels
-		}
-	}
-
-	n.Owner = nullableString(owner)
-	n.Criticality = nullableString(criticality)
-	n.Notes = nullableString(notes)
-	n.RunbookUrl = nullableString(runbookURL)
-	n.HardwareModel = nullableString(hardwareModel)
-	if len(annotationsJSON) > 0 {
-		var annotations map[string]string
-		if err := json.Unmarshal(annotationsJSON, &annotations); err != nil {
-			return api.Node{}, fmt.Errorf("unmarshal node annotations: %w", err)
-		}
-		if len(annotations) > 0 {
-			n.Annotations = &annotations
-		}
-	}
-	return n, nil
 }
 
 // unmarshalMapArray decodes a JSONB array of objects. Returns nil for
@@ -4069,71 +219,6 @@ func marshalLabels(labels *map[string]string) ([]byte, error) { //nolint:gocriti
 		return nil, fmt.Errorf("marshal labels: %w", err)
 	}
 	return b, nil
-}
-
-func scanCluster(row pgx.Row) (api.Cluster, error) {
-	var (
-		c                 api.Cluster
-		id                uuid.UUID
-		createdAt         time.Time
-		updatedAt         time.Time
-		displayName       sql.NullString
-		environment       sql.NullString
-		provider          sql.NullString
-		region            sql.NullString
-		kubernetesVersion sql.NullString
-		apiEndpoint       sql.NullString
-		labelsJSON        []byte
-		owner             sql.NullString
-		criticality       sql.NullString
-		notes             sql.NullString
-		runbookURL        sql.NullString
-		annotationsJSON   []byte
-	)
-	if err := row.Scan(
-		&id, &c.Name,
-		&displayName, &environment, &provider, &region,
-		&kubernetesVersion, &apiEndpoint,
-		&labelsJSON,
-		&owner, &criticality, &notes, &runbookURL, &annotationsJSON,
-		&createdAt, &updatedAt,
-	); err != nil {
-		return api.Cluster{}, fmt.Errorf("scan cluster: %w", err)
-	}
-
-	c.Id = &id
-	c.CreatedAt = &createdAt
-	c.UpdatedAt = &updatedAt
-	c.DisplayName = nullableString(displayName)
-	c.Environment = nullableString(environment)
-	c.Provider = nullableString(provider)
-	c.Region = nullableString(region)
-	c.KubernetesVersion = nullableString(kubernetesVersion)
-	c.ApiEndpoint = nullableString(apiEndpoint)
-	c.Owner = nullableString(owner)
-	c.Criticality = nullableString(criticality)
-	c.Notes = nullableString(notes)
-	c.RunbookUrl = nullableString(runbookURL)
-
-	if len(labelsJSON) > 0 {
-		var labels map[string]string
-		if err := json.Unmarshal(labelsJSON, &labels); err != nil {
-			return api.Cluster{}, fmt.Errorf("unmarshal labels: %w", err)
-		}
-		if len(labels) > 0 {
-			c.Labels = &labels
-		}
-	}
-	if len(annotationsJSON) > 0 {
-		var annotations map[string]string
-		if err := json.Unmarshal(annotationsJSON, &annotations); err != nil {
-			return api.Cluster{}, fmt.Errorf("unmarshal annotations: %w", err)
-		}
-		if len(annotations) > 0 {
-			c.Annotations = &annotations
-		}
-	}
-	return c, nil
 }
 
 func nullableString(s sql.NullString) *string {
@@ -4168,820 +253,6 @@ func decodeCursor(c string) (time.Time, uuid.UUID, error) {
 	return ts, id, nil
 }
 
-// accessModesValue normalises a nullable *[]string access-modes slice into a
-// never-nil []string that pgx will send as a TEXT[] (empty slice => empty
-// array, not NULL). Matches the store's DEFAULT '{}' column definition.
-func accessModesValue(modes *[]string) []string {
-	if modes == nil {
-		return []string{}
-	}
-	return *modes
-}
-
-func accessModesPointer(modes []string) *[]string {
-	if len(modes) == 0 {
-		return nil
-	}
-	return &modes
-}
-
-// CreatePersistentVolume inserts a cluster-scoped PV.
-//
-//nolint:gocritic // hugeParam: Store interface requires value param
-func (p *PG) CreatePersistentVolume(ctx context.Context, in api.PersistentVolumeCreate) (api.PersistentVolume, error) {
-	id := uuid.New()
-	now := time.Now().UTC()
-
-	labelsJSON, err := marshalLabels(in.Labels)
-	if err != nil {
-		return api.PersistentVolume{}, err
-	}
-
-	const q = `
-		INSERT INTO persistent_volumes (
-			id, cluster_id, name, capacity, access_modes,
-			reclaim_policy, phase, storage_class_name,
-			csi_driver, volume_handle,
-			claim_ref_namespace, claim_ref_name,
-			labels, created_at, updated_at
-		) VALUES ($1, $2, $3, $4, $5, $6, $7, $8, $9, $10, $11, $12, $13, $14, $14)
-	`
-	_, err = p.pool.Exec(ctx, q,
-		id, in.ClusterId, in.Name,
-		in.Capacity, accessModesValue(in.AccessModes),
-		in.ReclaimPolicy, in.Phase, in.StorageClassName,
-		in.CsiDriver, in.VolumeHandle,
-		in.ClaimRefNamespace, in.ClaimRefName,
-		labelsJSON, now,
-	)
-	if err != nil {
-		var pgErr *pgconn.PgError
-		if errors.As(err, &pgErr) {
-			switch pgErr.Code {
-			case "23505":
-				return api.PersistentVolume{}, fmt.Errorf(
-					"persistent volume %q in cluster %s already exists: %w",
-					in.Name, in.ClusterId, api.ErrConflict,
-				)
-			case "23503":
-				return api.PersistentVolume{}, fmt.Errorf("cluster %s does not exist: %w", in.ClusterId, api.ErrNotFound)
-			}
-		}
-		return api.PersistentVolume{}, fmt.Errorf("insert persistent volume: %w", err)
-	}
-
-	return api.PersistentVolume{
-		Id:                &id,
-		ClusterId:         in.ClusterId,
-		Name:              in.Name,
-		Capacity:          in.Capacity,
-		AccessModes:       in.AccessModes,
-		ReclaimPolicy:     in.ReclaimPolicy,
-		Phase:             in.Phase,
-		StorageClassName:  in.StorageClassName,
-		CsiDriver:         in.CsiDriver,
-		VolumeHandle:      in.VolumeHandle,
-		ClaimRefNamespace: in.ClaimRefNamespace,
-		ClaimRefName:      in.ClaimRefName,
-		Labels:            in.Labels,
-		CreatedAt:         &now,
-		UpdatedAt:         &now,
-	}, nil
-}
-
-// GetPersistentVolume fetches a PV by id.
-func (p *PG) GetPersistentVolume(ctx context.Context, id uuid.UUID) (api.PersistentVolume, error) {
-	const q = `
-		SELECT id, cluster_id, name, capacity, access_modes,
-		       reclaim_policy, phase, storage_class_name,
-		       csi_driver, volume_handle,
-		       claim_ref_namespace, claim_ref_name,
-		       labels, created_at, updated_at
-		FROM persistent_volumes
-		WHERE id = $1
-	`
-	row := p.pool.QueryRow(ctx, q, id)
-	pv, err := scanPersistentVolume(row)
-	if err != nil {
-		if errors.Is(err, pgx.ErrNoRows) {
-			return api.PersistentVolume{}, api.ErrNotFound
-		}
-		return api.PersistentVolume{}, fmt.Errorf("select persistent volume: %w", err)
-	}
-	return pv, nil
-}
-
-// ListPersistentVolumes returns up to limit PVs, optionally filtered by cluster.
-//
-//nolint:gocyclo // cursor-paginated query builder with optional filters
-func (p *PG) ListPersistentVolumes(ctx context.Context, clusterID *uuid.UUID, limit int, cursor string) ([]api.PersistentVolume, string, error) {
-	if limit <= 0 {
-		limit = 50
-	}
-	if limit > 200 {
-		limit = 200
-	}
-
-	sb := strings.Builder{}
-	sb.WriteString(`SELECT id, cluster_id, name, capacity, access_modes,
-	                       reclaim_policy, phase, storage_class_name,
-	                       csi_driver, volume_handle,
-	                       claim_ref_namespace, claim_ref_name,
-	                       labels, created_at, updated_at
-	                FROM persistent_volumes`)
-	args := make([]any, 0, 4)
-	conds := make([]string, 0, 2)
-
-	if clusterID != nil {
-		args = append(args, *clusterID)
-		conds = append(conds, fmt.Sprintf("cluster_id = $%d", len(args)))
-	}
-	if cursor != "" {
-		ts, cid, err := decodeCursor(cursor)
-		if err != nil {
-			return nil, "", err
-		}
-		args = append(args, ts)
-		tsIdx := len(args)
-		args = append(args, cid)
-		idIdx := len(args)
-		conds = append(conds, fmt.Sprintf("(created_at, id) < ($%d, $%d)", tsIdx, idIdx))
-	}
-	if len(conds) > 0 {
-		sb.WriteString(" WHERE ")
-		sb.WriteString(strings.Join(conds, " AND "))
-	}
-	args = append(args, limit+1)
-	fmt.Fprintf(&sb, " ORDER BY created_at DESC, id DESC LIMIT $%d", len(args))
-
-	rows, err := p.pool.Query(ctx, sb.String(), args...)
-	if err != nil {
-		return nil, "", fmt.Errorf("query persistent volumes: %w", err)
-	}
-	defer rows.Close()
-
-	items := make([]api.PersistentVolume, 0, limit)
-	for rows.Next() {
-		pv, err := scanPersistentVolume(rows)
-		if err != nil {
-			return nil, "", fmt.Errorf("scan persistent volume: %w", err)
-		}
-		items = append(items, pv)
-	}
-	if err := rows.Err(); err != nil {
-		return nil, "", fmt.Errorf("iterate persistent volumes: %w", err)
-	}
-
-	var next string
-	if len(items) > limit {
-		last := items[limit-1]
-		if last.CreatedAt != nil && last.Id != nil {
-			next = encodeCursor(*last.CreatedAt, *last.Id)
-		}
-		items = items[:limit]
-	}
-	return items, next, nil
-}
-
-// UpdatePersistentVolume applies merge-patch on mutable PV fields.
-//
-//nolint:gocyclo,gocritic // merge-patch nil checks are inherently repetitive; hugeParam: Store interface requires value param
-func (p *PG) UpdatePersistentVolume(ctx context.Context, id uuid.UUID, in api.PersistentVolumeUpdate) (api.PersistentVolume, error) {
-	sets := make([]string, 0, 8)
-	args := make([]any, 0, 10)
-	idx := 1
-	appendSet := func(column string, value any) {
-		sets = append(sets, fmt.Sprintf("%s=$%d", column, idx))
-		args = append(args, value)
-		idx++
-	}
-
-	if in.Capacity != nil {
-		appendSet("capacity", *in.Capacity)
-	}
-	if in.AccessModes != nil {
-		appendSet("access_modes", accessModesValue(in.AccessModes))
-	}
-	if in.ReclaimPolicy != nil {
-		appendSet("reclaim_policy", *in.ReclaimPolicy)
-	}
-	if in.Phase != nil {
-		appendSet("phase", *in.Phase)
-	}
-	if in.StorageClassName != nil {
-		appendSet("storage_class_name", *in.StorageClassName)
-	}
-	if in.CsiDriver != nil {
-		appendSet("csi_driver", *in.CsiDriver)
-	}
-	if in.VolumeHandle != nil {
-		appendSet("volume_handle", *in.VolumeHandle)
-	}
-	if in.ClaimRefNamespace != nil {
-		appendSet("claim_ref_namespace", *in.ClaimRefNamespace)
-	}
-	if in.ClaimRefName != nil {
-		appendSet("claim_ref_name", *in.ClaimRefName)
-	}
-	if in.Labels != nil {
-		b, err := marshalLabels(in.Labels)
-		if err != nil {
-			return api.PersistentVolume{}, err
-		}
-		appendSet("labels", b)
-	}
-	appendSet("updated_at", time.Now().UTC())
-	args = append(args, id)
-
-	q := fmt.Sprintf("UPDATE persistent_volumes SET %s WHERE id=$%d", strings.Join(sets, ", "), idx)
-	tag, err := p.pool.Exec(ctx, q, args...)
-	if err != nil {
-		return api.PersistentVolume{}, fmt.Errorf("update persistent volume: %w", err)
-	}
-	if tag.RowsAffected() == 0 {
-		return api.PersistentVolume{}, api.ErrNotFound
-	}
-	return p.GetPersistentVolume(ctx, id)
-}
-
-// DeletePersistentVolume removes a PV by id.
-func (p *PG) DeletePersistentVolume(ctx context.Context, id uuid.UUID) error {
-	tag, err := p.pool.Exec(ctx, "DELETE FROM persistent_volumes WHERE id=$1", id)
-	if err != nil {
-		return fmt.Errorf("delete persistent volume: %w", err)
-	}
-	if tag.RowsAffected() == 0 {
-		return api.ErrNotFound
-	}
-	return nil
-}
-
-// UpsertPersistentVolume inserts-or-updates a PV keyed by (cluster_id, name).
-//
-//nolint:gocritic // hugeParam: Store interface requires value param
-func (p *PG) UpsertPersistentVolume(ctx context.Context, in api.PersistentVolumeCreate) (api.PersistentVolume, api.UpsertOutcome, error) {
-	id := uuid.New()
-	now := time.Now().UTC()
-
-	labelsJSON, err := marshalLabels(in.Labels)
-	if err != nil {
-		return api.PersistentVolume{}, api.OutcomeNoChange, err
-	}
-
-	// AUDIT_BUSINESS_FIELDS: capacity, access_modes, reclaim_policy, phase,
-	// storage_class_name, csi_driver, volume_handle, claim_ref_namespace,
-	// claim_ref_name, labels. updated_at is a clock field — excluded.
-	const q = `
-		WITH old AS (
-		  SELECT capacity, access_modes, reclaim_policy, phase,
-		         storage_class_name, csi_driver, volume_handle,
-		         claim_ref_namespace, claim_ref_name, labels
-		    FROM persistent_volumes WHERE cluster_id=$2 AND name=$3
-		),
-		upserted AS (
-		  INSERT INTO persistent_volumes (
-		    id, cluster_id, name, capacity, access_modes,
-		    reclaim_policy, phase, storage_class_name,
-		    csi_driver, volume_handle,
-		    claim_ref_namespace, claim_ref_name,
-		    labels, created_at, updated_at
-		  ) VALUES ($1, $2, $3, $4, $5, $6, $7, $8, $9, $10, $11, $12, $13, $14, $14)
-		  ON CONFLICT (cluster_id, name) DO UPDATE SET
-		      capacity            = EXCLUDED.capacity,
-		      access_modes        = EXCLUDED.access_modes,
-		      reclaim_policy      = EXCLUDED.reclaim_policy,
-		      phase               = EXCLUDED.phase,
-		      storage_class_name  = EXCLUDED.storage_class_name,
-		      csi_driver          = EXCLUDED.csi_driver,
-		      volume_handle       = EXCLUDED.volume_handle,
-		      claim_ref_namespace = EXCLUDED.claim_ref_namespace,
-		      claim_ref_name      = EXCLUDED.claim_ref_name,
-		      labels              = EXCLUDED.labels,
-		      updated_at          = EXCLUDED.updated_at
-		  RETURNING id, cluster_id, name, capacity, access_modes,
-		            reclaim_policy, phase, storage_class_name,
-		            csi_driver, volume_handle,
-		            claim_ref_namespace, claim_ref_name,
-		            labels, created_at, updated_at, xmax
-		)
-		SELECT u.id, u.cluster_id, u.name, u.capacity, u.access_modes,
-		       u.reclaim_policy, u.phase, u.storage_class_name,
-		       u.csi_driver, u.volume_handle,
-		       u.claim_ref_namespace, u.claim_ref_name,
-		       u.labels, u.created_at, u.updated_at,
-		       (u.xmax = 0) AS inserted,
-		       (u.xmax <> 0 AND (
-		           o.capacity            IS DISTINCT FROM u.capacity            OR
-		           o.access_modes        IS DISTINCT FROM u.access_modes        OR
-		           o.reclaim_policy      IS DISTINCT FROM u.reclaim_policy      OR
-		           o.phase               IS DISTINCT FROM u.phase               OR
-		           o.storage_class_name  IS DISTINCT FROM u.storage_class_name  OR
-		           o.csi_driver          IS DISTINCT FROM u.csi_driver          OR
-		           o.volume_handle       IS DISTINCT FROM u.volume_handle       OR
-		           o.claim_ref_namespace IS DISTINCT FROM u.claim_ref_namespace OR
-		           o.claim_ref_name      IS DISTINCT FROM u.claim_ref_name      OR
-		           o.labels              IS DISTINCT FROM u.labels
-		       )) AS business_changed
-		  FROM upserted u LEFT JOIN old o ON true
-	`
-	row := p.pool.QueryRow(ctx, q,
-		id, in.ClusterId, in.Name,
-		in.Capacity, accessModesValue(in.AccessModes),
-		in.ReclaimPolicy, in.Phase, in.StorageClassName,
-		in.CsiDriver, in.VolumeHandle,
-		in.ClaimRefNamespace, in.ClaimRefName,
-		labelsJSON, now,
-	)
-	var inserted, businessChanged bool
-	pv, err := scanPersistentVolume(scanRowWith{row: row, extra: []any{&inserted, &businessChanged}})
-	if err != nil {
-		var pgErr *pgconn.PgError
-		if errors.As(err, &pgErr) && pgErr.Code == "23503" {
-			return api.PersistentVolume{}, api.OutcomeNoChange, fmt.Errorf("cluster %s does not exist: %w", in.ClusterId, api.ErrNotFound)
-		}
-		return api.PersistentVolume{}, api.OutcomeNoChange, fmt.Errorf("upsert persistent volume: %w", err)
-	}
-	return pv, classifyOutcome(inserted, businessChanged), nil
-}
-
-// DeletePersistentVolumesNotIn removes cluster-scoped PVs whose name is not in keepNames.
-func (p *PG) DeletePersistentVolumesNotIn(ctx context.Context, clusterID uuid.UUID, keepNames []string) (int64, error) {
-	tag, err := p.pool.Exec(ctx,
-		`DELETE FROM persistent_volumes
-		 WHERE cluster_id = $1
-		   AND name <> ALL(COALESCE($2::text[], ARRAY[]::text[]))`,
-		clusterID, keepNames,
-	)
-	if err != nil {
-		return 0, fmt.Errorf("delete persistent volumes not in: %w", err)
-	}
-	return tag.RowsAffected(), nil
-}
-
-func scanPersistentVolume(row pgx.Row) (api.PersistentVolume, error) {
-	var (
-		pv                api.PersistentVolume
-		id                uuid.UUID
-		clusterID         uuid.UUID
-		createdAt         time.Time
-		updatedAt         time.Time
-		capacity          sql.NullString
-		accessModes       []string
-		reclaimPolicy     sql.NullString
-		phase             sql.NullString
-		storageClassName  sql.NullString
-		csiDriver         sql.NullString
-		volumeHandle      sql.NullString
-		claimRefNamespace sql.NullString
-		claimRefName      sql.NullString
-		labelsJSON        []byte
-	)
-	if err := row.Scan(
-		&id, &clusterID, &pv.Name,
-		&capacity, &accessModes,
-		&reclaimPolicy, &phase, &storageClassName,
-		&csiDriver, &volumeHandle,
-		&claimRefNamespace, &claimRefName,
-		&labelsJSON,
-		&createdAt, &updatedAt,
-	); err != nil {
-		return api.PersistentVolume{}, fmt.Errorf("scan persistent volume: %w", err)
-	}
-	pv.Id = &id
-	pv.ClusterId = clusterID
-	pv.CreatedAt = &createdAt
-	pv.UpdatedAt = &updatedAt
-	pv.Capacity = nullableString(capacity)
-	pv.AccessModes = accessModesPointer(accessModes)
-	pv.ReclaimPolicy = nullableString(reclaimPolicy)
-	pv.Phase = nullableString(phase)
-	pv.StorageClassName = nullableString(storageClassName)
-	pv.CsiDriver = nullableString(csiDriver)
-	pv.VolumeHandle = nullableString(volumeHandle)
-	pv.ClaimRefNamespace = nullableString(claimRefNamespace)
-	pv.ClaimRefName = nullableString(claimRefName)
-	if len(labelsJSON) > 0 {
-		var labels map[string]string
-		if err := json.Unmarshal(labelsJSON, &labels); err != nil {
-			return api.PersistentVolume{}, fmt.Errorf("unmarshal pv labels: %w", err)
-		}
-		if len(labels) > 0 {
-			pv.Labels = &labels
-		}
-	}
-	return pv, nil
-}
-
-// classifyPVCFKError disambiguates 23503 foreign-key violations on the
-// persistent_volume_claims table into namespace vs bound-volume misses.
-func classifyPVCFKError(err error, namespaceID uuid.UUID, boundVolumeID *uuid.UUID) error {
-	var pgErr *pgconn.PgError
-	if !errors.As(err, &pgErr) || pgErr.Code != "23503" {
-		return nil
-	}
-	if strings.Contains(pgErr.ConstraintName, "bound_volume_id") {
-		target := "<nil>"
-		if boundVolumeID != nil {
-			target = boundVolumeID.String()
-		}
-		return fmt.Errorf("persistent volume %s does not exist: %w", target, api.ErrNotFound)
-	}
-	return fmt.Errorf("namespace %s does not exist: %w", namespaceID, api.ErrNotFound)
-}
-
-// CreatePersistentVolumeClaim inserts a namespace-scoped PVC.
-//
-//nolint:gocritic // hugeParam: Store interface requires value param
-func (p *PG) CreatePersistentVolumeClaim(ctx context.Context, in api.PersistentVolumeClaimCreate) (api.PersistentVolumeClaim, error) {
-	id := uuid.New()
-	now := time.Now().UTC()
-
-	labelsJSON, err := marshalLabels(in.Labels)
-	if err != nil {
-		return api.PersistentVolumeClaim{}, err
-	}
-
-	const q = `
-		INSERT INTO persistent_volume_claims (
-			id, namespace_id, name, phase, storage_class_name,
-			volume_name, bound_volume_id, access_modes, requested_storage,
-			labels, created_at, updated_at
-		) VALUES ($1, $2, $3, $4, $5, $6, $7, $8, $9, $10, $11, $11)
-	`
-	_, err = p.pool.Exec(ctx, q,
-		id, in.NamespaceId, in.Name,
-		in.Phase, in.StorageClassName,
-		in.VolumeName, in.BoundVolumeId,
-		accessModesValue(in.AccessModes), in.RequestedStorage,
-		labelsJSON, now,
-	)
-	if err != nil {
-		var pgErr *pgconn.PgError
-		if errors.As(err, &pgErr) {
-			if pgErr.Code == "23505" {
-				return api.PersistentVolumeClaim{}, fmt.Errorf("pvc %q in namespace %s already exists: %w", in.Name, in.NamespaceId, api.ErrConflict)
-			}
-			if pErr := classifyPVCFKError(err, in.NamespaceId, in.BoundVolumeId); pErr != nil {
-				return api.PersistentVolumeClaim{}, pErr
-			}
-		}
-		return api.PersistentVolumeClaim{}, fmt.Errorf("insert pvc: %w", err)
-	}
-
-	return api.PersistentVolumeClaim{
-		Id:               &id,
-		NamespaceId:      in.NamespaceId,
-		Name:             in.Name,
-		Phase:            in.Phase,
-		StorageClassName: in.StorageClassName,
-		VolumeName:       in.VolumeName,
-		BoundVolumeId:    in.BoundVolumeId,
-		AccessModes:      in.AccessModes,
-		RequestedStorage: in.RequestedStorage,
-		Labels:           in.Labels,
-		CreatedAt:        &now,
-		UpdatedAt:        &now,
-	}, nil
-}
-
-// PVC read-side projection with parent-name JOINs (ADR-0027).
-const pvcSelectColumns = `pvc.id, pvc.namespace_id, pvc.name, pvc.phase, pvc.storage_class_name,
-	pvc.volume_name, pvc.bound_volume_id, pvc.access_modes, pvc.requested_storage,
-	pvc.labels, pvc.created_at, pvc.updated_at,
-	n.name AS namespace_name, n.cluster_id AS namespace_cluster_id, c.name AS cluster_name`
-
-const pvcFromJoined = `FROM persistent_volume_claims pvc
-	LEFT JOIN namespaces n ON n.id = pvc.namespace_id
-	LEFT JOIN clusters   c ON c.id = n.cluster_id`
-
-// GetPersistentVolumeClaim fetches a PVC by id.
-func (p *PG) GetPersistentVolumeClaim(ctx context.Context, id uuid.UUID) (api.PersistentVolumeClaim, error) {
-	q := `SELECT ` + pvcSelectColumns + ` ` + pvcFromJoined + ` WHERE pvc.id = $1`
-	row := p.pool.QueryRow(ctx, q, id)
-	pvc, err := scanPersistentVolumeClaim(row)
-	if err != nil {
-		if errors.Is(err, pgx.ErrNoRows) {
-			return api.PersistentVolumeClaim{}, api.ErrNotFound
-		}
-		return api.PersistentVolumeClaim{}, fmt.Errorf("select pvc: %w", err)
-	}
-	return pvc, nil
-}
-
-// ListPersistentVolumeClaims returns up to limit PVCs, optionally filtered by namespace.
-//
-//nolint:gocyclo // cursor-paginated query builder with optional filters
-func (p *PG) ListPersistentVolumeClaims(
-	ctx context.Context, namespaceID *uuid.UUID, limit int, cursor string,
-) ([]api.PersistentVolumeClaim, string, error) {
-	if limit <= 0 {
-		limit = 50
-	}
-	if limit > 200 {
-		limit = 200
-	}
-
-	sb := strings.Builder{}
-	sb.WriteString(`SELECT `)
-	sb.WriteString(pvcSelectColumns)
-	sb.WriteString(` `)
-	sb.WriteString(pvcFromJoined)
-	args := make([]any, 0, 4)
-	conds := make([]string, 0, 2)
-
-	if namespaceID != nil {
-		args = append(args, *namespaceID)
-		conds = append(conds, fmt.Sprintf("pvc.namespace_id = $%d", len(args)))
-	}
-	if cursor != "" {
-		ts, cid, err := decodeCursor(cursor)
-		if err != nil {
-			return nil, "", err
-		}
-		args = append(args, ts)
-		tsIdx := len(args)
-		args = append(args, cid)
-		idIdx := len(args)
-		conds = append(conds, fmt.Sprintf("(pvc.created_at, pvc.id) < ($%d, $%d)", tsIdx, idIdx))
-	}
-	if len(conds) > 0 {
-		sb.WriteString(" WHERE ")
-		sb.WriteString(strings.Join(conds, " AND "))
-	}
-	args = append(args, limit+1)
-	fmt.Fprintf(&sb, " ORDER BY pvc.created_at DESC, pvc.id DESC LIMIT $%d", len(args))
-
-	rows, err := p.pool.Query(ctx, sb.String(), args...)
-	if err != nil {
-		return nil, "", fmt.Errorf("query pvcs: %w", err)
-	}
-	defer rows.Close()
-
-	items := make([]api.PersistentVolumeClaim, 0, limit)
-	for rows.Next() {
-		pvc, err := scanPersistentVolumeClaim(rows)
-		if err != nil {
-			return nil, "", fmt.Errorf("scan pvc: %w", err)
-		}
-		items = append(items, pvc)
-	}
-	if err := rows.Err(); err != nil {
-		return nil, "", fmt.Errorf("iterate pvcs: %w", err)
-	}
-
-	var next string
-	if len(items) > limit {
-		last := items[limit-1]
-		if last.CreatedAt != nil && last.Id != nil {
-			next = encodeCursor(*last.CreatedAt, *last.Id)
-		}
-		items = items[:limit]
-	}
-	return items, next, nil
-}
-
-// UpdatePersistentVolumeClaim applies merge-patch on mutable PVC fields.
-//
-//nolint:gocyclo // merge-patch nil checks are inherently repetitive
-func (p *PG) UpdatePersistentVolumeClaim(ctx context.Context, id uuid.UUID, in api.PersistentVolumeClaimUpdate) (api.PersistentVolumeClaim, error) {
-	sets := make([]string, 0, 8)
-	args := make([]any, 0, 10)
-	idx := 1
-	appendSet := func(column string, value any) {
-		sets = append(sets, fmt.Sprintf("%s=$%d", column, idx))
-		args = append(args, value)
-		idx++
-	}
-
-	if in.Phase != nil {
-		appendSet("phase", *in.Phase)
-	}
-	if in.StorageClassName != nil {
-		appendSet("storage_class_name", *in.StorageClassName)
-	}
-	if in.VolumeName != nil {
-		appendSet("volume_name", *in.VolumeName)
-	}
-	if in.BoundVolumeId != nil {
-		appendSet("bound_volume_id", *in.BoundVolumeId)
-	}
-	if in.AccessModes != nil {
-		appendSet("access_modes", accessModesValue(in.AccessModes))
-	}
-	if in.RequestedStorage != nil {
-		appendSet("requested_storage", *in.RequestedStorage)
-	}
-	if in.Labels != nil {
-		b, err := marshalLabels(in.Labels)
-		if err != nil {
-			return api.PersistentVolumeClaim{}, err
-		}
-		appendSet("labels", b)
-	}
-	appendSet("updated_at", time.Now().UTC())
-	args = append(args, id)
-
-	q := fmt.Sprintf("UPDATE persistent_volume_claims SET %s WHERE id=$%d", strings.Join(sets, ", "), idx)
-	tag, err := p.pool.Exec(ctx, q, args...)
-	if err != nil {
-		if pErr := classifyPVCFKError(err, uuid.Nil, in.BoundVolumeId); pErr != nil {
-			return api.PersistentVolumeClaim{}, pErr
-		}
-		return api.PersistentVolumeClaim{}, fmt.Errorf("update pvc: %w", err)
-	}
-	if tag.RowsAffected() == 0 {
-		return api.PersistentVolumeClaim{}, api.ErrNotFound
-	}
-	return p.GetPersistentVolumeClaim(ctx, id)
-}
-
-// DeletePersistentVolumeClaim removes a PVC by id.
-func (p *PG) DeletePersistentVolumeClaim(ctx context.Context, id uuid.UUID) error {
-	tag, err := p.pool.Exec(ctx, "DELETE FROM persistent_volume_claims WHERE id=$1", id)
-	if err != nil {
-		return fmt.Errorf("delete pvc: %w", err)
-	}
-	if tag.RowsAffected() == 0 {
-		return api.ErrNotFound
-	}
-	return nil
-}
-
-// UpsertPersistentVolumeClaim inserts-or-updates a PVC keyed by (namespace_id, name).
-//
-//nolint:gocritic // hugeParam: Store interface requires value param
-func (p *PG) UpsertPersistentVolumeClaim(
-	ctx context.Context,
-	in api.PersistentVolumeClaimCreate,
-) (api.PersistentVolumeClaim, api.UpsertOutcome, error) {
-	id := uuid.New()
-	now := time.Now().UTC()
-
-	labelsJSON, err := marshalLabels(in.Labels)
-	if err != nil {
-		return api.PersistentVolumeClaim{}, api.OutcomeNoChange, err
-	}
-
-	// AUDIT_BUSINESS_FIELDS: phase, storage_class_name, volume_name,
-	// bound_volume_id, access_modes, requested_storage, labels.
-	// updated_at is a clock field — excluded.
-	const q = `
-		WITH old AS (
-		  SELECT phase, storage_class_name, volume_name, bound_volume_id,
-		         access_modes, requested_storage, labels
-		    FROM persistent_volume_claims WHERE namespace_id=$2 AND name=$3
-		),
-		upserted AS (
-		  INSERT INTO persistent_volume_claims (
-		    id, namespace_id, name, phase, storage_class_name,
-		    volume_name, bound_volume_id, access_modes, requested_storage,
-		    labels, created_at, updated_at
-		  ) VALUES ($1, $2, $3, $4, $5, $6, $7, $8, $9, $10, $11, $11)
-		  ON CONFLICT (namespace_id, name) DO UPDATE SET
-		      phase              = EXCLUDED.phase,
-		      storage_class_name = EXCLUDED.storage_class_name,
-		      volume_name        = EXCLUDED.volume_name,
-		      bound_volume_id    = EXCLUDED.bound_volume_id,
-		      access_modes       = EXCLUDED.access_modes,
-		      requested_storage  = EXCLUDED.requested_storage,
-		      labels             = EXCLUDED.labels,
-		      updated_at         = EXCLUDED.updated_at
-		  RETURNING id, namespace_id, name, phase, storage_class_name,
-		            volume_name, bound_volume_id, access_modes, requested_storage,
-		            labels, created_at, updated_at, xmax
-		)
-		SELECT u.id, u.namespace_id, u.name, u.phase, u.storage_class_name,
-		       u.volume_name, u.bound_volume_id, u.access_modes, u.requested_storage,
-		       u.labels, u.created_at, u.updated_at,
-		       n.name AS namespace_name, n.cluster_id AS namespace_cluster_id, c.name AS cluster_name,
-		       (u.xmax = 0) AS inserted,
-		       (u.xmax <> 0 AND (
-		           o.phase              IS DISTINCT FROM u.phase              OR
-		           o.storage_class_name IS DISTINCT FROM u.storage_class_name OR
-		           o.volume_name        IS DISTINCT FROM u.volume_name        OR
-		           o.bound_volume_id    IS DISTINCT FROM u.bound_volume_id    OR
-		           o.access_modes       IS DISTINCT FROM u.access_modes       OR
-		           o.requested_storage  IS DISTINCT FROM u.requested_storage  OR
-		           o.labels             IS DISTINCT FROM u.labels
-		       )) AS business_changed
-		  FROM upserted u
-		  LEFT JOIN old o        ON true
-		  LEFT JOIN namespaces n ON n.id = u.namespace_id
-		  LEFT JOIN clusters   c ON c.id = n.cluster_id
-	`
-	row := p.pool.QueryRow(ctx, q,
-		id, in.NamespaceId, in.Name,
-		in.Phase, in.StorageClassName,
-		in.VolumeName, in.BoundVolumeId,
-		accessModesValue(in.AccessModes), in.RequestedStorage,
-		labelsJSON, now,
-	)
-	var inserted, businessChanged bool
-	pvc, err := scanPersistentVolumeClaim(scanRowWith{row: row, extra: []any{&inserted, &businessChanged}})
-	if err != nil {
-		if pErr := classifyPVCFKError(err, in.NamespaceId, in.BoundVolumeId); pErr != nil {
-			return api.PersistentVolumeClaim{}, api.OutcomeNoChange, pErr
-		}
-		return api.PersistentVolumeClaim{}, api.OutcomeNoChange, fmt.Errorf("upsert pvc: %w", err)
-	}
-	return pvc, classifyOutcome(inserted, businessChanged), nil
-}
-
-// DeletePersistentVolumeClaimsNotIn removes namespace-scoped PVCs whose name
-// is not in keepNames.
-func (p *PG) DeletePersistentVolumeClaimsNotIn(ctx context.Context, namespaceID uuid.UUID, keepNames []string) (int64, error) {
-	tag, err := p.pool.Exec(ctx,
-		`DELETE FROM persistent_volume_claims
-		 WHERE namespace_id = $1
-		   AND name <> ALL(COALESCE($2::text[], ARRAY[]::text[]))`,
-		namespaceID, keepNames,
-	)
-	if err != nil {
-		return 0, fmt.Errorf("delete pvcs not in: %w", err)
-	}
-	return tag.RowsAffected(), nil
-}
-
-func scanPersistentVolumeClaim(row pgx.Row) (api.PersistentVolumeClaim, error) {
-	var (
-		pvc                api.PersistentVolumeClaim
-		id                 uuid.UUID
-		namespaceID        uuid.UUID
-		createdAt          time.Time
-		updatedAt          time.Time
-		phase              sql.NullString
-		storageClassName   sql.NullString
-		volumeName         sql.NullString
-		boundVolumeID      *uuid.UUID
-		accessModes        []string
-		requestedStorage   sql.NullString
-		labelsJSON         []byte
-		namespaceName      sql.NullString
-		namespaceClusterID uuid.NullUUID
-		clusterName        sql.NullString
-	)
-	if err := row.Scan(
-		&id, &namespaceID, &pvc.Name,
-		&phase, &storageClassName,
-		&volumeName, &boundVolumeID,
-		&accessModes, &requestedStorage,
-		&labelsJSON,
-		&createdAt, &updatedAt,
-		&namespaceName, &namespaceClusterID, &clusterName,
-	); err != nil {
-		return api.PersistentVolumeClaim{}, fmt.Errorf("scan pvc: %w", err)
-	}
-	pvc.Id = &id
-	pvc.NamespaceId = namespaceID
-	pvc.CreatedAt = &createdAt
-	pvc.UpdatedAt = &updatedAt
-	pvc.Phase = nullableString(phase)
-	pvc.StorageClassName = nullableString(storageClassName)
-	pvc.VolumeName = nullableString(volumeName)
-	pvc.BoundVolumeId = boundVolumeID
-	pvc.AccessModes = accessModesPointer(accessModes)
-	pvc.RequestedStorage = nullableString(requestedStorage)
-	pvc.NamespaceName = nullableString(namespaceName)
-	if namespaceClusterID.Valid {
-		v := namespaceClusterID.UUID
-		pvc.ClusterId = &v
-	}
-	pvc.ClusterName = nullableString(clusterName)
-	if len(labelsJSON) > 0 {
-		var labels map[string]string
-		if err := json.Unmarshal(labelsJSON, &labels); err != nil {
-			return api.PersistentVolumeClaim{}, fmt.Errorf("unmarshal pvc labels: %w", err)
-		}
-		if len(labels) > 0 {
-			pvc.Labels = &labels
-		}
-	}
-	return pvc, nil
-}
-
-// --- Settings ------------------------------------------------------------
-
-// GetSettings returns the current runtime settings from the single-row
-// settings table.
-func (p *PG) GetSettings(ctx context.Context) (api.Settings, error) {
-	const q = `SELECT eol_enabled, mcp_enabled,
-		time_travel_enabled, time_travel_retention_days, time_travel_reaper_enabled,
-		image_versions_enabled,
-		flow_matrix_enabled,
-		updated_at FROM settings WHERE id = 1`
-	var s api.Settings
-	if err := p.pool.QueryRow(ctx, q).Scan(
-		&s.EOLEnabled, &s.MCPEnabled,
-		&s.TimeTravelEnabled, &s.TimeTravelRetentionDays, &s.TimeTravelReaperEnabled,
-		&s.ImageVersionsEnabled,
-		&s.FlowMatrixEnabled,
-		&s.UpdatedAt,
-	); err != nil {
-		return api.Settings{}, fmt.Errorf("get settings: %w", err)
-	}
-	return s, nil
-}
-
 // isUniqueViolation reports whether err is a PostgreSQL unique-constraint
 // violation (SQLSTATE 23505). Used by CreateImageRegistry and other
 // single-row insert helpers that want to map the error to api.ErrConflict.
@@ -4998,60 +269,4 @@ func escapeLike(s string) string {
 	s = strings.ReplaceAll(s, `%`, `\%`)
 	s = strings.ReplaceAll(s, `_`, `\_`)
 	return s
-}
-
-// UpdateSettings applies the merge-patch on the settings row.
-func (p *PG) UpdateSettings(ctx context.Context, in api.SettingsPatch) (api.Settings, error) {
-	sets := make([]string, 0, 6)
-	args := make([]any, 0, 6)
-	idx := 1
-
-	if in.EOLEnabled != nil {
-		sets = append(sets, fmt.Sprintf("eol_enabled=$%d", idx))
-		args = append(args, *in.EOLEnabled)
-		idx++
-	}
-	if in.MCPEnabled != nil {
-		sets = append(sets, fmt.Sprintf("mcp_enabled=$%d", idx))
-		args = append(args, *in.MCPEnabled)
-		idx++
-	}
-	if in.TimeTravelEnabled != nil {
-		sets = append(sets, fmt.Sprintf("time_travel_enabled=$%d", idx))
-		args = append(args, *in.TimeTravelEnabled)
-		idx++
-	}
-	if in.TimeTravelRetentionDays != nil {
-		sets = append(sets, fmt.Sprintf("time_travel_retention_days=$%d", idx))
-		args = append(args, *in.TimeTravelRetentionDays)
-		idx++
-	}
-	if in.TimeTravelReaperEnabled != nil {
-		sets = append(sets, fmt.Sprintf("time_travel_reaper_enabled=$%d", idx))
-		args = append(args, *in.TimeTravelReaperEnabled)
-		idx++
-	}
-	if in.ImageVersionsEnabled != nil {
-		sets = append(sets, fmt.Sprintf("image_versions_enabled=$%d", idx))
-		args = append(args, *in.ImageVersionsEnabled)
-		idx++
-	}
-	if in.FlowMatrixEnabled != nil {
-		sets = append(sets, fmt.Sprintf("flow_matrix_enabled=$%d", idx))
-		args = append(args, *in.FlowMatrixEnabled)
-		idx++
-	}
-
-	if len(sets) == 0 {
-		return p.GetSettings(ctx)
-	}
-
-	sets = append(sets, fmt.Sprintf("updated_at=$%d", idx))
-	args = append(args, time.Now().UTC())
-
-	q := fmt.Sprintf("UPDATE settings SET %s WHERE id=1", strings.Join(sets, ", "))
-	if _, err := p.pool.Exec(ctx, q, args...); err != nil {
-		return api.Settings{}, fmt.Errorf("update settings: %w", err)
-	}
-	return p.GetSettings(ctx)
 }
